### 2023-02-16 #5156 @kyleniemeyer
Hi @esclapez, I'm going to serve as editor for your submission. We will use this issue to address any paper compilation problems and find reviewers. Any reviewer suggestions you can make from our [list](https://bit.ly/joss-reviewers) would be appreciated!

For the paper, I think the error is that author Candace Gilet is missing an affiliation.

Also, can you take a look at the missing DOI messages above and address those?
