### 2023-02-17 #5156 @kyleniemeyer
@esclapez this is ok, you can add an "unaffiliated" affiliation by doing something like:
```
affiliations:
...
 - name: Independent Researcher, Country
   index: 4
```
