### 2023-02-17 23fcead
Fix DOIs and add Candace independent affiliation.
