+++ paper/paper.bib
@@ -81,6 +81,7 @@
    pages     = {102836},
    year      = {2021},
    publisher = {Elsevier}
+   doi       = {10.1016/j.parco.2021.102836}
 }
 
 @misc{AMReX-Hydro,
@@ -104,7 +105,7 @@
 @article{Nonaka12,
   title= {A Deferred Correction Coupling Strategy for Low Mach Number Flow with Complex Chemistry},
   journal={Combustion Theory and Modeling},
-  doi = {10.1080/13647832.2012.701019},
+  doi = {10.1080/13647830.2012.701019},
   number = {16},
   publisher = {Taylor \& Francis},
   url = {https://www.tandfonline.com/doi/abs/10.1080/13647830.2012.701019},
@@ -148,6 +149,7 @@
   pages={123--168},
   year={1998},
   publisher={Taylor \& Francis}
+  doi = {10.1080/00102209808915770}
 }
 
 
