+++ paper/paper.bib
@@ -80,7 +80,7 @@
    volume    = {108},
    pages     = {102836},
    year      = {2021},
-   publisher = {Elsevier}
+   publisher = {Elsevier},
    doi       = {10.1016/j.parco.2021.102836}
 }
 
