+++ paper/paper.bib
@@ -148,7 +148,7 @@
   number={1-6},
   pages={123--168},
   year={1998},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
   doi = {10.1080/00102209808915770}
 }
 
