+++ Docs/source/manual/Validation.rst
@@ -16,25 +16,31 @@ is obtained from a Cantera simulation using
 the DRM19 mechanism and the simulation is conducted starting with a coarse resolution
 :math:`d_{l,0}/\delta_x = 3.5` (3.5 grid cells in the flame thermal thickness) and progressively
 refined to :math:`d_{l,0}/\delta_x = 110` using AMR. Note that the simulation is carried for
-about 30 :math:`\tau_f = d_{l,0} / S_{l,0}` ensuring that initial numerical noise introduced
-by interpolating the Cantera solution onto the cartesian grid is completely removed.
+about 30 :math:`\tau_f = d_{l,0} / S_{l,0}` before adding refinement levels ensuring that initial 
+numerical noise introduced by interpolating the Cantera solution onto the cartesian grid is 
+completely removed.
 
-The `PeleLMeX` results obtained with the finest grid are compared to that of Cantera in the following
-figure. Profiles of major and intermediate species as well as temperature across the flame front are
+The `PeleLMeX` results obtained with the finest grid are compared to that of Cantera in :numref:`PMFLMeXCantera`.
+Profiles of major and intermediate species as well as temperature across the flame front are
 displayed, with Cantera results shown as black ticked lined underlying each continuous line obtained
 with `PeleLMeX`, demonstrating the solver accuracy.
 
 .. figure:: images/validations/PremixedFlame/LMeX_DRM19_Phi0p9.png
+   :name: PMFLMeXCantera
    :align: center
-   :figwidth: 60%
+   :figwidth: 50%
 
-It is interesting to see how resolution affect the accuracy of `PeleLMeX` results. The following table
+   : One-dimentional flame profiles with PeleLMeX and Cantera (underlying black dashed lines)
+
+It is interesting to see how resolution affect the prediction of `PeleLMeX`. The :numref:`PMFTableResolution`
 shows the laminar flame speed, obtained by numerically integrating the fuel consumption rate, at
 increasing level of refinement. For each AMR level addition, simulation are conducted for about 10 :math:`\tau_f`,
-until the flame speed stabilize. With only a handful of cells across the flame thermal thickness, the flame
-speed is recovered with an accuracy close to 1.0%.
+until the flame speed stabilizes. With only a handful of cells across the flame thermal thickness, the flame
+speed is recovered with an accuracy close to 1.5%.
 
 .. list-table:: Effect of resolution on the laminar flame speed
+    :name: PMFTableResolution
+    :align: center
     :widths: 50 25 25 25 25 25 25
     :header-rows: 1
 
@@ -60,15 +66,18 @@ speed is recovered with an accuracy close to 1.0%.
       - 0.58
       - 0.57
 
-Additionally, the following figure compared profiles of major and intermediate species as well as 
+Additionally, :numref:`PMFLMeXCrseFine` compares profiles of major and intermediate species as well as 
 temperature across the flame front for the coarsest and the finest resolution employed. Results
 indicates that major species (or temperature) are well captured whereas short lived species peak
 values are locally off due to the lack of resolution but are stil reasonably well located within
 the flame front.
 
 .. figure:: images/validations/PremixedFlame/LMeX_FinevsCoarse.png
+   :name: PMFLMeXCrseFine
    :align: center
-   :figwidth: 60%
+   :figwidth: 50%
+
[…]
