### 2023-02-21 fc85dbe
allow restart with fewer levels than previous checkpoint
