+++ paper/paper.bib
@@ -152,14 +152,13 @@
   doi = {10.1080/00102209808915770}
 }
 
-
-@article{Berger:2021,
-  title={A state redistribution algorithm for finite volume schemes on cut cell meshes},
-  author={Berger, Marsha and Giuliani, Andrew},
+@article{Giuliani:2022,
+  title={A weighted state redistribution algorithm for embedded boundary grids},
+  author={Giuliani, Andrew and Almgren, Ann S and Bell, John B and Berger, Marsha J and de Frahan, MT Henry and Rangarajan, Deepak},
   journal={Journal of Computational Physics},
-  doi = {10.1016/j.jcp.2020.109820},
-  volume={428},
-  pages={109820},
-  year={2021},
+  doi = {10.1016/j.jcp.2022.111305},
+  volume={464},
+  pages={111305},
+  year={2022},
   publisher={Elsevier}
 }
