### 2023-02-28 81c60f2
During regrid, ensure rho = sum rhoYs and recompute temperature.
