+++ Source/PeleLMRegrid.cpp
@@ -81,6 +81,9 @@ void PeleLM::MakeNewLevelFromCoarse( int lev,
          setRhoToSumRhoY(lev, AmrNewTime);
       }
 
+      // Recompute temperature from interpolated rhoYs and rhoH 
+      setTemperature(lev, AmrNewTime);
+
       // Initialize thermodynamic pressure
       setThermoPress(lev, AmrNewTime);
    }
@@ -176,6 +179,15 @@ void PeleLM::RemakeLevel( int lev,
    m_resetCoveredMask = 1;
 
    if (!m_incompressible) {
+      // Enforce density / species density consistency
+      // only usefull when using cell cons interp
+      if (m_regrid_interp_method == 1) {
+         setRhoToSumRhoY(lev, AmrNewTime);
+      }
+
+      // Recompute temperature from interpolated rhoYs and rhoH 
+      setTemperature(lev, AmrNewTime);
+
       // Initialize thermodynamic pressure
       setThermoPress(lev, AmrNewTime);
    }
