+++ Source/PeleLMBC.cpp
@@ -99,7 +99,7 @@ InterpBase* PeleLM::getInterpolator(int a_method) {
 #endif
   } else if (a_method == 2) {
 #ifdef AMREX_USE_EB
-    Abort("Regrid interpolation method = 2 not available with EB !"); 
+    Abort("Regrid interpolation method = 2 not available with EB !");
 #else
     mapper = &mf_linear_slope_minmax_interp;
 #endif
