### 2023-02-28 3fd1187
Add 2 new run-time diagnostics: PDF and conditional stats (#173)
* Add PDF and conditionals. WIP.

* Fix mask in DiagPDF.

* More WIP on Conditional.

* Add Barrier in FramePlane.

* Update diags to have better control over the outputed variables.
PDF and conditional write plain ASCII.

* Add a light descriptor of the reaction state.

* Add coordinate as a derived.

* Manage diagnostic variables.

* Add reactVariables checks. One can now tag based on reaction rate of a species.

* Add diagnostic example in FlameSheet 3D.

* Trailing whitespace.

* More trailing whitespace.

* Add documentation
[…]
