+++ Source/PeleLMDeriveFunc.cpp
@@ -12,7 +12,7 @@ using namespace amrex;
 //
 void pelelm_dertemp (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int dcomp, int ncomp,
                      const FArrayBox& statefab, const FArrayBox& /*reactfab*/, const FArrayBox& /*pressfab*/,
-                     const Geometry& /*geomdata*/,
+                     const Geometry& /*geom*/,
                      Real /*time*/, const Vector<BCRec>& /*bcrec*/, int /*level*/)
 
 {
@@ -34,7 +34,7 @@ void pelelm_dertemp (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int dco
 //
 void pelelm_derheatrelease (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int dcomp, int ncomp,
                             const FArrayBox& statefab, const FArrayBox& reactfab, const FArrayBox& /*pressfab*/,
-                            const Geometry& /*geomdata*/,
+                            const Geometry& /*geom*/,
                             Real /*time*/, const Vector<BCRec>& /*bcrec*/, int level)
 
 {
@@ -66,7 +66,7 @@ void pelelm_derheatrelease (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, 
 //
 void pelelm_dermassfrac (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int dcomp, int ncomp,
                          const FArrayBox& statefab, const FArrayBox& /*reactfab*/, const FArrayBox& /*pressfab*/,
-                         const Geometry& /*geomdata*/,
+                         const Geometry& /*geom*/,
                          Real /*time*/, const Vector<BCRec>& /*bcrec*/, int /*level*/)
 
 {
@@ -91,7 +91,7 @@ void pelelm_dermassfrac (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int
 //
 void pelelm_dermolefrac (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int dcomp, int ncomp,
                          const FArrayBox& statefab, const FArrayBox& /*reactfab*/, const FArrayBox& /*pressfab*/,
-                         const Geometry& /*geomdata*/,
+                         const Geometry& /*geom*/,
                          Real /*time*/, const Vector<BCRec>& /*bcrec*/, int /*level*/)
 {
     AMREX_ASSERT(derfab.box().contains(bx));
@@ -124,7 +124,7 @@ void pelelm_dermolefrac (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int
 //
 void pelelm_derrhomrhoy (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int dcomp, int ncomp,
                          const FArrayBox& statefab, const FArrayBox& /*reactfab*/, const FArrayBox& /*pressfab*/,
-                         const Geometry& /*geomdata*/,
+                         const Geometry& /*geom*/,
                          Real /*time*/, const Vector<BCRec>& /*bcrec*/, int /*level*/)
 
 {
@@ -151,7 +151,7 @@ void pelelm_derrhomrhoy (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int
 //
 void pelelm_deravgpress (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
                          const FArrayBox& /*statefab*/, const FArrayBox& /*reactfab*/, const FArrayBox& pressfab,
-                         const Geometry& /*geomdata*/,
[…]
