+++ manual/html/_sources/LMeXControls.rst.txt
@@ -148,6 +148,9 @@ The following list of derived variables are available in PeleLMeX:
     * - `rhominsumrhoY`
       - 1
       - Rho minus sum of rhoYs, for debug purposes
+    * - `coordinates`
+      - AMREX_SPACEDIM
+      - Cell-center coordinates
 
 Note that `mixture_fraction` and `progress_variable` requires additional inputs from the users as described below.
 
@@ -331,28 +334,58 @@ state:
     peleLM.progressVariable.hotState = CO:0.000002 CO2:0.0666
 
 
-A set of diagnostics available at runtime are currently under development. The following provide an example for extracting
-the state variables on a 'x','y' or 'z' aligned plane and writting a 2D plotfile compatible with Amrvis, Paraview or yt:
+Analysing the data a-posteriori can become extremely cumbersome when dealing with extreme datasets.
+PeleLMeX offers a set of diagnostics available at runtime and more are under development.
+Currently, the list of diagnostic contains:
+
+* `DiagFramePlane` : extract a plane aligned in the 'x','y' or 'z' direction across the AMR hierarchy, writing
+  a 2D plotfile compatible with Amrvis, Paraview or yt. Only available for 3D simulations.
+* `DiagPDF` : extract the PDF of a given variable and write it to an ASCII file.
+* `DiagConditional` : extract statistics (average and standard deviation, integral or sum) of a
+  set of variables conditioned on the value of given variable and write it to an ASCII file.
+
+When using `DiagPDF` or `DiagConditional`, it is possible to narrow down the diagnostic to a region of interest
+by specifying a set of filters, defining a range of interest for a variable. Note also the for these two diagnostics,
+fine-covered regions are masked. The following provide examples for each diagnostic:
 
 ::
 
     #--------------------------DIAGNOSTICS------------------------
 
-    peleLM.diagnostics = xnormal ynormal
-    peleLM.xnormal.type = DiagFramePlane
-    peleLM.xnormal.file = xNorm5mm
-    peleLM.xnormal.normal = 0
-    peleLM.xnormal.center = 0.005
-    peleLM.xnormal.int    = 5
-    peleLM.xnormal.interpolation = Linear
+    peleLM.diagnostics = xnormP condT pdfTest
 
-    peleLM.ynormal.type = DiagFramePlane
-    peleLM.ynormal.file = yNormCent
-    peleLM.ynormal.normal = 1
-    peleLM.ynormal.center = 0.0
-    peleLM.ynormal.int    = 10
-    peleLM.ynormal.interpolation = Quadratic
+    peleLM.xnormP.type = DiagFramePlane                             # Diagnostic type
+    peleLM.xnormP.file = xNorm5mm                                   # Output file prefix
+    peleLM.xnormP.normal = 0                                        # Plane normal (0, 1 or 2 for x, y or z)
+    peleLM.xnormP.center = 0.005                                    # Coordinate in the normal direction
+    peleLM.xnormP.int    = 5                                        # Frequency (as step #) for performing the diagnostic
+    peleLM.xnormP.interpolation = Linear                            # [OPT, DEF=Linear] Interpolation type : Linear or Quadratic
+    peleLM.xnormP.field_names = x_velocity mag_vort density         # List of variables outputed to the 2D pltfile
 
+    peleLM.condT.type = DiagConditional                             # Diagnostic type
+    peleLM.condT.file = condTest                                    # Output file prefix
+    peleLM.condT.int  = 5                                           # Frequency (as step #) for performing the diagnostic
+    peleLM.condT.filters = xHigh stoich                             # [OPT, DEF=None] List of filters
+    peleLM.condT.xHigh.field_name = x                               # Filter field
[…]
