+++ paper/paper.bib
@@ -162,3 +162,14 @@
   year={2022},
   publisher={Elsevier}
 }
+ 
+@article{Perry:2022,
+  title={Co-optimized machine-learned manifold models for large eddy simulation of turbulent combustion},
+  author={Perry, Bruce A and de Frahan, Marc T Henry and Yellapantula, Shashank},
+  journal={Combustion and Flame},
+  doi = {10.1016/j.combustflame.2022.112286},
+  volume={244},
+  pages={112286},
+  year={2022},
+  publisher={Elsevier}
+}
