### 2023-03-01 5751900
Extend validation section with EBchannelFlow and 1D PMF. (#180)
* Add EBChannelFlow validation test case. LES results are still in process.

* Add data from a laminar premixed flame case and comparisons
with Cantera.

* A fresh pass on the doc.

* Add a couple of missing files.

* Add the Re_t = 934 case to the validation and stop there for now.

* Trailing whitespace.
