+++ Source/Soot/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_include_directories(pelelmex PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}>)
+
+target_sources(pelelmex
+    PRIVATE
+         PeleLMSoot.cpp
+    )
