### 2023-03-09 ae1ab95
First pass at setting CMake framework. (using initCMake branch out of PelePhysics).
