+++ Utils/CMake/SetPelePhysicsOptions.cmake
@@ -1,0 +1,10 @@
+set(PELEPHYSICS_SPACEDIM ${PELELMEX_DIM} CACHE STRING "Spatial dimensions" FORCE)
+set(PELEPHYSICS_CHEMISTRY_MODEL ${PELELMEX_CHEMISTRY_MODEL} CACHE STRING "Chemical mechanism")
+set(PELEPHYSICS_EOS_MODEL ${PELELMEX_EOS_MODEL} CACHE STRING "Eq. of state model")
+set(PELEPHYSICS_TRANSPORT_MODEL ${PELELMEX_TRANSPORT_MODEL} CACHE STRING "Transport model")
+
+set(PELEPHYSICS_MPI ${PELELMEX_MPI})
+set(PELEPHYSICS_OMP ${PELELMEX_OPENMP})
+set(PELEPHYSICS_CUDA ${PELELMEX_CUDA})
+set(PELEPHYSICS_HIP ${PELELMEX_HIP})
+set(PELEPHYSICS_SYCL ${PELELMEX_SYCL})
