### 2023-03-14 0b5b7ab
Explose more GMG options for the TensorSolve.
