+++ Utils/CMake/SetPelePhysicsOptions.cmake
@@ -1,4 +1,5 @@
 set(PELEPHYSICS_SPACEDIM ${PELELMEX_DIM} CACHE STRING "Spatial dimensions" FORCE)
+set(PELEPHYSICS_PRECISION $PELELMEX_PRECISION} CACHE STRING "Floating point precision" FORCE)
 set(PELEPHYSICS_CHEMISTRY_MODEL ${PELELMEX_CHEMISTRY_MODEL} CACHE STRING "Chemical mechanism")
 set(PELEPHYSICS_EOS_MODEL ${PELELMEX_EOS_MODEL} CACHE STRING "Eq. of state model")
 set(PELEPHYSICS_TRANSPORT_MODEL ${PELELMEX_TRANSPORT_MODEL} CACHE STRING "Transport model")
