+++ Source/PeleLMInit.cpp
@@ -89,7 +89,7 @@ void PeleLM::MakeNewLevelFromScratch( int lev,
    m_t_old[lev] = time - 1.0e200;
 
    // Mac projector
-#if AMREX_USE_EB
+#ifdef AMREX_USE_EB
    macproj.reset(new Hydro::MacProjector(Geom(0,finest_level),
                                          MLMG::Location::FaceCentroid,  // Location of mac velocity
                                          MLMG::Location::FaceCentroid,  // Location of beta
@@ -102,7 +102,7 @@ void PeleLM::MakeNewLevelFromScratch( int lev,
                                        MFInfo(), *m_factory[lev]));
    m_extSource[lev]->setVal(0.);
 
-#if AMREX_USE_EB
+#ifdef AMREX_USE_EB
    if ( lev == 0 && m_signDistNeeded) {
       // Set up CC signed distance container to control EB refinement
       m_signedDist0.reset(new MultiFab(grids[lev], dmap[lev], 1, 1, MFInfo(), *m_factory[lev]));
