+++ CMakeLists.txt
@@ -1,8 +1,9 @@
-cmake_minimum_required(VERSION 3.14)
+cmake_minimum_required(VERSION 3.23)
 
 project( PeleLMeX
          DESCRIPTION "An AMR Low Mach Number Reactive Flow Simulation Code Without Level Sub-cycling"
          HOMEPAGE_URL "https://github.com/AMReX-Combustion/PeleLMeX"
+         VERSION 23.03
          LANGUAGES CXX
        )
 
@@ -13,11 +14,12 @@ list(APPEND CMAKE_MODULE_PATH "${CMAKE_SOURCE_DIR}/Utils/CMake")
 #
 # Base options
 #
-set(PELELMEX_DIM             "3"      CACHE STRING "Number of physical dimensions")
-set(PELELMEX_PRECISION       "DOUBLE" CACHE STRING "Floating point precision SINGLE or DOUBLE")
-set(PELELMEX_CHEMISTRY_MODEL drm19    CACHE STRING "Chemical mechanism")
-set(PELELMEX_EOS_MODEL       Fuego    CACHE STRING "Equation of state model")
-set(PELELMEX_TRANSPORT_MODEL Simple   CACHE STRING "Transport model")
+set(PELELMEX_DIM             "3"        CACHE STRING "Number of physical dimensions")
+set(PELELMEX_PRECISION       "DOUBLE"   CACHE STRING "Floating point precision SINGLE or DOUBLE")
+set(PELELMEX_CHEMISTRY_MODEL drm19      CACHE STRING "Chemical mechanism")
+set(PELELMEX_EOS_MODEL       Fuego      CACHE STRING "Equation of state model")
+set(PELELMEX_TRANSPORT_MODEL Simple     CACHE STRING "Transport model")
+set(PELELMEX_CASE            FlameSheet CACHE STRING "Case folder")
 
 #
 # Physics options
@@ -25,7 +27,7 @@ set(PELELMEX_TRANSPORT_MODEL Simple   CACHE STRING "Transport model")
 option(PELELMEX_USE_EB        "Enable Embedded Boundary"     OFF)
 option(PELELMEX_USE_EFIELD    "Enable Electric field module" OFF)
 option(PELELMEX_USE_PARTICLES "Enable Spray module"          OFF)
-option(PELELMEX_USE_SOOR      "Enable Soot module"           OFF)
+option(PELELMEX_USE_SOOT      "Enable Soot module"           OFF)
 
 #
 # Misc options
@@ -48,37 +50,68 @@ set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_EXTENSIONS OFF)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
+#
+# Setup Languages
+#
+if (PELELMEX_CUDA)
+   enable_language(CUDA)
+endif ()
+if (PELELMEX_HIP)
+   enable_language(HIP)
+endif ()
+
 # Checks
 if(PELELMEX_DIM EQUAL 1)
     message(FATAL_ERROR "PeleLMeX does not support 1D.")
 endif()
 
+#
+# Cases
+#
+set(PELELMEX_EXEC_SUBDIRS Exec/RegTests Exec/Cases)
+list(TRANSFORM PELELMEX_EXEC_SUBDIRS PREPEND "${CMAKE_CURRENT_LIST_DIR}/")
+set(EXEC_DIR "")
+foreach(subdir ${PELELMEX_EXEC_SUBDIRS})
+    FILE(GLOB caselist ${subdir}/*)
+    foreach(case ${caselist})
+        if(IS_DIRECTORY ${case})
+            get_filename_component( case_dir ${case} NAME)
+            LIST(APPEND PELELMEX_CASE_LIST ${case_dir})
[…]
