### 2023-03-16 0aff5bb
Add basic CMake capabilities (#183)
* Add source (sub-)folders CMakeLists.

* Add CMakeLists to FlameSheet case.

* First pass at setting CMake framework.
(using initCMake branch out of PelePhysics).

* Add Utils/CMake, mostly pulling from PeleC.

* Missing files in sources.

* Pass precision to PelePhysics CMake.

* Add CMakeLists.txt in Cases folders.

* CMakeLists in RegTests.

* Fix preprocessor in PeleLMInit.cpp.

* Update base CMakeLists.txt.
