+++ Exec/RegTests/HITDecay/CMakeLists.txt
@@ -1,7 +1,7 @@
 target_include_directories(pelelmex PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}>)
 
-set(PELELMEX_CHEMISTRY_MODEL dodecane_lu PARENT_SCOPE)
-set(PELELMEX_EOS_MODEL       Fuego PARENT_SCOPE)
+set(PELELMEX_EOS_MODEL Fuego PARENT_SCOPE)
+set(PELELMEX_CHEMISTRY_MODEL air PARENT_SCOPE)
 set(PELELMEX_TRANSPORT_MODEL Simple PARENT_SCOPE)
 set(PELELMEX_DIM "3" PARENT_SCOPE)
 set(PELELMEX_MPI ON PARENT_SCOPE)
