+++ Utils/CMake/SetAmrexHydroOptions.cmake
@@ -1,0 +1,6 @@
+set(HYDRO_SPACEDIM ${PELELMEX_DIM} CACHE STRING "Spatial dimensions" FORCE)
+set(HYDRO_EB ${PELELMEX_USE_EB}) 
+set(HYDRO_MPI ${PELELMEX_MPI})
+set(HYDRO_OMP ${PELELMEX_OPENMP})
+set(HYDRO_GPU_BACKEND ${AMReX_GPU_BACKEND} CACHE STRING "AMReX GPU type" FORCE)
+set(HYDRO_PROJECTIONS ON)
