+++ .github/workflows/codeql.yml
@@ -1,0 +1,45 @@
+name: 🔍 CodeQL
+
+on:
+  push:
+    branches: [ "development" ]
+  pull_request:
+    branches: [ "development" ]
+  schedule:
+    - cron: "27 3 * * 0"
+
+concurrency:
+  group: ${{ github.ref }}-${{ github.head_ref }}-codeql
+  cancel-in-progress: true
+
+jobs:
+  analyze:
+    name: Analyze
+    runs-on: ubuntu-latest
+    permissions:
+      actions: read
+      contents: read
+      security-events: write
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+        with:
+          submodules: true
+      - name: System Dependencies
+        run: .github/workflows/dependencies/dependencies_gcc10.sh
+      - name: Initialize CodeQL
+        uses: github/codeql-action/init@v2
+        with:
+          languages: cpp
+          queries: +security-and-quality
+      - name: Configure
+        run: |
+          cmake -S . -B build
+      - name: Build
+        run: |
+          cmake --build build --parallel ${{env.NPROCS}}
+      - name: Perform CodeQL Analysis
+        uses: github/codeql-action/analyze@v2
+        with:
+          category: "/language:cpp"
