+++ .github/workflows/codeql.yml
@@ -38,7 +38,7 @@ jobs:
           cmake -S . -B build
       - name: Build
         run: |
-          cmake --build build --parallel ${{env.NPROCS}}
+          cmake --build build --parallel 2
       - name: Perform CodeQL Analysis
         uses: github/codeql-action/analyze@v2
         with:
