### 2023-03-17 fdb27b2
Address CodeQL warnings.
