### 2023-03-17 2d969fe
Init CodeQL CI. (#185)
* Init CodeQL CI.

* Update PelePhysics

* Limit to 2 procs in cmake --build.

* Follow PeleC and filter out AMReX/AMReX-Hydro/SUNDIALS in CodeQL output.

* Small amrex.

* Address CodeQL warnings.

* Switch to native GH Action concurrency.
