+++ README.md
@@ -6,15 +6,21 @@
 
 ## Overview
 
-*PeleLMeX* is a non-subcycling version of [*PeleLM*](https://github.com/AMReX-Combustion/PeleLM) based on AMReX's [AmrCore](https://amrex-codes.github.io/amrex/docs_html/AmrCore.html) and borrowing from the incompressible solver [incflo](https://github.com/AMReX-Codes/incflo).
+![CUDA build](https://github.com/AMReX-Combustion/PeleLMeX/workflows/cuda/badge.svg)
+![HIP build](https://github.com/AMReX-Combustion/PeleLMeX/workflows/hip/badge.svg)
+![SYCL build](https://github.com/AMReX-Combustion/PeleLMeX/workflows/intel/badge.svg)
+
+*PeleLMeX* is a non-subcycling version of [*PeleLM*](https://github.com/AMReX-Combustion/PeleLM) based on AMReX's [AmrCore](https://amrex-codes.github.io/amrex/docs_html/AmrCore.html) and borrowing from the incompressible solver [incflo](https://github.com/AMReX-Codes/incflo). It is designed to run efficiently on small workstations as well as the largest ExaScale platforms currently available.
 
 *PeleLMeX* is part of the [Pele combustion Suite](https://amrex-combustion.github.io/).
 
 ## Documentation
 
+![Documentation](https://github.com/AMReX-Combustion/PeleLMeX/workflows/docs/badge.svg)
+
 *PeleLMeX* solves of the multispecies reactive Navier-Stokes equations in the low Mach number limit as described in the [documentation](https://amrex-combustion.github.io/PeleLMeX/manual/html/index.html). It inherits most of *PeleLM* algorithmic features, but differs significantly in its implementation stemming from the non-subcycling approach.
 
-A overview of PeleLMeX controls is provided in the [documentation](https://amrex-combustion.github.io/PeleLMeX/manual/html/LMeXControls.html).
+A overview of *PeleLMeX* controls is provided in the [documentation](https://amrex-combustion.github.io/PeleLMeX/manual/html/LMeXControls.html).
 
 ### Core Algorithm
 
@@ -32,27 +38,35 @@ https://amrex-combustion.github.io/PeleLMeX/manual/html/Tutorials.html
 
 ## Installation
 
-### Using git submodules
+### Download
 
-You can clone the PeleLMeX and tested versions of its submodules
-([amrex](https://github.com/AMReX-Codes/amrex), [AMReX-Hydro](https://github.com/AMReX-Codes/AMReX-Hydro) and [PelePhysics](https://github.com/AMReX-Combustion/PelePhysics)) using:
+The prefered method consist in cloning the *PeleLMeX* and its submodules using a recursive `git clone`:
+([amrex](https://github.com/AMReX-Codes/amrex), [AMReX-Hydro](https://github.com/AMReX-Codes/AMReX-Hydro) and [PelePhysics](https://github.com/AMReX-Combustion/PelePhysics), [PeleMP](https://github.com/AMReX-Combustion/PeleMP)) using:
 
 ```
-git clone --recursive https://github.com/AMReX-Combustion/PeleLMeX.git
+git clone --depth 1 --recursive https://github.com/AMReX-Combustion/PeleLMeX.git
 ```
 
-You can optionally setup the following environment variables (e.g. using bash):
+Alternatively, you can use separate `git clone` of the each of the submodules.
+The default location for *PeleLMeX* dependencies is the `Submodule` folder but you optionnally
+setup the following environment variables (e.g. using bash) to an other location:
 
 ```
 export PELELMEX_HOME=<path_to_PeleLMeX>
[…]
