### 2023-03-23 04a7890
EB diffusion fluxes have a single component.
