### 2023-03-23 6a8b6c3
Simplify diff flux divergence.
