### 2023-03-25 aedc392
Need a ghost cell when processing the diagnostics.
