### 2023-03-25 77ce8c4
Directly handle 2D in here instead of re-writting after 3D AMReX call.
