+++ Source/Diagnostics/DiagFramePlane.cpp
@@ -301,6 +301,8 @@ DiagFramePlane::Write2DMultiLevelPlotfile(const std::string &a_pltfile,
     amrex::ParallelDescriptor::Barrier();
 
     if (amrex::ParallelDescriptor::MyProc() == amrex::ParallelDescriptor::NProcs()-1) {
+
+        // Write 2D pltfile header
         amrex::Vector<amrex::BoxArray> boxArrays(a_nlevels);
         for(int level(0); level < boxArrays.size(); ++level) {
             boxArrays[level] = a_slice[level]->boxArray();
@@ -317,12 +319,32 @@ DiagFramePlane::Write2DMultiLevelPlotfile(const std::string &a_pltfile,
         Write2DPlotfileHeader(HeaderFile, a_nlevels, boxArrays, a_varnames,
                               a_geoms, a_time, a_steps, a_rref, versionName,
                               levelPrefix, mfPrefix);
+        HeaderFile.flush();
+        HeaderFile.close();
+
+        // Add extra file providing the plane information
+        // because the output is always a z-normal plane at z=0
+        std::string PlaneFileName(a_pltfile + "/PlaneData");
+        std::ofstream PlaneFile;
+        PlaneFile.rdbuf()->pubsetbuf(io_buffer.dataPtr(), io_buffer.size());
+        PlaneFile.open(PlaneFileName.c_str(), std::ofstream::out   |
+                                              std::ofstream::trunc |
+                                              std::ofstream::binary);
+        if( ! PlaneFile.good()) amrex::FileOpenFailed(PlaneFileName);
+
+        PlaneFile << "Pele plane definition: \n";
+        PlaneFile << m_normal << "\n";
+        PlaneFile << AMREX_D_TERM(m_center[0],
+                                  << " " << m_center[1],
+                                  << " " << m_center[2]) << "\n";
+
+        PlaneFile.flush();
+        PlaneFile.close();
     }
+
+    // Write a 2D version of the MF at each level
     for (int level = 0; level < a_nlevels; ++level) {
-        VisMF2D(*a_slice[level], amrex::MultiFabFileFullPrefix(level, a_pltfile, levelPrefix, mfPrefix+"2D"));
-        amrex::VisMF::Write(*a_slice[level], amrex::MultiFabFileFullPrefix(level, a_pltfile, levelPrefix, mfPrefix));
-        amrex::ParallelDescriptor::Barrier();
-        ReWriteLevelVisMFHeader(amrex::MultiFabFileFullPrefix(level, a_pltfile, levelPrefix, ""));
+        VisMF2D(*a_slice[level], amrex::MultiFabFileFullPrefix(level, a_pltfile, levelPrefix, mfPrefix));
     }
 }
 
@@ -399,147 +421,6 @@ DiagFramePlane::Write2DPlotfileHeader(std::ostream &HeaderFile,
             }
         }
         HeaderFile << amrex::MultiFabHeaderPath(level, levelPrefix, mfPrefix) << '\n';
-    }
-}
-
-void
-DiagFramePlane::ReWriteLevelVisMFHeader(const std::string &a_HeaderPath)
-{
-    std::string OldHeaderFileName(a_HeaderPath+ "Cell_H");
-    amrex::Vector<char> oldfileCharPtr;
-    amrex::ParallelDescriptor::ReadAndBcastFile(OldHeaderFileName, oldfileCharPtr);
-
-    if (amrex::ParallelDescriptor::IOProcessor()) {
-        amrex::VisMF::IO_Buffer io_buffer_new(amrex::VisMF::IO_Buffer_Size);
-        std::string HeaderFileName(a_HeaderPath+ "Cell_temp_H");
[…]
