### 2023-03-27 3757986
Abort if a derived with more than 1 component is required in the Diagnostics.
