### 2023-03-27 7e41cde
Untag covered cells and update backward step flame.
