### 2023-03-28 e207cc2
allow single components of multicomponent derive
