+++ Source/PeleLMDiagnostics.cpp
@@ -35,7 +35,7 @@ PeleLM::createDiagnostics()
         } else {
             if (derive_lst.canDerive(v)) {
                 const PeleLMDeriveRec* rec = derive_lst.get(v);
-                if (rec->numDerive() > 1) {
+                if (rec->variableComp(v) < 0) {
                     std::string errmsg = "Diagnostics can't handle derived with more than 1 component at the moment.\n";
                     errmsg += "Add the desired components individually.\n";
                     Abort(errmsg);
