### 2023-03-29 fb2870b
EBDiff is not a pointer.
