### 2023-03-29 5af2050
Add function to control EB type using a diffusivity pre-factor.
