### 2023-03-30 e8422f7
Provide documentation on isothermal EB and also a section on refinement criteria.
