### 2023-04-04 3537ce5
Add a note on isothermal EB + LES.
