### 2023-04-04 63c6bf0
Enable isothermal EB specification (#84)
* Remove unused kernel and minor ifdef changes.

* First pass at isoEB. Something is up.

* FlowPastCylinder with hot EB.

* Add isoT EB input file.

* Need to scale the EBFlux by dx because area is a fraction.

* Add a diffuse_scalar with EB dirichlet version.

* Try using a non-homogeneous dirichlet BC for the deltaT solve ...
not convinced it works better.

* Fix arguments order.

* Need a zero dirichlet on the EB boundary for the deltaT solve.

* Add EB temperature in PipeFlow case

* Update the isoT input of FlowPastCylinder

* Add EB_BackwardStepF
[…]
