+++ Source/PeleLMUtils.cpp
@@ -29,12 +29,38 @@ void PeleLM::fluxDivergence(const Vector<MultiFab*> &a_divergence,
    }
 }
 
+void PeleLM::fluxDivergence(const Vector<MultiFab*> &a_divergence,
+                            int div_comp,
+                            const Vector<Array<MultiFab*,AMREX_SPACEDIM> > &a_fluxes,
+                            int flux_comp,
+                            const Vector<MultiFab* > &a_EBfluxes,
+                            int ebflux_comp,
+                            int ncomp,
+                            int intensiveFluxes,
+                            Real scale) {
+
+   BL_PROFILE("PeleLM::fluxDivergence()");
+   if (intensiveFluxes) {        // Fluxes are intensive -> need area scaling in div
+      for (int lev = 0; lev <= finest_level; ++lev) {
+         intFluxDivergenceLevelEB(lev,*a_divergence[lev], div_comp, a_fluxes[lev], flux_comp,
+                                  a_EBfluxes[lev], ebflux_comp, ncomp, scale);
+      }
+   } else {                      // Fluxes are extensive
+      for (int lev = 0; lev <= finest_level; ++lev) {
+         extFluxDivergenceLevel(lev,*a_divergence[lev], div_comp, a_fluxes[lev], flux_comp,
+                                ncomp, scale);
+      }
+   }
+}
+
 void PeleLM::fluxDivergenceRD(const Vector<const MultiFab*> &a_state,
                               int state_comp,
                               const Vector<MultiFab*> &a_divergence,
                               int div_comp,
                               const Vector<Array<MultiFab*,AMREX_SPACEDIM> > &a_fluxes,
                               int flux_comp,
+                              const Vector<MultiFab* > &a_EBfluxes,
+                              int ebflux_comp,
                               int ncomp,
                               int intensiveFluxes,
                               const BCRec *state_bc_d,
@@ -43,6 +69,7 @@ void PeleLM::fluxDivergenceRD(const Vector<const MultiFab*> &a_state,
 {
     BL_PROFILE("PeleLM::fluxDivergenceRD()");
 #ifdef AMREX_USE_EB
+    int have_ebfluxes = (a_EBfluxes.empty()) ? 0 : 1;
     for (int lev = 0; lev <= finest_level; ++lev) {
         //----------------------------------------------------------------
         // Use a temporary MF to hold divergence before redistribution
@@ -50,7 +77,11 @@ void PeleLM::fluxDivergenceRD(const Vector<const MultiFab*> &a_state,
         MultiFab divTmp(grids[lev],dmap[lev],ncomp,nGrow_divTmp,MFInfo(),EBFactory(lev));
         divTmp.setVal(0.0);
         if (intensiveFluxes) {        // Fluxes are intensive -> need area scaling in div
-            intFluxDivergenceLevel(lev, divTmp, 0, a_fluxes[lev], flux_comp, ncomp, scale);
+            if (have_ebfluxes) {
+               intFluxDivergenceLevelEB(lev, divTmp, 0, a_fluxes[lev], flux_comp, a_EBfluxes[lev], ebflux_comp, ncomp, scale);
+            } else {
+               intFluxDivergenceLevel(lev, divTmp, 0, a_fluxes[lev], flux_comp, ncomp, scale);
+            }
         } else {                      // Fluxes are extensive
             extFluxDivergenceLevel(lev, divTmp, 0, a_fluxes[lev], flux_comp, ncomp, scale);
         }
@@ -141,9 +172,9 @@ void PeleLM::extFluxDivergenceLevel(int lev,
                }
             }
          });
[…]
