+++ Exec/RegTests/EB_FlowPastCylinder/EBUserDefined.H
@@ -27,33 +27,28 @@ EBUserDefined(const Geometry& /*geom*/,
 AMREX_GPU_DEVICE
 AMREX_FORCE_INLINE
 void
-setEBState(const amrex::Real * /*xEBface[AMREX_SPACEDIM]*/,
-           amrex::Real * /*s_ext[NVAR]*/,
+setEBState(const amrex::Real xEBface[AMREX_SPACEDIM],
+           amrex::Real s_ext[NVAR],
            const amrex::Real /*time*/,
            amrex::GeometryData const& /*geomdata*/,
            ProbParm const& /*prob_parm*/)
 {
-  // Provide the state on the EB centroid provided
-  // by xEBface
-  // Not the entire has to be specified. For instance, if using isothermal
-  // boundary, only s_ext[TEMP] will be used.
-  // Aborting here (message will not show on GPUs)
-  Abort("Default setEBState called. Copy and edit this file for your needs");
+  if ( xEBface[1] > 0.02 ) {
+     s_ext[TEMP] = 500.0;
+  } else {
+     s_ext[TEMP] = 300.0;
+  }
 }
 
 AMREX_GPU_DEVICE
 AMREX_FORCE_INLINE
 void
-setEBType(const amrex::Real * /*xEBface[AMREX_SPACEDIM]*/,
-          amrex::Real & /*EBflagType*/,
+setEBType(const amrex::Real xEBface[AMREX_SPACEDIM],
+          amrex::Real & EBflagType,
           amrex::GeometryData const& /*geomdata*/,
           ProbParm const& /*prob_parm*/)
 {
-  // Provide a factor applied on the EB face diffusion
-  // coefficient: 0.0 will turn off the diffusion (homogeneous Neumann BC)
-  // while 1.0 will let it unchanged. Intermediate values can be specified
-  // for partially Dirichlet BCs
-  Abort("Default setEBType called. Copy and edit this file for your needs");
+  EBflagType = 1.0;
 }
 #endif
 #endif
