+++ Source/PeleLM.H
@@ -330,6 +330,7 @@ class PeleLM : public amrex::AmrCore {
    // get species/enthalpy differential diffusion fluxes
    void computeDifferentialDiffusionFluxes(const PeleLM::TimeStamp &a_time,
                                            const amrex::Vector<amrex::Array<amrex::MultiFab*,AMREX_SPACEDIM>> &a_fluxes,
+                                           const amrex::Vector<amrex::MultiFab*> &a_EBfluxes,
                                            const amrex::Vector<amrex::Array<amrex::MultiFab*,AMREX_SPACEDIM>> &a_wbarfluxes,
                                            const amrex::Vector<amrex::Array<amrex::MultiFab*,AMREX_SPACEDIM>> &a_soretfluxes);
 
@@ -372,6 +373,7 @@ class PeleLM : public amrex::AmrCore {
 
    void deltaTIter_update(int a_dtiter,
                           const amrex::Vector<amrex::Array<amrex::MultiFab*,AMREX_SPACEDIM>> &a_fluxes,
+                          const amrex::Vector<amrex::MultiFab*> &a_ebfluxes,
                           const amrex::Vector<amrex::MultiFab const*> &a_Tsave,
                           std::unique_ptr<AdvanceDiffData> &diffData,
                           amrex::Real &a_deltaT_norm);
@@ -655,12 +657,24 @@ class PeleLM : public amrex::AmrCore {
                        int intensiveFluxes,
                        amrex::Real scale);
 
+   void fluxDivergence(const amrex::Vector<amrex::MultiFab*> &a_divergence,
+                       int div_comp,
+                       const amrex::Vector<amrex::Array<amrex::MultiFab*,AMREX_SPACEDIM> > &a_fluxes,
+                       int flux_comp,
+                       const amrex::Vector<amrex::MultiFab* > &a_EBfluxes,
+                       int ebflux_comp,
+                       int ncomp,
+                       int intensiveFluxes,
+                       amrex::Real scale);
+
    void fluxDivergenceRD(const amrex::Vector<const amrex::MultiFab*> &a_state,
                          int state_comp,
                          const amrex::Vector<amrex::MultiFab*> &a_divergence,
                          int div_comp,
                          const amrex::Vector<amrex::Array<amrex::MultiFab*,AMREX_SPACEDIM> > &a_fluxes,
                          int flux_comp,
+                         const amrex::Vector<amrex::MultiFab* > &a_EBfluxes,
+                         int ebflux_comp,
                          int ncomp,
                          int intensiveFluxes,
                          const amrex::BCRec *state_bc_d,
@@ -695,6 +709,16 @@ class PeleLM : public amrex::AmrCore {
                                int flux_comp,
                                int ncomp,
                                amrex::Real scale);
+
+   void intFluxDivergenceLevelEB(int lev,
+                                 amrex::MultiFab &a_divergence,
+                                 int div_comp,
+                                 const amrex::Array<amrex::MultiFab*,AMREX_SPACEDIM> &a_fluxes,
+                                 int flux_comp,
+                                 const amrex::MultiFab* a_EBfluxes,
+                                 int ebflux_comp,
+                                 int ncomp,
+                                 amrex::Real scale);
 
    void floorSpecies(const PeleLM::TimeStamp &a_time);
 
@@ -794,6 +818,17 @@ class PeleLM : public amrex::AmrCore {
 
    void extendSignedDistance( amrex::MultiFab *a_signDist,
                               amrex::Real a_extendFactor);
+
[…]

[…]
