### 2023-04-04 b90099b
Updated spray functions to use static unique pointers instead
