+++ Source/Spray/PeleLMSprayParticles.cpp
@@ -8,19 +8,6 @@
 using namespace amrex;
 
 namespace {
-//
-// Containers for the real "active" Particles
-//
-SprayParticleContainer* SprayPC = nullptr;
-//
-// Container for temporary, virtual Particles
-//
-SprayParticleContainer* VirtPC = nullptr;
-//
-// Container for temporary, ghost Particles
-//
-SprayParticleContainer* GhostPC = nullptr;
-
 SprayData sprayData;
 // Indices for spray source MultiFab
 int sprayMomSrcIndx = VELX;
@@ -35,16 +22,6 @@ Vector<int> spray_ghost_num;
 Vector<int> prev_state;
 Vector<int> prev_source;
 
-void
-RemoveParticlesOnExit()
-{
-  delete SprayPC;
-  SprayPC = nullptr;
-  delete GhostPC;
-  GhostPC = nullptr;
-  delete VirtPC;
-  VirtPC = nullptr;
-}
 bool mesh_regrid = true;
 std::string init_file;
 int init_function = 1;
@@ -52,6 +29,9 @@ int spray_verbose = 0;
 Real max_spray_cfl = 5.;
 } // namespace
 
+std::unique_ptr<SprayParticleContainer> PeleLM::SprayPC = nullptr;
+std::unique_ptr<SprayParticleContainer> PeleLM::VirtPC = nullptr;
+std::unique_ptr<SprayParticleContainer> PeleLM::GhostPC = nullptr;
 bool PeleLM::do_spray_particles = true;
 // momentum + density + fuel species + enthalpy
 int PeleLM::num_spray_src = AMREX_SPACEDIM + 2 + SPRAY_FUEL_NUM;
@@ -59,34 +39,16 @@ int PeleLM::num_spray_src = AMREX_SPACEDIM + 2 + SPRAY_FUEL_NUM;
 int PeleLM::write_spray_ascii_files = 0;
 int PeleLM::plot_spray_src = 0;
 
-SprayParticleContainer*
-PeleLM::theSprayPC()
-{
-  return SprayPC;
-}
-
-SprayParticleContainer*
-PeleLM::theVirtPC()
-{
-  return VirtPC;
-}
-
-SprayParticleContainer*
-PeleLM::theGhostPC()
-{
-  return GhostPC;
-}
-
 Real
 PeleLM::SprayEstDt()
 {
   Real estdt = 1.0e200;
-  if (!do_spray_particles || theSprayPC() == nullptr) {
+  if (!do_spray_particles || SprayPC == nullptr) {
     return estdt;
   }
   BL_PROFILE("PeleLM::SprayEstDt()");
   for (int lev = 0; lev <= finest_level; ++lev) {
-    Real estdt_lev = theSprayPC()->estTimestep(lev, max_spray_cfl);
+    Real estdt_lev = SprayPC->estTimestep(lev, max_spray_cfl);
     if (estdt_lev > 0. && estdt_lev < estdt) {
       estdt = estdt_lev;
     } else if (lev > 0) {
@@ -123,7 +85,7 @@ PeleLM::SpraySetup()
   // There must be at least as many fuel species in the spray as
   // there are species in the fluid
   if (SPRAY_FUEL_NUM > NUM_SPECIES) {
-    amrex::Abort("Cannot have more spray fuel species than fluid species");
+    Abort("Cannot have more spray fuel species than fluid species");
   }
   SprayParticleContainer::spraySetup(sprayData);
   // Component indices for conservative variables
@@ -143,15 +105,15 @@ void
 PeleLM::setupVirtualParticles(const int level)
 {
   BL_PROFILE("PeleLM::setupVirtualParticles()");
-  if (theSprayPC() != nullptr) {
[…]
