### 2023-04-05 f88e669
A set of small cosmetic changes to make the stdout more readable in the tutorial.
