+++ Docs/source/manual/Tutorials_FlowPastCyl.rst
@@ -69,22 +69,17 @@ Direct Numerical Simulations (DNS) is performed on a 12x4 :math:`cm^2` 2D comput
 The base grid is decomposed into 192x64 cells and up to 3 levels of refinement (although we will start with a single level).
 The refinement ratio between each level is set to 2.
 The maximum box size is fixed at 64, and the base (level 0) grid is composed of 3 boxes,
-as shown in Fig :numref:`fig:FPC_NumSetup`.
+as shown in :numref:`FPC_Domain`.
 
 Periodic boundary conditions are used in the transverse (:math:`y`) direction, while ``Inflow`` (dirichlet) and ``Outflow`` (neumann) boundary conditions are used in the main flow direction (:math:`x`). The flow goes from left to right.
 A cylinder of radius 0.0035 m is placed in the middle of the flow at (0.0:0.0).
 
-.. |FPC_a| image:: /images/tutorials/FPC_SetupSketch.png
-     :width: 100%
+.. figure:: images/tutorials/FPC_SetupSketch.png
+   :name: FPC_Domain
+   :align: center
+   :figwidth: 80%
 
-.. _fig:FPC_NumSetup:
-
-.. table:: Sketch of the computational domain with level 0 box decomposition.
-     :align: center
-
-     +---------+
-     | |FPC_a| |
-     +---------+
+   : Sketch of the computational domain with level 0 box decomposition.
 
 The geometry of the problem is specified in the first block of the ``input.2d-Re500``: ::
 
@@ -94,7 +89,7 @@ The geometry of the problem is specified in the first block of the ``input.2d-Re
    geometry.prob_lo     = -0.02 -0.02     # x_lo y_lo
    geometry.prob_hi     =  0.10  0.02     # x_hi y_hi
 
-The second block determines the boundary conditions. Note that `Interior` is used to indicate periodic boundary conditions. Refer to Fig :numref:`fig:FPC_NumSetup`: ::
+The second block determines the boundary conditions. Note that `Interior` is used to indicate periodic boundary conditions. Refer to :numref:`FPC_Domain`: ::
 
    #--------------------------BC FLAGS-----------------------------
    # Interior, Inflow, Outflow, Symmetry,
@@ -275,22 +270,17 @@ Whether you have used one or the other command, the computation finishes within 
    amrvis -a plt_?????
 
 
-.. |FPC_b| image:: /images/tutorials/FPC_Coarse_25ms.png
-     :width: 100%
+.. figure:: images/tutorials/FPC_Coarse_25ms.png
+   :name: FPC_Coarse
+   :align: center
+   :figwidth: 80%
 
-.. _fig:FPC_Coarse:
-
-.. table:: Contour plots of velocity components, vorticity, pressure and volume fraction at t = 25 ms on the coarse grid.
-     :align: center
-
-     +---------+
-     | |FPC_b| |
-     +---------+
+   : Contour plots of velocity components, vorticity, pressure and volume fraction at t = 25 ms on the coarse grid.
 
 At this point, you have established a flow with a large recirculation zone in the wake of the cylinder, but the flow has not yet fully transitioned to periodic vortex shedding.
-The flow is depicted in Fig :numref:`fig:FPC_Coarse` showing a few of the available contour plots at 25 ms. Note that the value of the fully covered cells is set to zero.
+The flow is depicted in :numref:`FPC_Coarse` showing a few of the available contour plots at 25 ms. Note that the value of the fully covered cells is set to zero.
 
-As can be seen from Fig :numref:`fig:FPC_Coarse`, the flow has not yet transitioned to the familiar Von-Karman alleys and two aspects of the current simulation can delay or even prevent the onset of vortex shedding:
[…]
