+++ Docs/source/manual/Tutorials_FlameSheet.rst
@@ -1,0 +1,523 @@
+.. highlight:: rst
+
+.. _sec:tutorialFlameSheet:
+
+Premixed flame sheet with harmonic perturbations
+================================================
+
+.. _sec:TUTO_FS::Intro:
+
+Introduction
+------------
+
+`PeleLMeX` primary objective is to enable simulation of reactive flows on platforms ranging
+from small personal computer to Exascale supercomputer. This short tutorial describes
+the case of a 2D laminar methane/hydrogen/air premixed flame, perturbed using harmonic fluctuations
+on the initial conditions.
+
+The goal of this tutorial is to demonstrate `PeleLMeX` basic controls when dealing with reactive simulations.
+This document provides step by step instruction reviewing how to set-up the domain and boundary conditions,
+and to construct an initial solution.
+
+..  _sec:TUTO_FS::PrepStep:
+
+Setting-up your environment
+---------------------------
+
+Getting a functioning environment in which to compile and run `PeleLMeX` is the first step of this tutorial.
+Follow the steps listed below to get to this point:
+
+#. The first step consist in getting `PeleLMeX` and its dependencies. To do so, use a recursive *git clone*: ::
+
+    git clone --recursive https://github.com/AMReX-Combustion/PeleLMeX.git
+
+#. Move into the Exec folder containing the ``FlameSheet``. To do so: ::
+
+    cd PeleLMeX/Exec/RegTests/FlameSheet
+
+Note that the makefile system is set up such that default paths are automatically set to the 
+submodules obtained with the recursive *git clone*, however the user can set its own dependencies
+in the `GNUmakefile` by updating the top-most lines as follows: ::
+
+       PELELMEX_HOME     = <path_to_PeleLMeX>
+       AMREX_HOME        = <path_to_MyAMReX>
+       AMREX_HYDRO_HOME  = <path_to_MyAMReXHydro>
+       PELE_PHYSICS_HOME = <path_to_MyPelePhysics>
+
+or directly through shell environement variables (using *bash* for instance): ::
+
+       export PELELMEX_HOME=<path_to_PeleLMeX>
+       export AMREX_HOME=<path_to_MyAMReX>
+       export AMREX_HYDRO_HOME=<path_to_MyAMReXHydro>
+       export PELE_PHYSICS_HOME=<path_to_MyPelePhysics>
+
+You're good to go !
+
+Case setup
+----------
+
+A `PeleLMeX` case folder generally contains a minimal set of files to enable compilation,
+provide user-defined function defining initial and boundary conditions, input file(s) and
+any additional files necessary for the simulation (solution of a Cantera 1D flame for instance).
+
+The following three files in particular are necessary: ::
+
+        pelelm_prob_parm.H
+        pelelm_prob.H
+        pelelm_prob.cpp
+
+The first file provides a C++ struct `ProbParm` containing the set of user-defined variables
+used during the simulation (value of inlet temperature, amplitude of the initial
+perturbation, ...). The `.cpp` file uses AMReX `ParmParse` to read these run-time
+parameter and initialize the `ProbParm` container values. Finally, `pelelm_prob.H`
+provides C++ kernels for the initial and boundary conditions and will be detailed
+later in this tutorial.
+
+The following review the content of the various files required for the flame sheet test case.
+User keys listed in the `input.2d-regt` file are reviewed and linked to the specific aspects
+of the simulation setup. To get additional information about the keywords discussed, 
+the reader is referred to :doc:`LMeXControls`.
+
+Geometry, grid and boundary conditions
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+This Direct Numerical Simulations (DNS) is performed on a 0.016x0.032 :math:`m^2` 2D computational domain,
[…]
