+++ Docs/source/manual/Tutorials_FlameSheet.rst
@@ -35,7 +35,7 @@ Follow the steps listed below to get to this point:
 
     cd PeleLMeX/Exec/RegTests/FlameSheet
 
-Note that the makefile system is set up such that default paths are automatically set to the 
+Note that the makefile system is set up such that default paths are automatically set to the
 submodules obtained with the recursive *git clone*, however the user can set its own dependencies
 in the `GNUmakefile` by updating the top-most lines as follows: ::
 
@@ -75,7 +75,7 @@ later in this tutorial.
 
 The following review the content of the various files required for the flame sheet test case.
 User keys listed in the `input.2d-regt` file are reviewed and linked to the specific aspects
-of the simulation setup. To get additional information about the keywords discussed, 
+of the simulation setup. To get additional information about the keywords discussed,
 the reader is referred to :doc:`LMeXControls`.
 
 Geometry, grid and boundary conditions
@@ -158,13 +158,13 @@ contained in `ProbParm` using AMReX's ParmParse: ::
     void PeleLM::readProbParm()
     {
        amrex::ParmParse pp("prob");
-    
+
        std::string type;
        pp.query("P_mean",   PeleLM::prob_parm->P_mean);
        pp.query("standoff", PeleLM::prob_parm->standoff);
        pp.query("pertmag",  PeleLM::prob_parm->pertmag);
        pp.query("pertlength",  PeleLM::prob_parm->pertlength);
-    
+
        PeleLM::pmf_data.initialize();
     }
 
@@ -208,7 +208,7 @@ version of the function reads:
 
 * Compute the harmonic perturbation.
 
-* Using ``standoff`` and the pertubation, use the ``PMF`` function to get cell-average temperature, mole fractions and 
+* Using ``standoff`` and the pertubation, use the ``PMF`` function to get cell-average temperature, mole fractions and
   velocity from the Cantera solution.
 
 * Use the data from the ``PMF`` to set the state array: velocities, density, rhoYs, rhoH and temperature. Relying on
@@ -216,9 +216,9 @@ version of the function reads:
 
 Some of the arguments of the `bcnormal` should now be familiar. The coordinates of the cell where the function
 is called are now directly passed into the function and the outgoing state vector is now ``s_ext``. The ``idir``
-and ``sgn`` `ints` can be used to easily determine on which domain face the function in called. Once again, the 
-state vector is extracted from the ``PMF`` function to match the operating conditions of the Cantera flame. This 
-function is only called in the direction/orientation where a Dirichlet boundary condition is imposed, i.e. the 
+and ``sgn`` `ints` can be used to easily determine on which domain face the function in called. Once again, the
+state vector is extracted from the ``PMF`` function to match the operating conditions of the Cantera flame. This
+function is only called in the direction/orientation where a Dirichlet boundary condition is imposed, i.e. the
 :math:`y`-low domain face here since the transverse direction is periodic and the outflow is an homogeneous
 Neumann for the state components.
 
@@ -229,7 +229,7 @@ Numerical parameters
 
 The ``PeleLM CONTROL`` block contains a few of the `PeleLMeX` algorithmic parameters. Many more
 unspecified parameters are relying on their default values which can be found in :doc:`LMeXControls`.
-Of particular interest are the ``peleLM.sdc_iterMax`` parameter controlling the number of 
+Of particular interest are the ``peleLM.sdc_iterMax`` parameter controlling the number of
 SDC iterations (see :doc:`Model` for more details on SDC in `PeleLMeX`) and the
 ``peleLM.num_init_iter`` one controlling the number of initial iteration the solver will do
 after initialization to obtain a consistent pressure and velocity field.
@@ -294,7 +294,7 @@ Checking the initial conditions
 -------------------------------
 
[…]
