### 2023-04-07 9fbe4ef
Add a Tutorial on the FlameSheet case (#193)
* Clean up the input.2d-regt file in FlameSheet regtest for
the tutorial.

* A set of small cosmetic changes to make the stdout more readable in the
tutorial.

* Minor tweaks for tutorial.

* Fix typo

* Acutally add the tutorial.

* Trailing whitespace.
