### 2023-04-10 c67e32c
Updated MP submodule and redid spray derived variable name access
