+++ Source/PeleLMPlot.cpp
@@ -112,7 +112,7 @@ void PeleLM::WritePlotFile() {
    ncomp += deriveEntryCount;
 #ifdef PELELM_USE_SPRAY
    if (do_spray_particles) {
-     ncomp += SprayParticleContainer::spray_derive_vars.size();
+     ncomp += SprayParticleContainer::NumDeriveVars();
    }
 #endif
 
@@ -204,12 +204,12 @@ void PeleLM::WritePlotFile() {
       }
    }
 #ifdef PELELM_USE_SPRAY
-   if (SprayParticleContainer::spray_derive_vars.size() > 0) {
+   if (SprayParticleContainer::NumDeriveVars() > 0) {
      // We need virtual particles for the lower levels
      setupVirtualParticles(0);
-     for (int ivar = 0; ivar < SprayParticleContainer::spray_derive_vars.size();
-          ivar++) {
-       plt_VarsName.push_back(SprayParticleContainer::spray_derive_vars[ivar]);
+     for (const auto& spray_derive_name :
+          SprayParticleContainer::DeriveVarNames()) {
+       plt_VarsName.push_back(spray_derive_name);
      }
    }
 #endif
@@ -293,15 +293,15 @@ void PeleLM::WritePlotFile() {
          cnt += mf->nComp();
       }
 #ifdef PELELM_USE_SPRAY
-      if (SprayParticleContainer::spray_derive_vars.size() > 0) {
-        int num_spray_derive = SprayParticleContainer::spray_derive_vars.size();
+      if (SprayParticleContainer::NumDeriveVars() > 0) {
+        const int num_spray_derive = SprayParticleContainer::NumDeriveVars();
         mf_plt[lev].setVal(0., cnt, num_spray_derive);
-        theSprayPC()->computeDerivedVars(mf_plt[lev], lev, cnt);
+        SprayPC->computeDerivedVars(mf_plt[lev], lev, cnt);
         if (lev < finest_level) {
           MultiFab tmp_plt(
             grids[lev], dmap[lev], num_spray_derive, 0, MFInfo(), Factory(lev));
           tmp_plt.setVal(0.);
-          theVirtPC()->computeDerivedVars(tmp_plt, lev, 0);
+          VirtPC->computeDerivedVars(tmp_plt, lev, 0);
           MultiFab::Add(mf_plt[lev], tmp_plt, 0, cnt, num_spray_derive, 0);
         }
         cnt += num_spray_derive;
@@ -356,8 +356,7 @@ void PeleLM::WritePlotFile() {
    if (do_spray_particles) {
      bool is_spraycheck = false;
      for (int lev = 0; lev <= finest_level; ++lev) {
-       theSprayPC()->SprayParticleIO(
-         lev, is_spraycheck, write_spray_ascii_files, plotfilename);
+       SprayPC->SprayParticleIO(lev, is_spraycheck, write_spray_ascii_files, plotfilename);
        // Remove virtual particles that were made for derived variables
        removeVirtualParticles(lev);
      }
@@ -470,8 +469,7 @@ void PeleLM::WriteCheckPointFile()
      int write_ascii = 0; // Not for checkpoints
      bool is_spraycheck = true;
      for (int lev = 0; lev <= finest_level; ++lev) {
-       theSprayPC()->SprayParticleIO(
-         lev, is_spraycheck, write_ascii, checkpointname);
+       SprayPC->SprayParticleIO(lev, is_spraycheck, write_ascii, checkpointname);
      }
    }
 #endif
