+++ Docs/source/manual/Tutorials_BFSFlame.rst
@@ -1,0 +1,366 @@
+.. highlight:: rst
+
+.. _sec:tutorialBFS:
+
+Backward facing step anchored premixed flame
+============================================
+
+.. _sec:TUTO_BFS::Intro:
+
+Introduction
+------------
+
+`PeleLMeX` primary objective is to enable simulations of reactive flows on platforms ranging
+from small personal computer to Exascale supercomputer. This tutorial describes the case
+of a 2D laminar premixed methane/air flame anchored behind a backward facing step.
+
+The goal of this tutorial is to introduce `PeleLMeX` users to more advanced reactive simulations setup as
+well as embedded boundaries.
+
+..  _sec:TUTO_BFS::PrepStep:
+
+Setting-up your environment
+---------------------------
+
+Getting a functioning environment in which to compile and run `PeleLMeX` is the first step of this tutorial.
+Follow the steps listed below to get to this point:
+
+#. The first step consist in getting `PeleLMeX` and its dependencies. To do so, use a recursive *git clone*: ::
+
+    git clone --recursive https://github.com/AMReX-Combustion/PeleLMeX.git
+
+#. Move into the Exec folder containing the ``FlameSheet``. To do so: ::
+
+    cd PeleLMeX/Exec/RegTests/FlameSheet
+
+Note that the makefile system is set up such that default paths are automatically set to the
+submodules obtained with the recursive *git clone*, however the user can set its own dependencies
+in the `GNUmakefile` by updating the top-most lines as follows: ::
+
+       PELELMEX_HOME     = <path_to_PeleLMeX>
+       AMREX_HOME        = <path_to_MyAMReX>
+       AMREX_HYDRO_HOME  = <path_to_MyAMReXHydro>
+       PELE_PHYSICS_HOME = <path_to_MyPelePhysics>
+
+or directly through shell environement variables (using *bash* for instance): ::
+
+       export PELELMEX_HOME=<path_to_PeleLMeX>
+       export AMREX_HOME=<path_to_MyAMReX>
+       export AMREX_HYDRO_HOME=<path_to_MyAMReXHydro>
+       export PELE_PHYSICS_HOME=<path_to_MyPelePhysics>
+
+You're good to go !
+
+Case setup
+----------
+
+A `PeleLMeX` case folder generally contains a minimal set of files to enable compilation,
+and the reader is referred to the FlameSheet tutorial :doc:`Tutorials_FlameSheet` for a
+more detailed description of `PeleLMeX` case setup. The case of interest for this 
+tutorial can be found in `PeleLMeX` Exec folder: ::
+
+    Exec/RegTests/EB_BackwardStepFlame
+
+Geometry, grid and boundary conditions
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+This simulation is performed on a 0.08x0.02 :math:`m^2` 2D computational domain,
+with the bottom left corner located at (-0.01:-0.01) and the top right corner at (0.07:0.01). The flow
+is primarily aligned with the :math:`x` direction with an ``Inflow`` (dirichlet) boundary on the :math:`x`-low 
+and ``Outflow`` (0-neumann) boundary on :math:`x`-high. No-slip wall condition are imposed the transverse direction.
+Finally a Cartesian coordinate system is used here. A overview of the computational domain is provided in :numref:`BFS_SetupGeom`.
+
+.. figure:: images/tutorials/BFS_SetupSketch.png
+   :name: BFS_SetupGeom
+   :align: center
+   :figwidth: 95%
+
+   : Setup of the computational domain for the backward facing step flame case.
+
+All of the geometrical information can be specified the first two blocks of the input file (`input.2d`): ::
+
[…]
