+++ Docs/source/manual/Tutorials_BFSFlame.rst
@@ -356,7 +356,8 @@ the step size is of the order of 10 :math:`\mu s`.
 Visualizing the `plt00250` file, we can see that the solution has eveolved, with a
 vortex propagating downstream along the flame surface, while intermediate species
 can be found. Looking at the heat release rate and the H2 mass fraction, we can see that
-the flame front is very poorly resolved.
+the flame front is very poorly resolved. The density along the isothermal EB also increased under
+the effect of the cold wall.
 
 .. figure:: images/tutorials/BFS_250steps.png
    :name: BFS_250steps
@@ -364,3 +365,233 @@ the flame front is very poorly resolved.
    :figwidth: 95%
 
    : Contour plots of density, H2 mass fraction, :math:`x`-velocity component and heat release rate after 250 steps.
+
+In order to illustrate one of `PeleLMeX` failure mode, we will now continue the simulation for 
+another 50 steps, starting from `chk00250`, while increasing the CFL number to 0.6. Update the 
+following keys in the input file: ::
+
+    #---------------------- Time Stepping CONTROL --------------------
+    amr.max_step = 300                          # Maximum number of time steps
+    ...
+    amr.cfl = 0.6                               # CFL number for hyperbolic system
+
+    ...
+    #---------------------- IO CONTROL -------------------------------
+    amr.restart = chk00250                      # Restart checkpoint file
+
+and restart the simulation: ::
+
+    mpirun -n 4 ./PeleLMeX2d.gnu.MPI.ex input.2d > logFail.dat &
+
+The simulation will proceed, with the step size progressively increasing due to the higher CFL
+and changes to the velocity field, but after ~30 steps `PeleLMeX` will fail with the following error: ::
+
+    ====================   NEW TIME STEP   ====================
+    Est. time step - Conv: 1.550410967e-05, divu: 5.712875519e-05
+    STEP [282] - Time: 0.001701680124, dt 1.550410967e-05
+      SDC iter [1]
+      SDC iter [2]
+    amrex::Abort::3::Aborting from CVODE !!!
+    SIGABRT
+    From CVODE: At t = 1.43329e-05, mxstep steps taken before reaching tout.
+
+The combination of large time step size and poor flame resolution leads to a very stiff chemical system,
+where fuel, oxidizer, intermediate species and heat are mixed within the cell averaged state representation
+associated with finite volume. The CVODE error clearly states that the internal sub-stepping of the ODE
+integrator was not able to integrate past 1.43329e-05. This is an indication that the CFL constraint
+is too loose compared to chemical stiffness, even though a implicit solve is performed. This generally occurs
+in laminar flows with coarse resolution, but could also occurs in midly turbulent flames with stiff
+chemical mechanisms. If you plan on pushing the simulation forward without adding refinement, it is advised
+to reduce CFL to smaller value.
+
+Refine the simulation
+---------------------
+
+Instead, let's add a first level of refinement and keep the CFL at a value of 0.6, while restarting again from
+`chk00250`. Enable AMR refinement by increasing the ``amr.max_level``: ::
+
+    #---------------------- AMR CONTROL ------------------------------
+    ...
+    amr.max_level       = 1                     # maximum level number allowed
+    ...
+
+And increase the maximum number of steps to 500: ::
+
+    #---------------------- Time Stepping CONTROL --------------------
+    amr.max_step = 500                          # Maximum number of time steps
+    ...
+
+Restart the simulation: ::
+
+    mpirun -n 4 ./PeleLMeX2d.gnu.MPI.ex input.2d > log1AMR.dat &
+
[…]
