+++ Docs/source/manual/Tutorials_BFSFlame.rst
@@ -56,7 +56,7 @@ Case setup
 
 A `PeleLMeX` case folder generally contains a minimal set of files to enable compilation,
 and the reader is referred to the FlameSheet tutorial :doc:`Tutorials_FlameSheet` for a
-more detailed description of `PeleLMeX` case setup. The case of interest for this 
+more detailed description of `PeleLMeX` case setup. The case of interest for this
 tutorial can be found in `PeleLMeX` Exec folder: ::
 
     Exec/RegTests/EB_BackwardStepFlame
@@ -66,7 +66,7 @@ Geometry, grid and boundary conditions
 
 This simulation is performed on a 0.08x0.02 :math:`m^2` 2D computational domain,
 with the bottom left corner located at (-0.01:-0.01) and the top right corner at (0.07:0.01). The flow
-is primarily aligned with the :math:`x` direction with an ``Inflow`` (dirichlet) boundary on the :math:`x`-low 
+is primarily aligned with the :math:`x` direction with an ``Inflow`` (dirichlet) boundary on the :math:`x`-low
 and ``Outflow`` (0-neumann) boundary on :math:`x`-high. No-slip wall condition are imposed the transverse direction.
 Finally a Cartesian coordinate system is used here. A overview of the computational domain is provided in :numref:`BFS_SetupGeom`.
 
@@ -115,11 +115,11 @@ All of those parameters are specified in the `AMR CONTROL` block: ::
    amr.max_grid_size   = 64               # maximum box size
 
 
-Finally, this case use Embedded Boundaries to represent the backward facing step. The EB is 
-defined as a box on the lower-left corner of the domain. For such an easy geometry, 
+Finally, this case use Embedded Boundaries to represent the backward facing step. The EB is
+defined as a box on the lower-left corner of the domain. For such an easy geometry,
 AMReX native simple CGS are sufficient. The box will extend from a point beyond
 the computational domain bottom left corner to (0.01:0.0). Because the intersection of the
-EB with the computational grid can lead to arbitrarily small cells, AMReX provides 
+EB with the computational grid can lead to arbitrarily small cells, AMReX provides
 ``eb2.small_volfrac`` to set a cell volume fraction limit below which a cell
 is considered fully covered. In the present simulation, we will treat the EB
 as an isothermal boundary, with control over the wall temperature described in the
@@ -191,10 +191,10 @@ geometries (not the case here) and to prescribe EB isothermal wall condition. It
 * ``setEBType()`` : even though ``peleLM.isothermal_EB=1`` is activated, the user can locally decide to use
   an adiabatic wall on part of the EB. To do so, this function takes in the EB face center coordinates
   and return a ``Real`` flag that should be set to 1.0 on isothermal areas and 0.0 on adiabatic areas. The
-  flag is later used to pre-multiply the thermal diffusivity effectively zeroing the thermal flux where the flag 
+  flag is later used to pre-multiply the thermal diffusivity effectively zeroing the thermal flux where the flag
   is 0.0.
 
-In the present case, we set the EB temperature to ``T_wall`` everywhere on the EB in ``setEBState()`` but 
+In the present case, we set the EB temperature to ``T_wall`` everywhere on the EB in ``setEBState()`` but
 the EB flag is only set to 1.0 on the vertical EB faces (:math:`x` normal) such that the top of the EB box
 is adiabatic.
 
@@ -237,7 +237,7 @@ The next few lines specify AMReX compilation options and compiler selection: ::
    USE_HIP         = FALSE
 
 It allows users to specify the number of spatial dimensions (2D), activate the compilation of the EB aware AMReX source code,
-trigger debug compilation and other AMReX options. The compiler (``gnu``) and the parallelism paradigm 
+trigger debug compilation and other AMReX options. The compiler (``gnu``) and the parallelism paradigm
[…]
