+++ Source/PeleLMDeriveFunc.H
@@ -98,6 +98,11 @@ void pelelm_derlambda (PeleLM* a_pelelm, const amrex::Box& bx, amrex::FArrayBox&
                        const amrex::Geometry& geom,
                        amrex::Real time, const amrex::Vector<amrex::BCRec> &bcrec, int level);
 
+void pelelm_derdmap (PeleLM* a_pelelm, const amrex::Box& bx, amrex::FArrayBox& derfab, int dcomp, int ncomp,
+                     const amrex::FArrayBox& statefab, const amrex::FArrayBox& reactfab, const amrex::FArrayBox& pressfab,
+                     const amrex::Geometry& geom,
+                     amrex::Real time, const amrex::Vector<amrex::BCRec> &bcrec, int level);
+
 amrex::Vector<std::string> pelelm_setuserderives();
 
 void pelelm_deruserdef (PeleLM* a_pelelm, const amrex::Box& bx, amrex::FArrayBox& derfab, int dcomp, int ncomp,
