### 2023-04-20 8f6ef98
fix some typos in BFS tutorial
