+++ Docs/source/manual/Tutorials_BFSFlame.rst
@@ -10,11 +10,11 @@ Backward facing step anchored premixed flame
 Introduction
 ------------
 
-`PeleLMeX` primary objective is to enable simulations of reactive flows on platforms ranging
-from small personal computer to Exascale supercomputer. This tutorial describes the case
+The primary objective of `PeleLMeX` is to enable simulations of reactive flows on platforms ranging
+from small personal computers to Exascale supercomputers. This tutorial describes the case
 of a 2D laminar premixed methane/air flame anchored behind a backward facing step.
 
-The goal of this tutorial is to introduce `PeleLMeX` users to more advanced reactive simulations setup as
+The goal of this tutorial is to introduce `PeleLMeX` users to more advanced reactive simulation setup as
 well as embedded boundaries.
 
 ..  _sec:TUTO_BFS::PrepStep:
@@ -25,13 +25,13 @@ Setting-up your environment
 Getting a functioning environment in which to compile and run `PeleLMeX` is the first step of this tutorial.
 Follow the steps listed below to get to this point:
 
-#. The first step consist in getting `PeleLMeX` and its dependencies. To do so, use a recursive *git clone*: ::
+#. The first step is to get `PeleLMeX` and its dependencies. To do so, use a recursive *git clone*: ::
 
     git clone --recursive https://github.com/AMReX-Combustion/PeleLMeX.git
 
-#. Move into the Exec folder containing the ``FlameSheet``. To do so: ::
+#. Move into the Exec folder containing the ``EB_BackwardStepFlame``. To do so: ::
 
-    cd PeleLMeX/Exec/RegTests/FlameSheet
+    cd PeleLMeX/Exec/RegTests/EB_BackwardStepFlame
 
 Note that the makefile system is set up such that default paths are automatically set to the
 submodules obtained with the recursive *git clone*, however the user can set its own dependencies
@@ -67,8 +67,8 @@ Geometry, grid and boundary conditions
 This simulation is performed on a 0.08x0.02 :math:`m^2` 2D computational domain,
 with the bottom left corner located at (-0.01:-0.01) and the top right corner at (0.07:0.01). The flow
 is primarily aligned with the :math:`x` direction with an ``Inflow`` (dirichlet) boundary on the :math:`x`-low
-and ``Outflow`` (0-neumann) boundary on :math:`x`-high. No-slip wall condition are imposed the transverse direction.
-Finally a Cartesian coordinate system is used here. A overview of the computational domain is provided in :numref:`BFS_SetupGeom`.
+and ``Outflow`` (0-neumann) boundary on :math:`x`-high. No-slip wall conditions are imposed the transverse direction.
+Finally a Cartesian coordinate system is used here. An overview of the computational domain is provided in :numref:`BFS_SetupGeom`.
 
 .. figure:: images/tutorials/BFS_SetupSketch.png
    :name: BFS_SetupGeom
@@ -94,13 +94,14 @@ All of the geometrical information can be specified the first two blocks of the 
 .. note::
     Note that when running 2D simulations, it is not necessary to specify entries for the third dimension.
 
-The base grid is decomposed into a 256x64 cells array with AMR initially not activated.
+The base grid is decomposed into a 256x64 cell array with AMR initially not activated.
 
 The refinement ratio between each level is set to 2 and `PeleLMeX` currently does not support
 refinement ratio of 4. Regrid operation will be performed every 5 steps. ``amr.n_error_buf`` specifies,
 for each level, the number of buffer cells used around the cell tagged for refinement, while ``amr.grid_eff``
 describes the grid efficiency, i.e. how much of the new grid contains tagged cells. Higher values lead
-to tighter grids around the tagged cells.
+to tighter grids around the tagged cells. For more information on how these parameters affect grid generation,
[…]
