+++ Docs/source/manual/Implementation.rst
@@ -1,0 +1,196 @@
+.. highlight:: rst
+
+.. _sec:code:
+
+The `PeleLMeX` code
+===================
+
+The following provides an overview of *PeleLMeX* source code, basic information on the data structure and is 
+useful for any user intending to use and/or do some development in the code.
+
+Overview of source code
+-----------------------
+
+*PeleLMeX* is based upon AMReX's [AmrCore](https://amrex-codes.github.io/amrex/docs_html/AmrCore.html) from which it inherits
+an AMR hierarchy data structure and basic regridding functionnalities. The code is entirely written in C++, with low level
+compute-intensive kernels implemented as lambda functions to seamlessly run on CPU and various GPU backends through AMReX
+high performance portatbility abstraction.
+
+The core of the algorithm is implementation in the ``advance()`` function which acts on all the levels concurrently.
+Projection operators and advection scheme functions are imported the [AMReX-Hydro library](https://amrex-codes.github.io/AMReX-Hydro) 
+while the core of the thermo-chemistry functionalities comes from [PelePhysics](https://amrex-combustion.github.io/PelePhysics/).
+Users are responsible for providing initial and boundary conditions in the local subfolder implementing their case, i.e. it is
+not possible to compile and run *PeleLMeX* without actually writting a few lines of codes. However, numerous example are provided
+in ``Exec/RegTests`` from which new users can pull for their new case.
+
+The source code contains a few dozen files, organized around the pieces of the algorithm and major functionalities:
+* ``PeleLMEvolve``: top level time advance loop, with IO/exit controls
+* ``PeleLMSetup``: setting up the simulation parameters, parsing the input file
+* ``PeleLMInit``: generating the initial solution from scratch or checkpoint file, performing initial projections/iteration(s)
+* ``PeleLMAdvance``: top level implementation of the time step algorithm
+* ``PeleLMProjection``: implement the various flavors of the nodal projection
+* ``PeleLMUmac``: implement the construction and projection of MAC-velocities
+* ``PeleLMAdvection``: functions to compute the explicit advection terms
+* ``PeleLMDiffusion``: functions to compute the diffusion terms, using the operators defined in ``DiffusionOp``
+* ``PeleLMReaction``: function using *PelePhysics* reactors to integrate the chemistry and linearized advection/diffusion
+* ``PeleLMPlot``: implementation of plotfile and checkpoint file IOs
+* ``PeleLMBC``: functions filling the ghost cells (at fine/fine, coarse/fine and domain boundaries)
+* ``PeleLMRegrid``: creating new AMR level or remaking modified AMR level during adaptive refinement
+* ``PeleLMTagging``: mark cells for refinement
+* ``PeleLM_K.H``: low-level kernel functions
+
+Data structure and containers
+-----------------------------
+
+AMReX 101
+^^^^^^^^^
+
+The basic AMReX`s data structure is the [MultiFab](https://amrex-codes.github.io/amrex/docs_html/Basics.html#fabarray-multifab-and-imultifab)
+(historically, multi Fortran Array Box (FAB)).
+Within the block-structured AMR approach of AMReX, the domain is decomposed into non-overlapping rectangular `boxes`,
+which can be assembled into a `boxArray`. Each AMR level has a `boxArray` providing the list of `boxes` of that level.
+The `boxes` are distributed accross the MPI ranks, the mapping of which is described by a `DistributionMap`. Given a 
+`boxArray` and a `DistributionMap`, one can define an actual data container (`boxes` are only lightweight descriptor
+of the geometrical rectangular object, containing bounds and centering information only), where each rank will
[…]
