+++ Docs/source/manual/Implementation.rst
@@ -2,8 +2,8 @@
 
 .. _sec:code:
 
-The `PeleLMeX` code
-===================
+Source code
+===========
 
 The following provides an overview of *PeleLMeX* source code, basic information on the data structure and is 
 useful for any user intending to use and/or do some development in the code.
@@ -11,19 +11,20 @@ useful for any user intending to use and/or do some development in the code.
 Overview of source code
 -----------------------
 
-*PeleLMeX* is based upon AMReX's [AmrCore](https://amrex-codes.github.io/amrex/docs_html/AmrCore.html) from which it inherits
+*PeleLMeX* is based upon AMReX's `AmrCore <https://amrex-codes.github.io/amrex/docs_html/AmrCore.html>`_ from which it inherits
 an AMR hierarchy data structure and basic regridding functionnalities. The code is entirely written in C++, with low level
 compute-intensive kernels implemented as lambda functions to seamlessly run on CPU and various GPU backends through AMReX
 high performance portatbility abstraction.
 
 The core of the algorithm is implementation in the ``advance()`` function which acts on all the levels concurrently.
-Projection operators and advection scheme functions are imported the [AMReX-Hydro library](https://amrex-codes.github.io/AMReX-Hydro) 
-while the core of the thermo-chemistry functionalities comes from [PelePhysics](https://amrex-combustion.github.io/PelePhysics/).
+Projection operators and advection scheme functions are imported the `AMReX-Hydro library <https://amrex-codes.github.io/AMReX-Hydro>`_
+while the core of the thermo-chemistry functionalities comes from `PelePhysics <https://amrex-combustion.github.io/PelePhysics/>`_ .
 Users are responsible for providing initial and boundary conditions in the local subfolder implementing their case, i.e. it is
 not possible to compile and run *PeleLMeX* without actually writting a few lines of codes. However, numerous example are provided
 in ``Exec/RegTests`` from which new users can pull for their new case.
 
 The source code contains a few dozen files, organized around the pieces of the algorithm and major functionalities:
+
 * ``PeleLMEvolve``: top level time advance loop, with IO/exit controls
 * ``PeleLMSetup``: setting up the simulation parameters, parsing the input file
 * ``PeleLMInit``: generating the initial solution from scratch or checkpoint file, performing initial projections/iteration(s)
@@ -45,7 +46,7 @@ Data structure and containers
 AMReX 101
 ^^^^^^^^^
 
-The basic AMReX`s data structure is the [MultiFab](https://amrex-codes.github.io/amrex/docs_html/Basics.html#fabarray-multifab-and-imultifab)
+The basic AMReX`s data structure is the `MultiFab <https://amrex-codes.github.io/amrex/docs_html/Basics.html#fabarray-multifab-and-imultifab>`_
 (historically, multi Fortran Array Box (FAB)).
 Within the block-structured AMR approach of AMReX, the domain is decomposed into non-overlapping rectangular `boxes`,
 which can be assembled into a `boxArray`. Each AMR level has a `boxArray` providing the list of `boxes` of that level.
@@ -55,10 +56,10 @@ of the geometrical rectangular object, containing bounds and centering informati
 allocate a FAB for the boxes it owns in the `boxArray`, resulting in a collection of FABs or a MultiFab, distributed
 accross the MPI ranks.
 
-To access the data in a MultiFab, one uses a [MFIter](https://amrex-codes.github.io/amrex/docs_html/Basics.html#mfiter-and-tiling)
+To access the data in a MultiFab, one uses a `MFIter <https://amrex-codes.github.io/amrex/docs_html/Basics.html#mfiter-and-tiling>`_
[…]
