### 2023-04-22 5cbf244
Tutorial Backward facing step flame (#196)
* Update/clean input file for Backfacing step tutorial.

* Remove unused code in EB_BFS case.

* Update base input file for Tutorial.

* Initiate a EB Backward facing step flame tutorial.

* Add missing images for tutorial.

* Cont' BFS Tuto.

* Wrap up BFS tutorial.

* fix some typos in BFS tutorial

* Address Bruce's comment.

---------

Co-authored-by: Bruce Perry <bruce.perry@nrel.gov>
