+++ Exec/RegTests/EB_BackwardStepFlame/pelelm_prob_parm.H
@@ -2,19 +2,18 @@
 #define PELELM_PROB_PARM_H_
 
 #include <AMReX_REAL.H>
-#include <AMReX_GpuMemory.H>
 
 using namespace amrex::literals;
 
 struct ProbParm
-    : amrex::Gpu::Managed
 {
    // Shared params
    amrex::Real T_mean = 298.0_rt;
    amrex::Real P_mean = 101325.0_rt;
    amrex::Real Y_fuel = 0.0445_rt;
-   amrex::Real Y_o2 = 0.223_rt;
+   amrex::Real Y_o2   = 0.223_rt;
+   amrex::Real T_hot  = 1800.0_rt;
+   amrex::Real Twall  = 300.0_rt;
    amrex::Real meanFlowMag = 0.0;
-   int  meanFlowDir = 1;
 };
 #endif
