+++ Exec/RegTests/EB_BackwardStepFlame/pelelm_prob.cpp
@@ -7,6 +7,9 @@ void PeleLM::readProbParm()
 
    pp.query("T_mean", prob_parm->T_mean);
    pp.query("P_mean", prob_parm->P_mean);
-   pp.query("meanFlowDir", prob_parm->meanFlowDir);
+   pp.query("Y_fuel", prob_parm->Y_fuel);
+   pp.query("Y_oxid", prob_parm->Y_o2);
+   pp.query("T_hot", prob_parm->T_hot);
+   pp.query("T_wall", prob_parm->Twall);
    pp.query("meanFlowMag", prob_parm->meanFlowMag);
 }
