+++ manual/html/_sources/Tutorials.rst.txt
@@ -13,3 +13,4 @@ This sections includes several self-contained tutorials.
 
    Tutorials_FlameSheet.rst
    Tutorials_FlowPastCyl.rst
+   Tutorials_BFSFlame.rst
