### 2023-04-24 fb3787d
Use LMeX version of the diagnostics only if requested.
