### 2023-04-27 59ab0e5
Pull active control out of LM. Data now live in a separate struct that can or can not be in the ProbParm.
