+++ Source/PeleLMFlowController.cpp
@@ -1,0 +1,403 @@
+#include "PeleLM.H"
+#include "pelelm_prob.H"
+
+#include <sys/stat.h>
+
+using namespace amrex;
+
+void
+PeleLM::initActiveControl()
+{
+    ParmParse pp("active_control");
+
+    pp.query("on",m_ctrl_active);
+
+    if (!m_ctrl_active) {
+        return;
+    }
+
+    Print() << " Initialization of active control \n";
+
+    pp.query("use_temp",m_ctrl_useTemp);
+    pp.query("temperature",m_ctrl_temperature);
+    pp.query("tau",m_ctrl_tauControl);
+    pp.query("v",m_ctrl_verbose);
+    pp.query("height",m_ctrl_h);
+    pp.query("changeMax",m_ctrl_changeMax);
+    pp.query("velMax",m_ctrl_velMax);
+    pp.query("pseudo_gravity",m_ctrl_pseudoGravity);
+    pp.query("flow_dir",m_ctrl_flameDir);
+    pp.query("AC_history",m_ctrl_AChistory);
+    pp.query("npoints_average",m_ctrl_NavgPts);
+    pp.query("method",m_ctrl_method);
+
+    // Active control checks
+    if ( m_ctrl_useTemp && (m_ctrl_temperature <= 0.0) ) {
+        amrex::Error("active_control.temperature MUST be set with active_control.use_temp = 1");
+    }
+
+    if ( m_ctrl_active && (m_ctrl_tauControl <= 0.0) ) {
+        amrex::Error("active_control.tau MUST be set when using active_control");
+    }
+
+    if ( m_ctrl_h <= 0.0 ) {
+        amrex::Error("active_control.height MUST be set when using active_control");
+    }
+
+    if ( m_ctrl_flameDir > 2 ) {
+        amrex::Error("active_control.flame_direction MUST be 0, 1 or 2 for X, Y and Z resp.");
+    }
+
+    // Initialize flow controller
+    if (hasFlowControllerData<ProbParm>::value == false) {
+        Abort("ProbParm doesn't have a FCData FlowControllerData member variable");
+    }
+
+    // Get FlowControllerData from ProbParm if it exists
+    FlowControllerData* fcdata_host{nullptr};
+    fcdata_host = getFCDataPtr(*prob_parm, hasFlowControllerData<ProbParm>{});
+
+    // Resize vector for temporal average
+    m_ctrl_time_pts.resize(m_ctrl_NavgPts+1,-1.0);
+    m_ctrl_velo_pts.resize(m_ctrl_NavgPts+1,-1.0);
+    m_ctrl_cntl_pts.resize(m_ctrl_NavgPts+1,-1.0);
+
+    // Extract data from BC: assumes flow comes in from lo side of ctrl_flameDir
+    ProbParm const* lprobparm = prob_parm_d;
+    pele::physics::PMF::PmfData::DataContainer const* lpmfdata = pmf_data.getDeviceData();
+
+    Gpu::DeviceVector<Real> s_ext_v(NVAR);
+    Real* s_ext_d = s_ext_v.data();
+    Real x[AMREX_SPACEDIM] = {AMREX_D_DECL(Geom(0).ProbLo(0),
+                                           Geom(0).ProbLo(1),
+                                           Geom(0).ProbLo(2))};
+    x[m_ctrl_flameDir] -= 1.0; 
+    const int ctrl_flameDir_l = m_ctrl_flameDir;
+    const amrex::Real time_l = -1.0;
+    const auto geomdata = Geom(0).data();
+
[…]
