### 2023-04-27 d027e92
Add a ctrl_int controlling how often AC is triggered.
