### 2023-04-28 e5864ff
Add load balancing control (#198)
* Redo load balancing (only).

* Update submodules.

* Describe additional derived vars.

* Add a brief description of the code base.

* A few fixes.

* Add control doc.

* Trailing whitespaces.

* Use LMeX version of the diagnostics only if requested.

* Add a couple of details to address Bruce's comments.
