+++ Docs/source/manual/Tutorials_TripleFlame.rst
@@ -1,0 +1,567 @@
+.. highlight:: rst
+
+.. _sec:tutorialTF:
+
+A simple triple flame
+=====================
+
+.. _sec:TUTO_TF::Intro:
+
+Introduction
+------------
+
+Laminar flames have the potential to reveal the fundamental structure of combustion 
+without the added complexities of turbulence. 
+They also aid in our understanding of the more complex turbulent flames. 
+Depending on the fuel involved and the flow configuration, the laminar flames can take on a number of interesting geometries. 
+For example, as practical combustion systems often operate in partially premixed mode,
+with one or more fuel injections, a wide range of fresh gas compositions can be observed; 
+and these conditions favor the appearance of edge flames, see Fig. :numref:`TripleFlameIntro`. 
+
+.. figure:: images/tutorials/TF_Intro.png
+   :name: TripleFlameIntro
+   :align: center
+   :figwidth: 95%
+
+   : Normalized heat release rate (top) and temperature (bottom) contours of two-dimensional (2D) laminar lifted flames of ethylene.
+
+Edge flames are composed of lean and rich premixed flame wings usually surrounding a central
+anchoring diffusion flame extending from a single point [PCI2007]_. Edge flames play
+an important role in flame stabilization, re-ignition and propagation.
+Simple fuels can exhibit up to three burning branches while diesel fuel, with a low temperature combustion mode, 
+can exhibit up to 5 branches.
+
+The goal of this tutorial is to setup a simple 2D laminar triple edge flame configuration with `PeleLMeX`. 
+This document provides step by step instructions to properly set-up the domain and boundary conditions, 
+construct an initial solution, and provides guidance on how to monitor and influence the initial transient to reach
+a final steady-state solution. 
+In a final Section, post-processing tools available in `PeleAnalysis` are used to extract information about 
+the triple flame.
+
+..  _sec:TUTO_TF::PrepStep:
+
+Setting-up your environment
+---------------------------
+
+Getting a functioning environment in which to compile and run `PeleLMeX` is the first step of this tutorial.
+Follow the steps listed below to get to this point:
+
+#. The first step is to get `PeleLMeX` and its dependencies. To do so, use a recursive *git clone*: ::
+
+    git clone --recursive https://github.com/AMReX-Combustion/PeleLMeX.git
+
+#. Move into the Exec folder containing the ``TripleFLame``. To do so: ::
+
+    cd PeleLMeX/Exec/RegTests/TripleFLame
+
+Note that the makefile system is set up such that default paths are automatically set to the
+submodules obtained with the recursive *git clone*, however the user can set its own dependencies
+in the `GNUmakefile` by updating the top-most lines as follows: ::
+
+       PELELMEX_HOME     = <path_to_PeleLMeX>
+       AMREX_HOME        = <path_to_MyAMReX>
+       AMREX_HYDRO_HOME  = <path_to_MyAMReXHydro>
+       PELE_PHYSICS_HOME = <path_to_MyPelePhysics>
+
+or directly through shell environement variables (using *bash* for instance): ::
+
+       export PELELMEX_HOME=<path_to_PeleLMeX>
+       export AMREX_HOME=<path_to_MyAMReX>
+       export AMREX_HYDRO_HOME=<path_to_MyAMReXHydro>
+       export PELE_PHYSICS_HOME=<path_to_MyPelePhysics>
+
+You're good to go !
+
+Numerical setup
+---------------
+
+In this section we review the content of the various input files for the Triple Flame test case. To get additional information about the keywords discussed, the user is referred to :doc:`LMeXControls`.
+
+Test case and boundary conditions
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
[…]
