+++ Docs/source/manual/Tutorials.rst
@@ -14,3 +14,4 @@ This sections includes several self-contained tutorials.
    Tutorials_FlameSheet.rst
    Tutorials_FlowPastCyl.rst
    Tutorials_BFSFlame.rst
+   Tutorials_TripleFlame.rst
