+++ Source/PeleLMSetup.cpp
@@ -376,6 +376,7 @@ void PeleLM::readParameters() {
    // -----------------------------------------
    pp.query("sdc_iterMax",m_nSDCmax);
    pp.query("floor_species",m_floor_species);
+   pp.query("dPdt_factor",m_dpdtFactor);
    pp.query("memory_checks",m_checkMem);
 
    // -----------------------------------------
