+++ Exec/RegTests/TripleFlame/pelelm_prob.H
@@ -16,16 +16,16 @@ void
 set_Y_from_Phi(ProbParm const &prob_parm,
                const amrex::Real &phi_in,
                amrex::Real Y[])
-{   
+{
    auto eos = pele::physics::PhysicsType::eos();
-   amrex::Real Xt[NUM_SPECIES] = {0.0};   
-   amrex::Real a = 0.0;   
-   if ( prob_parm.fuelID == CH4_ID ) {   
+   amrex::Real Xt[NUM_SPECIES] = {0.0};
+   amrex::Real a = 0.0;
+   if ( prob_parm.fuelID == CH4_ID ) {
       a = 2.0;
-   }      
+   }
    Xt[prob_parm.oxidID] = 1.0 / ( 1.0 + phi_in / a + 0.79 / 0.21 );
-   Xt[prob_parm.fuelID] = phi_in * Xt[prob_parm.oxidID] / a;   
-   Xt[prob_parm.bathID] = 1.0 - Xt[prob_parm.oxidID] - Xt[prob_parm.fuelID];   
+   Xt[prob_parm.fuelID] = phi_in * Xt[prob_parm.oxidID] / a;
+   Xt[prob_parm.bathID] = 1.0 - Xt[prob_parm.oxidID] - Xt[prob_parm.fuelID];
 
    eos.X2Y(Xt,Y);
 }
@@ -36,8 +36,8 @@ void
 set_Y_from_Ksi(ProbParm const &prob_parm,
                const amrex::Real &ksi_in,
                amrex::Real Y[])
-{   
-   amrex::Real phi = ksi_in / ( 1.0 - amrex::min(ksi_in,0.999999) ) * (1.0 - prob_parm.Zst ) / prob_parm.Zst;   
+{
+   amrex::Real phi = ksi_in / ( 1.0 - amrex::min(ksi_in,0.999999) ) * (1.0 - prob_parm.Zst ) / prob_parm.Zst;
 
    set_Y_from_Phi(prob_parm,phi,Y);
 }
@@ -75,25 +75,25 @@ void pelelm_initdata(int i, int j, int k,
 
     // Define air
     amrex::Real air_Y[NUM_SPECIES] = {0.0};
-    air_Y[prob_parm.bathID] = 0.767;  
-    air_Y[prob_parm.oxidID] = 0.233;  
+    air_Y[prob_parm.bathID] = 0.767;
+    air_Y[prob_parm.oxidID] = 0.233;
 
     // Get mixture fraction and corresponding composition
-    amrex::Real tanhval = 0.5 * ( 1.0 + std::tanh((x - prob_parm.midtanh) / prob_parm.widthtanh));  
+    amrex::Real tanhval = 0.5 * ( 1.0 + std::tanh((x - prob_parm.midtanh) / prob_parm.widthtanh));
     set_Y_from_Ksi(prob_parm,tanhval,massfrac);
 
-    amrex::Real rad = std::sqrt(  (x - prob_parm.splitx)      * (x - prob_parm.splitx) 
+    amrex::Real rad = std::sqrt(  (x - prob_parm.splitx)      * (x - prob_parm.splitx)
                                 + (amrex::min(y,y_lo) - y_lo) * (amrex::min(y,y_lo) - y_lo));
-    
-    amrex::Real Gaussian_T = std::exp( - (rad*rad) / (2.0 * ( Gauss_T_width * ( 1.0 + (y - y_lo) / 0.03)) 
-                                                          * ( Gauss_T_width * ( 1.0 + (y - y_lo) / 0.03)) ));  
+
+    amrex::Real Gaussian_T = std::exp( - (rad*rad) / (2.0 * ( Gauss_T_width * ( 1.0 + (y - y_lo) / 0.03))
[…]
