### 2023-05-01 327f0b1
Try fixing reduction for CUDA.
