### 2023-05-01 0811775
More fixing for CUDA.
