### 2023-05-01 8addb96
Use the m_ctrl_flamedir to apply pseudo-gravity term.
