+++ Source/PeleLMForces.cpp
@@ -136,11 +136,11 @@ void PeleLM::getVelForces(int lev,
    int is_incomp   = m_incompressible;
    Real rho_incomp = m_rho;
 
-   amrex::ParallelFor(bx, [=,grav=m_gravity, gp0=m_background_gp]
+   amrex::ParallelFor(bx, [=,grav=m_gravity, gp0=m_background_gp, ps_dir=m_ctrl_flameDir]
    AMREX_GPU_DEVICE(int i, int j, int k) noexcept
    {
       makeVelForce(i,j,k, is_incomp, rho_incomp,
-                   pseudo_gravity, a_time, grav, gp0, dV_control, dx,
+                   pseudo_gravity, ps_dir, a_time, grav, gp0, dV_control, dx,
                    vel, rho, rhoY, rhoh, temp, extMom, extRho, force);
    });
 }
