### 2023-05-02 66bc888
Remove unecessary TODO in Eos.
