+++ Source/PeleLMEos.cpp
@@ -155,7 +155,7 @@ void PeleLM::calcDivU(int is_init,
    }
 
    // fillPatch a_time divu to get properly filled ghost cells
-   int nGrowDivu = 1;   // TODO: need to make sure it's consistent across
+   int nGrowDivu = 1;
    for (int lev = 0; lev <= finest_level; ++lev) {
       Real time = getTime(lev,a_time);
       auto ldata_p = getLevelDataPtr(lev,a_time);
