### 2023-05-02 0075d01
Remove unused scale from DiffFluxes function.
