### 2023-05-02 194c10f
Minor tweaks in Plot functions.
