### 2023-05-02 7a1ffdf
Fix warnings and TODO in diffusion.
