+++ Source/PeleLMDiffusion.cpp
@@ -107,7 +107,6 @@ void PeleLM::computeDifferentialDiffusionTerms(const TimeStamp &a_time,
    }
 
    //----------------------------------------------------------------
-   // TODO simplify the following ...
    // Compute divergence/fill a_viscTerm
    // [0:NUM_SPECIES-1] Species           : \nabla \cdot \Flux_k
    // [NUM_SPECIES]     Temperature       : \nabla \cdot (-\lambda \nabla T)
@@ -123,21 +122,17 @@ void PeleLM::computeDifferentialDiffusionTerms(const TimeStamp &a_time,
                     GetVecOfArrOfPtrs(fluxes), 0,
                     {}, 0,
                     NUM_SPECIES, intensiveFluxes, bcRecSpec_d.dataPtr(), -1.0, m_dt);
+
    auto bcRecTemp = fetchBCRecArray(TEMP,1);
    auto bcRecTemp_d = convertToDeviceVector(bcRecTemp);
-   if (m_isothermalEB) {
-      fluxDivergenceRD(GetVecOfConstPtrs(getTempVect(a_time)), 0,
-                       diffTermVec, NUM_SPECIES,
-                       GetVecOfArrOfPtrs(fluxes), NUM_SPECIES,
-                       GetVecOfPtrs(EBfluxes), 0,
-                       1, intensiveFluxes, bcRecTemp_d.dataPtr(), -1.0, m_dt);
-   } else {
-      fluxDivergenceRD(GetVecOfConstPtrs(getTempVect(a_time)), 0,
-                       diffTermVec, NUM_SPECIES,
-                       GetVecOfArrOfPtrs(fluxes), NUM_SPECIES,
-                       {}, 0,
-                       1, intensiveFluxes, bcRecTemp_d.dataPtr(), -1.0, m_dt);
-   }
+   Vector<MultiFab*> EBFluxesVec = (m_isothermalEB) ? GetVecOfPtrs(EBfluxes)
+                                                    : Vector<MultiFab*>{};
+   fluxDivergenceRD(GetVecOfConstPtrs(getTempVect(a_time)), 0,
+                    diffTermVec, NUM_SPECIES,
+                    GetVecOfArrOfPtrs(fluxes), NUM_SPECIES,
+                    EBFluxesVec, 0,
+                    1, intensiveFluxes, bcRecTemp_d.dataPtr(), -1.0, m_dt);
+
    auto bcRecRhoH = fetchBCRecArray(RHOH,1);
    auto bcRecRhoH_d = convertToDeviceVector(bcRecRhoH);
    fluxDivergenceRD(GetVecOfConstPtrs(getRhoHVect(a_time)), 0,
@@ -153,8 +148,6 @@ void PeleLM::computeDifferentialDiffusionTerms(const TimeStamp &a_time,
    // Get the wbar term if appropriate
    if (!is_init && m_use_wbar) {
 #ifdef AMREX_USE_EB
-      auto bcRecSpec = fetchBCRecArray(FIRSTSPEC,NUM_SPECIES);
-      auto bcRecSpec_d = convertToDeviceVector(bcRecSpec);
       fluxDivergenceRD(GetVecOfConstPtrs(getSpeciesVect(a_time)), 0,
                        GetVecOfPtrs(diffData->Dwbar), 0,
                        GetVecOfArrOfPtrs(diffData->wbar_fluxes), 0,
@@ -169,8 +162,6 @@ void PeleLM::computeDifferentialDiffusionTerms(const TimeStamp &a_time,
    // Get the Soret term if appropriate
    if (!is_init && m_use_soret) {
 #ifdef AMREX_USE_EB
-      auto bcRecSpec = fetchBCRecArray(FIRSTSPEC,NUM_SPECIES);
[…]
