### 2023-05-02 699326b
Make vorticity, Qcrit and enstropy derived EB-aware
