+++ Source/PeleLMDeriveFunc.cpp
@@ -1,9 +1,14 @@
-#include <PeleLMDeriveFunc.H>
-#include <PeleLM_Index.H>
+#include "PeleLM_Index.H"
+#include "PeleLM.H"
+#include "PeleLM_K.H"
+#include "PeleLMDeriveFunc.H"
+
 #include <PelePhysics.H>
 #include <mechanism.H>
-#include <PeleLM.H>
-#include <PeleLM_K.H>
+#ifdef AMREX_USE_EB
+#include <AMReX_EBFabFactory.H>
+#include <AMReX_EBFArrayBox.H>
+#endif
 
 using namespace amrex;
 
@@ -35,7 +40,7 @@ void pelelm_dertemp (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int dco
 void pelelm_derheatrelease (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int dcomp, int ncomp,
                             const FArrayBox& statefab, const FArrayBox& reactfab, const FArrayBox& /*pressfab*/,
                             const Geometry& /*geom*/,
-                            Real /*time*/, const Vector<BCRec>& /*bcrec*/, int level)
+                            Real /*time*/, const Vector<BCRec>& /*bcrec*/, int /*level*/)
 
 {
     AMREX_ASSERT(derfab.box().contains(bx));
@@ -149,7 +154,7 @@ void pelelm_derrhomrhoy (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int
 //
 // Compute cell averaged pressure from nodes
 //
-void pelelm_deravgpress (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
+void pelelm_deravgpress (PeleLM* /*a_pelelm*/, const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
                          const FArrayBox& /*statefab*/, const FArrayBox& /*reactfab*/, const FArrayBox& pressfab,
                          const Geometry& /*geom*/,
                          Real /*time*/, const Vector<BCRec>& /*bcrec*/, int /*level*/)
@@ -177,7 +182,7 @@ void pelelm_deravgpress (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int
 //
 // Compute the velocity magnitude
 //
-void pelelm_dermgvel (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
+void pelelm_dermgvel (PeleLM* /*a_pelelm*/, const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
                       const FArrayBox& statefab, const FArrayBox& /*reactfab*/, const FArrayBox& /*pressfab*/,
                       const Geometry& /*geom*/,
                       Real /*time*/, const Vector<BCRec>& /*bcrec*/, int /*level*/)
@@ -199,13 +204,12 @@ void pelelm_dermgvel (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int dc
 //
 // Compute vorticity magnitude
 //
-void pelelm_dermgvort (PeleLM* a_pelelm, const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
+void pelelm_dermgvort (PeleLM* /*a_pelelm*/, const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
                        const FArrayBox& statefab, const FArrayBox& /*reactfab*/, const FArrayBox& /*pressfab*/,
                        const Geometry& geom,
[…]
