### 2023-05-02 7550a43
Switch to Array4<const Real>.
