+++ Source/DiffusionOp.cpp
@@ -1072,9 +1072,8 @@ void DiffusionTensorOp::diffuse_velocity (Vector<MultiFab*> const& a_vel,
        m_solve_op->setLevelBC(lev, a_vel[lev]);
    }
 
-   // TODO check Why * rho ??
    Vector<MultiFab> rhs(finest_level+1);
-   for (int lev = 0; lev <= finest_level; ++lev) {
+   for (int lev = 0; lev <= finest_level;DiffusionOp.cpp ++lev) {
        rhs[lev].define(a_vel[lev]->boxArray(),
                        a_vel[lev]->DistributionMap(), AMREX_SPACEDIM, 0);
 #ifdef AMREX_USE_OMP
@@ -1085,7 +1084,7 @@ void DiffusionTensorOp::diffuse_velocity (Vector<MultiFab*> const& a_vel,
            auto const& rhs_a = rhs[lev].array(mfi);
            auto const& vel_a = a_vel[lev]->const_array(mfi);
            auto const& rho_a = (have_density) ? a_density[lev]->const_array(mfi)
-                                              : a_vel[lev]->const_array(mfi);          // Dummy unused Array4
+                                              : Array4<const Real>{};
            amrex::ParallelFor(bx, AMREX_SPACEDIM, [=,rho_incomp=m_pelelm->m_rho,
                                                      is_incomp=m_pelelm->m_incompressible]
            AMREX_GPU_DEVICE (int i, int j, int k, int n) noexcept
