+++ Source/PeleLMEos.cpp
@@ -306,7 +306,7 @@ PeleLM::adjustPandDivU(std::unique_ptr<AdvanceAdvData> &advData)
     for (int lev = 0; lev <= finest_level; ++lev) {
         auto const& tma   = ThetaHalft[lev]->arrays();
         auto const& uma   = advData->mac_divu[lev].arrays();
-        amrex::ParallelFor(*ThetaHalft[lev], [=,dp0dt=m_dp0dt]
+        amrex::ParallelFor(*ThetaHalft[lev], [=]
         AMREX_GPU_DEVICE (int box_no, int i, int j, int k) noexcept
         {
            auto theta = tma[box_no];
