### 2023-05-02 46d29ad
Restore alternative method to estimate divU dt and add check.
