+++ Source/PeleLMTimestep.cpp
@@ -7,7 +7,7 @@ Real
 PeleLM::computeDt(int is_init,
                   const TimeStamp &a_time)
 {
-   BL_PROFILE_VAR("PeleLM::computeDt()", computeDt);
+   BL_PROFILE("PeleLM::computeDt()");
 
    Real estdt = 1.0e200;
 
@@ -153,8 +153,8 @@ PeleLM::estDivUDt(const TimeStamp &a_time) {
 
    Real estdt = 1.0e200;
 
-   // Note: only method 1 of PeleLM is available here
-   AMREX_ASSERT(m_divu_checkFlag==1);
+   // Note: only methods 1 & 2 of PeleLM are available here
+   AMREX_ASSERT(m_divu_checkFlag>=0 && m_divu_checkFlag<=2);
 
    for (int lev = 0; lev <= finest_level; ++lev) {
 
@@ -163,29 +163,95 @@ PeleLM::estDivUDt(const TimeStamp &a_time) {
 
       auto dtfac = m_divu_dtFactor;
       auto rhoMin = m_divu_rhoMin;
-      Real divu_dt = amrex::ReduceMin(*density, ldata_p->divu, 0,
-                                      [dtfac, rhoMin]
-      AMREX_GPU_HOST_DEVICE (Box const& bx, Array4<Real const> const& rho,
-                                            Array4<Real const> const& divu ) -> Real
-      {
-         using namespace amrex::literals;
-         const auto lo = amrex::lbound(bx);
-         const auto hi = amrex::ubound(bx);
-         amrex::Real dt = 1.e37_rt;
-         for       (int k = lo.z; k <= hi.z; ++k) {
-            for    (int j = lo.y; j <= hi.y; ++j) {
-               for (int i = lo.x; i <= hi.x; ++i) {
-                  Real dtcell = est_divu_dt_1(i, j, k, dtfac, rhoMin, rho, divu );
-                  dt = amrex::min(dt,dtcell);
-               }
-            }
-         }
-         return dt;
-      });
-      estdt = std::min(divu_dt,estdt);
+      if (m_divu_checkFlag == 1) {
+          Real divu_dt = amrex::ReduceMin(*density, ldata_p->divu, 0,
+                                          [dtfac, rhoMin]
+          AMREX_GPU_HOST_DEVICE (Box const& bx, Array4<Real const> const& rho,
+                                                Array4<Real const> const& divu ) -> Real
+          {
+             using namespace amrex::literals;
+             const auto lo = amrex::lbound(bx);
+             const auto hi = amrex::ubound(bx);
+             amrex::Real dt = 1.e37_rt;
+             for       (int k = lo.z; k <= hi.z; ++k) {
+                for    (int j = lo.y; j <= hi.y; ++j) {
+                   for (int i = lo.x; i <= hi.x; ++i) {
+                      Real dtcell = est_divu_dt_1(i, j, k, dtfac, rhoMin, rho, divu );
+                      dt = amrex::min(dt,dtcell);
+                   }
+                }
+             }
+             return dt;
+          });
+          estdt = std::min(divu_dt,estdt);
+      } else if (m_divu_checkFlag == 2) {
+          const auto& dxinv    = geom[lev].InvCellSizeArray();
+          std::unique_ptr<MultiFab> velo = std::make_unique <MultiFab>(ldata_p->state, amrex::make_alias, VELX, AMREX_SPACEDIM);
+          Real divu_dt = amrex::ReduceMin(*density, ldata_p->divu, *velo, 0,
+                                          [dtfac, rhoMin, dxinv]
+          AMREX_GPU_HOST_DEVICE (Box const& bx, Array4<Real const> const& rho,
[…]
