### 2023-05-02 e9bca18
Add a flag to set EB-covered values to zero in plotfile. Default is ON.
