### 2023-05-02 acd9a55
Add an imask in reaction -1 in EB-covered region, 1 otherwise. Change default EB-covered state.
