+++ Source/PeleLMEB.cpp
@@ -14,7 +14,6 @@ using namespace amrex;
 void PeleLM::makeEBGeometry()
 {
     BL_PROFILE("PeleLM::makeEBGeometry()");
-    // TODO extend
     int max_coarsening_level = 100;
     int req_coarsening_level = geom.size()-1;
 
@@ -95,10 +94,11 @@ void PeleLM::redistributeAofS(int a_lev,
                 // FluxRedistribute
                 Box gbx = bx;
 
-                if (m_adv_redist_type == "StateRedist")
-                  gbx.grow(3);
-                else if (m_adv_redist_type == "FluxRedist")
-                  gbx.grow(2);
+                if (m_adv_redist_type == "StateRedist") {
+                    gbx.grow(3);
+                } else if (m_adv_redist_type == "FluxRedist") {
+                    gbx.grow(2);
+                }
 
                 FArrayBox tmpfab(gbx, ncomp, The_Async_Arena());
                 Array4<Real> scratch = tmpfab.array(0);
@@ -106,7 +106,7 @@ void PeleLM::redistributeAofS(int a_lev,
                 {
                     amrex::ParallelFor(Box(scratch),
                     [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
-                    { scratch(i,j,k) = 1.;});   // TODO might want to test volfrac
+                    { scratch(i,j,k) = 1.;});
                 }
                 Redistribution::Apply( bx, ncomp, aofs_ar, divT_ar,
                                        a_state.const_array(mfi, state_comp), scratch, flag,
@@ -120,6 +120,57 @@ void PeleLM::redistributeAofS(int a_lev,
                 { aofs_ar(i,j,k,n) = divT_ar(i,j,k,n);});
             }
         }
+    }
+}
+
+void PeleLM::getCoveredIMask(int a_lev,
+                             iMultiFab &a_imask)
+{
+    const auto& ebfact = EBFactory(a_lev);
+    const auto& flags = ebfact.getMultiEBCellFlagFab();
+
+    if ((a_imask.boxArray() == flags.boxArray()) &&
+        (a_imask.DistributionMap() == flags.DistributionMap())) {
+
+#ifdef AMREX_USE_OMP
+#pragma omp parallel if (Gpu::notInLaunchRegion())
+#endif
+        for (MFIter mfi(a_imask,TilingIfNotGPU()); mfi.isValid(); ++mfi)
+        {
+            const Box& bx = mfi.tilebox();
+            const auto& flagarr = flags.const_array(mfi);
+            Array4<int> const& arr = a_imask.array(mfi);
+            AMREX_HOST_DEVICE_PARALLEL_FOR_3D ( bx, i, j, k,
+            {
+                if (flagarr(i,j,k).isCovered()) {
+                    arr(i,j,k) = -1;
+                } else {
+                    arr(i,j,k) = 1;
+                }
+            });
+        }
+    } else {
+
+        iMultiFab mask_tmp(flags.boxArray(),flags.DistributionMap(),1,0);
+
+#ifdef AMREX_USE_OMP
+#pragma omp parallel if (Gpu::notInLaunchRegion())
+#endif
+        for (MFIter mfi(mask_tmp,TilingIfNotGPU()); mfi.isValid(); ++mfi)
+        {
+            const Box& bx = mfi.tilebox();
+            const auto& flagarr = flags.const_array(mfi);
+            Array4<int> const& arr = mask_tmp.array(mfi);
+            AMREX_HOST_DEVICE_PARALLEL_FOR_3D ( bx, i, j, k,
+            {
+                if (flagarr(i,j,k).isCovered()) {
+                    arr(i,j,k) = -1;
+                } else {
[…]
