### 2023-05-03 1b167ad
Prints in checkDt only on CPU.
