### 2023-05-04 be5d1a2
Add active control capabilities out of PeleLM (#201)
* Pull active control out of LM. Data now live in a separate struct
that can or can not be in the ProbParm.

* Add calls to the AC.

* Restore the TripleFlame regtest.

* Set input for regtest.

* Add a ctrl_int controlling how often AC is triggered.

* Parse diffusionOp mg_max_order

* Fix EFIELD bug.

* Add the TripleFlame tutorial.

* ParmParse dPdt factor for debug purposes.

* Ran the tutorial with linear control, update default.

* Wrap up TF tutorial.

* Add active control input to LMeXControls.

* Remove trailing whitespaces.

* Incl
[…]
