+++ Exec/RegTests/TripleFlame/pelelm_prob.H
@@ -1,0 +1,192 @@
+#ifndef PELELM_PROB_H_
+#define PELELM_PROB_H_
+
+#include "PeleLM_Index.H"
+#include "pelelm_prob_parm.H"
+#include "PMFData.H"
+#include "PelePhysics.H"
+
+#include <AMReX_Geometry.H>
+#include <AMReX_FArrayBox.H>
+#include <AMReX_SPACE.H>
+
+AMREX_GPU_DEVICE
+AMREX_FORCE_INLINE
+void
+set_Y_from_Phi(ProbParm const &prob_parm,
+               const amrex::Real &phi_in,
+               amrex::Real Y[])
+{
+   auto eos = pele::physics::PhysicsType::eos();
+   amrex::Real Xt[NUM_SPECIES] = {0.0};
+   amrex::Real a = 0.0;
+   if ( prob_parm.fuelID == CH4_ID ) {
+      a = 2.0;
+   }
+   Xt[prob_parm.oxidID] = 1.0 / ( 1.0 + phi_in / a + 0.79 / 0.21 );
+   Xt[prob_parm.fuelID] = phi_in * Xt[prob_parm.oxidID] / a;
+   Xt[prob_parm.bathID] = 1.0 - Xt[prob_parm.oxidID] - Xt[prob_parm.fuelID];
+
+   eos.X2Y(Xt,Y);
+}
+
+AMREX_GPU_DEVICE
+AMREX_FORCE_INLINE
+void
+set_Y_from_Ksi(ProbParm const &prob_parm,
+               const amrex::Real &ksi_in,
+               amrex::Real Y[])
+{
+   amrex::Real phi = ksi_in / ( 1.0 - amrex::min(ksi_in,0.999999) ) * (1.0 - prob_parm.Zst ) / prob_parm.Zst;
+
+   set_Y_from_Phi(prob_parm,phi,Y);
+}
+
+AMREX_GPU_DEVICE
+AMREX_FORCE_INLINE
+void pelelm_initdata(int i, int j, int k,
+                     int /*is_incompressible*/,
+                     amrex::Array4<amrex::Real> const& state,
+                     amrex::Array4<amrex::Real> const& /*aux*/,
+                     amrex::GeometryData const& geomdata,
+                     ProbParm const& prob_parm,
+                     pele::physics::PMF::PmfData::DataContainer const * pmf_data)
+{
+    const amrex::Real* prob_lo = geomdata.ProbLo();
+    const amrex::Real* prob_hi = geomdata.ProbHi();
+    const amrex::Real* dx      = geomdata.CellSize();
+
+    AMREX_D_TERM(const amrex::Real x = prob_lo[0] + (i+0.5)*dx[0];,
+                 const amrex::Real y = prob_lo[1] + (j+0.5)*dx[1];,
+                 const amrex::Real z = prob_lo[2] + (k+0.5)*dx[2];);
+
+    AMREX_D_TERM(,
+                 const amrex::Real L_x = prob_hi[0] - prob_lo[0];,
+                 const amrex::Real L_y = prob_hi[1] - prob_lo[1]);
+
+    auto eos = pele::physics::PhysicsType::eos();
+    amrex::Real massfrac[NUM_SPECIES] = {0.0};
+
+    // Hot gaussian properties
+    amrex::Real y_lo = 0.01;
+    amrex::Real Gauss_maxT = 1740.0;
+    amrex::Real Gauss_T_width = 0.0012;
+    amrex::Real Gauss_Spec_width = 0.0020;
+
+    // Define air
+    amrex::Real air_Y[NUM_SPECIES] = {0.0};
[…]
