+++ Source/PeleLMInit.cpp
@@ -291,6 +291,10 @@ void PeleLM::initData() {
          }
       }
 
+      // Active control
+      int is_restart = 0;
+      activeControl(is_restart);
+
       //----------------------------------------------------------------
       // Initial pressure iterations
       initialIterations();
@@ -372,6 +376,10 @@ void PeleLM::initData() {
       m_resetCoveredMask = 1;
       resetCoveredMask();
       updateDiagnostics();
+
+      // Active control
+      int is_restart = 1;
+      activeControl(is_restart);
    }
 
 }
