### 2023-05-04 dcd53fb
Use ParseUserKey to handle domain BCs.
