+++ Source/PeleLMEvaluate.cpp
@@ -233,7 +233,7 @@ PeleLM::evaluateChemExtForces(const amrex::Vector<amrex::MultiFab*> &a_chemForce
    //----------------------------------------------------------------
    // Data for the advance, only live for the duration of the advance
    std::unique_ptr<AdvanceDiffData> diffData;
-   diffData.reset(new AdvanceDiffData(finest_level, grids, dmap, m_factory, m_nGrowAdv, m_use_wbar));
+   diffData.reset(new AdvanceDiffData(finest_level, grids, dmap, m_factory, m_nGrowAdv, m_use_wbar, m_use_soret));
    std::unique_ptr<AdvanceAdvData> advData;
    advData.reset(new AdvanceAdvData(finest_level, grids, dmap, m_factory, m_incompressible,
                                     m_nGrowAdv, m_nGrowMAC));
