+++ Source/PeleLM.H
@@ -291,7 +291,19 @@ class PeleLM : public amrex::AmrCore {
    */
    void checkRunParams();
 
+   /**
+   * \brief Compute load balancing cost estimate on a given level
+   * filling the PeleLM class LayoutData with default method
+   * \param a_lev target level
+   */
    void computeCosts(int a_lev);
+
+   /**
+   * \brief Compute load balancing cost estimate on a given level
+   * \param a_lev target level
+   * \param a_costs LayoutData holding the cost for each box
+   * \param a_costMethod cost method
+   */
    void computeCosts(int a_lev, amrex::LayoutData<amrex::Real> &a_costs,
                      int a_costMethod);
 
@@ -301,7 +313,7 @@ class PeleLM : public amrex::AmrCore {
    // NODAL PROJECTION
 
    /**
-   * \brief Initial velocity projection. The pressure and 
+   * \brief Initial velocity projection. The pressure and
    * pressure gradient are reset to zero after the projection.
    * Sigma is 1/rho instead of dt/rho.
    */
@@ -309,7 +321,7 @@ class PeleLM : public amrex::AmrCore {
 
    /**
    * \brief Initial velocity projection considering zero
-   * velocity divergence (no cell-centered RHS) to establish 
+   * velocity divergence (no cell-centered RHS) to establish
    * the isostatic pressure when using gravity
    * Sigma is 1/rho instead of dt/rho.
    */
@@ -362,25 +374,61 @@ class PeleLM : public amrex::AmrCore {
    //-----------------------------------------------------------------------------
    // MAC PROJECTION
 
+   /**
+   * \brief Re-create the MAC projector (reset the std::unique_ptr)
+   * if the circonstances require it
+   */
    void resetMacProjector();
 
-   // Predict face velocity using Godunov
+   /**
+   * \brief Use Godunov to obtain face-centered velocities
+   * from cell-centered ones
+   * \param advData container for the face-centered velocity umac
+   */
    void predictVelocity(std::unique_ptr<AdvanceAdvData> &advData);
 
-   // Create the MAC projection RHS
+   /**
+   * \brief Create MAC projection RHS by fillpatching divU
+   *  between old and new time
+   * \param advData container for mac_divu
+   */
    void createMACRHS(std::unique_ptr<AdvanceAdvData> &advData);
 
-   // Append the chi increment
+   /**
+   * \brief Create MAC projection RHS by fillpatching divU
+   *  between old and new time
+   * \param a_sdcIter current SDC iteration index
+   * \param a_time time stamp to access thermodynamic pressure
+   * \param advData container for chi and incremented mac_divu
+   */
    void addChiIncrement(int a_sdcIter,
                         const PeleLM::TimeStamp &a_time,
                         std::unique_ptr<AdvanceAdvData> &advData);
 
-   // Project face velocity
+   /**
+   * \brief Perform MAC projection to obtained divergence constrained
+   * face-centered (MAC) velocities
+   * \param a_time time stamp to access the density used in the projection
+   * \param advData container for the face-centered velocities umac
+   * \param a_divu velocity divergence constraint (can be empty)
+   */
    void macProject(const PeleLM::TimeStamp &a_time,
                    std::unique_ptr<AdvanceAdvData> &advData,
                    const amrex::Vector<amrex::MultiFab*> &a_divu);
 
-   // Umac ghost cells
+   /**
+   * \brief Fill the projected face velocities ghost faces, ensuring
+   * that the ghost cell velocities are constrained by the ghost cell
+   * divergence constraint
+   * \param a_lev target level
+   * \param a_nGrow number of ghost faces
+   * \param crse_geom pointer to the geometry of the next coarser level
[…]
