+++ Source/PeleLM.H
@@ -291,6 +291,9 @@ class PeleLM : public amrex::AmrCore {
    */
    void checkRunParams();
 
+   //-----------------------------------------------------------------------------
+   // REGRID / LOAD BALANCE
+
    /**
    * \brief Compute load balancing cost estimate on a given level
    * filling the PeleLM class LayoutData with default method
@@ -306,6 +309,17 @@ class PeleLM : public amrex::AmrCore {
    */
    void computeCosts(int a_lev, amrex::LayoutData<amrex::Real> &a_costs,
                      int a_costMethod);
+
+   /**
+   * \brief Create/update the DMap used for chemistry on all levels
+   */
+   void loadBalanceChem();
+
+   /**
+   * \brief Create/update the DMap used for chemistry on a given level
+   * \param a_lev level of interest
+   */
+   void loadBalanceChemLev(int a_lev);
 
    //-----------------------------------------------------------------------------
 
@@ -590,6 +604,16 @@ class PeleLM : public amrex::AmrCore {
                            std::unique_ptr<AdvanceAdvData> &advData,
                            std::unique_ptr<AdvanceDiffData> &diffData);
 
+   /**
+   * \brief Restore state, update enthalpy and compute fluxes after deltaT linear solve
+   * within the deltaT iteration scheme
+   * \param a_dtiter deltaT iteration index
+   * \param a_fluxes diffusion fluxes, with updated Fourier and differential species fluxes
+   * \param a_ebfluxes optional temperature fluxes at the EB
+   * \param a_Tsave temporary container to restore state temperature
+   * \param diffData container for diffusion forcing used to build RHS
+   * \param a_deltaT_norm norm of deltaT residual
+   */
    void deltaTIter_update(int a_dtiter,
                           const amrex::Vector<amrex::Array<amrex::MultiFab*,AMREX_SPACEDIM>> &a_fluxes,
                           const amrex::Vector<amrex::MultiFab*> &a_ebfluxes,
@@ -598,20 +622,39 @@ class PeleLM : public amrex::AmrCore {
                           amrex::Real &a_deltaT_norm);
 
 
-   // Viscous forces
+   /**
+   * \brief Compute the viscous force term in the momentum equation using AMReX
+   * TensorOp
+   * \param a_time state time for velocity, density and viscosity, either Old or New
+   * \param a_divtau outgoing divTau term
+   * \param use_density flag to scale divTau by 1/rho
+   * \param scale prefactor of divTau, used to weight by num. scheme property (e.g. 0.5 for 
+   * Crank Nicholson)
+   */
    void computeDivTau(const PeleLM::TimeStamp &a_time,
                       const amrex::Vector<amrex::MultiFab*> &a_divtau,
                       int use_density, amrex::Real scale = 1.0);
 
-   // Diffuse velocity
+   /**
+   * \brief Implicit diffusion of the of velocity field using AMReX
+   * TensorOp. Always act on New time data and include Crank Nicholson factor
+   */
    void diffuseVelocity();
 
-   // diffusion Linear operator BC
+   /**
+   * \brief Return LinOpBC for the scalar diffusion operator
+   * \param a_side AMReX orientaion
+   * \param a_bc BCRec for the scalar of interest
+   */
    amrex::Array<amrex::LinOpBCType,AMREX_SPACEDIM>
    getDiffusionLinOpBC(amrex::Orientation::Side a_side,
                        const amrex::BCRec &a_bc);
 
-   // diffusion Tensor operator BC
+   /**
+   * \brief Return AMReX_SPACEDIM LinOpBC vector for the velocity tensor operator
+   * \param a_side AMReX orientaion
[…]
