+++ Source/PeleLMEvaluate.cpp
@@ -18,7 +18,6 @@ void PeleLM::Evaluate() {
       if ( !itexists ) {
          amrex::Error("PeleLM::evaluate(): unknown variable: "+m_evaluatePlotVars[ivar]);
       }
-      Print() << " --> Evaluating " << m_evaluatePlotVars[ivar] << "\n";
       if ( derive_lst.canDerive(m_evaluatePlotVars[ivar]) ) {
          const PeleLMDeriveRec* rec = derive_lst.get(m_evaluatePlotVars[ivar]);
          ncomp += rec->numDerive();
@@ -45,10 +44,18 @@ void PeleLM::Evaluate() {
    }
 
    //----------------------------------------------------------------
+   // Prepare a few things if not restarting from a chkfile
+   if (m_restart_chkfile.empty()) {
+       m_nstep = 0;
+   }
+
+   //----------------------------------------------------------------
    // Fill the outgoing container
    int cnt = 0;
    for (int ivar = 0; ivar < m_evaluatePlotVarCount; ivar++ ) {
       int cntIncr = 0;
+
+      Print() << " --> Evaluating " << m_evaluatePlotVars[ivar] << "\n";
 
       // Evaluate function calls actual PeleLM::Evolve pieces and may require
       // the entire multi-level hierarchy
@@ -153,8 +160,12 @@ PeleLM::MLevaluate(const Vector<MultiFab *> &a_MFVec,
         }
         nComp = NUM_SPECIES+2;
     } else if ( a_var == "advTerm" ) {
-        // TODO
-        Abort("advTerm not available yet, soon hopefully");
+        Vector<std::unique_ptr<MultiFab> > aliasMF(finest_level+1);
+        for (int lev = 0; lev <= finest_level; ++lev) {
+            aliasMF[lev] = std::make_unique<MultiFab> (*a_MFVec[lev], amrex::make_alias, a_comp, NVAR-2);
+        }
+        evaluateAdvectionTerms(GetVecOfPtrs(aliasMF));
+        nComp = NVAR-2;
     } else if ( a_var == "chemTest" ) {
         // 'Old' chemical state (rhoYs + rhoH + T) and adv/diff forcing for chem. integration
         // Replicate most of the advance function
@@ -259,6 +270,10 @@ PeleLM::evaluateChemExtForces(const amrex::Vector<amrex::MultiFab*> &a_chemForce
    copyTransportOldToNew();
    copyDiffusionOldToNew(diffData);
 
+   // Compute the instantaneous reaction rate
+   // on Old state
+   computeInstantaneousReactionRate(getIRVect(),AmrOldTime);
+
    //----------------------------------------------------------------
    // Perform a single SDC iteration
    m_sdcIter = 1;
@@ -298,6 +313,98 @@ PeleLM::evaluateChemExtForces(const amrex::Vector<amrex::MultiFab*> &a_chemForce
    // Copy external forcing for chemistry into outgoing container
    for (int lev = 0; lev <= finest_level; ++lev) {
       MultiFab::Copy(*a_chemForces[lev],advData->Forcing[lev],0,0,NUM_SPECIES+1,0);
+   }
+
+   // Reset state
+   copyStateOldToNew();
+}
+
+void
+PeleLM::evaluateAdvectionTerms(const amrex::Vector<amrex::MultiFab*> &a_advTerms)
+{
+   //----------------------------------------------------------------
+   // Copy old <- new state
+   copyStateNewToOld(1);
+   copyPressNewToOld();
+
+   //----------------------------------------------------------------
+   // TIME
+   // Compute time-step size
+   m_dt = computeDt(0,AmrOldTime);
+
[…]
