### 2023-05-06 #5450 @editorialbot
[REVIEW]: PeleLMeX: An AMR Low Mach Number Reactive Flow Simulation Code Without Level Sub-cycling
**Submitting author:** @esclapez (<a href="http://orcid.org/0000-0002-2438-7292">Lucas Esclapez</a>)
**Repository:** https://github.com/AMReX-Combustion/PeleLMeX
**Branch with paper.md** (empty if default branch): JOSS_paper
**Version:** v23.11

@mameehan5, @xzz105, & @muellerm-pu, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
