### 2023-05-09 d84be9b
Add ReactEval readable data (#202)
* Add an Evaluate function to extract the state and chemistry forcing for
ReactEval.

* Fix missing soret argument.

* More description.

* More description.

* Trailing whitespace.

* Add a advTerm in evaluate mode.
