+++ Source/PeleLM.H
@@ -149,7 +149,9 @@ class PeleLM : public amrex::AmrCore {
    //-----------------------------------------------------------------------------
 
    //-----------------------------------------------------------------------------
-   // Gather State Data of a level into a struct
+   /**
+   * \brief Data container for the state data and long lived derived data
+   */
    struct LevelData {
       LevelData () = default;
       LevelData (amrex::BoxArray const& ba,
@@ -179,6 +181,9 @@ class PeleLM : public amrex::AmrCore {
 #endif
    };
 
+   /**
+   * \brief Data container for reaction data
+   */
    struct LevelDataReact {
       LevelDataReact () = default;
       LevelDataReact (const amrex::BoxArray &ba,
@@ -193,7 +198,9 @@ class PeleLM : public amrex::AmrCore {
    //-----------------------------------------------------------------------------
 
    //-----------------------------------------------------------------------------
-   // Gather Advance Datas into a structs
+   /**
+   * \brief Advance function data container for diffusion term
+   */
    struct AdvanceDiffData {
       AdvanceDiffData () = default;
       AdvanceDiffData (int a_finestLevel,
@@ -212,6 +219,9 @@ class PeleLM : public amrex::AmrCore {
       amrex::Vector<amrex::MultiFab> DT;        // Diffusion term of T (soret flux)
       amrex::Vector<amrex::Array<amrex::MultiFab,AMREX_SPACEDIM>> soret_fluxes;     // Soret flux correction
    };
+   /**
+   * \brief Advance function data container for advection term
+   */
    struct AdvanceAdvData {
       AdvanceAdvData () = default;
       AdvanceAdvData (int a_finestLevel,
@@ -236,13 +246,35 @@ class PeleLM : public amrex::AmrCore {
    //-----------------------------------------------------------------------------
    // INIT
 
-   // initData
+   /**
+   * \brief Top level initialization function
+   */
    void initData();
 
-   // initLevelData
+   /**
+   * \brief Fill the initial level data container
+   * using the user defined init function
+   */
    void initLevelData(int lev);
 
-   // initialIterations
+   /**
+   * \brief Fill the initial level data container
+   * using the data from a pltfile
+   * \param a_pltFile path to PeleLMeX plot file
+   */
+   void initLevelDataFromPlt(int a_lev, const std::string &a_pltFile);
+
+   /**
+   * \brief Project the initial solution velocity field,
+   * initializing the reaction term and providing a first
+   * estimate of the time step size
+   */
+   void projectInitSolution();
+
+   /**
+   * \brief Iterate the advance function to provide an
+   * initial pressure field.
+   */
    void initialIterations();
 
    // initFromGridFile
@@ -253,29 +285,81 @@ class PeleLM : public amrex::AmrCore {
                            amrex::Real time,
                            bool initial);
 
-   // checkRunParams
+   /**
+   * \brief Check the consistency of the run
+   * parameters.
+   */
    void checkRunParams();
 
+   //-----------------------------------------------------------------------------
+   // REGRID / LOAD BALANCE
+
+   /**
+   * \brief Compute load balancing cost estimate on a given level
+   * filling the PeleLM class LayoutData with default method
+   * \param a_lev target level
+   */
    void computeCosts(int a_lev);
+
+   /**
+   * \brief Compute load balancing cost estimate on a given level
+   * \param a_lev target level
+   * \param a_costs LayoutData holding the cost for each box
+   * \param a_costMethod cost method
+   */
    void computeCosts(int a_lev, amrex::LayoutData<amrex::Real> &a_costs,
                      int a_costMethod);
 
+   /**
[…]
