+++ Source/Diagnostics/DiagPDF.cpp
@@ -1,197 +1,0 @@
-#include "DiagPDF.H"
-#include "AMReX_MultiFabUtil.H"
-
-void
-DiagPDF::init(const std::string &a_prefix,
-              std::string_view a_diagName)
-{
-    DiagBase::init(a_prefix,a_diagName);
-
-    amrex::ParmParse pp(a_prefix);
-    pp.get("field_name",m_fieldName);
-    pp.get("nBins",m_nBins);
-    AMREX_ASSERT(m_nBins > 0);
-    pp.query("normalized",m_normalized);
-    pp.query("volume_weighted",m_volWeighted);
-
-    if (pp.countval("range")) {
-        amrex::Vector<amrex::Real> range{0.0};
-        pp.getarr("range", range, 0, 2);
-        m_lowBnd = std::min(range[0],range[1]);
-        m_highBnd = std::max(range[0],range[1]);
-        m_useFieldMinMax = false;
-    }
-}
-
-void
-DiagPDF::addVars(amrex::Vector<std::string> &a_varList) {
-    DiagBase::addVars(a_varList);
-    a_varList.push_back(m_fieldName);
-}
-
-void
-DiagPDF::prepare(int a_nlevels,
-                 const amrex::Vector<amrex::Geometry> &a_geoms,
-                 const amrex::Vector<amrex::BoxArray> &a_grids,
-                 const amrex::Vector<amrex::DistributionMapping> &a_dmap,
-                 const amrex::Vector<std::string> &a_varNames)
-{
-    if (first_time) {
-        DiagBase::prepare(a_nlevels, a_geoms, a_grids, a_dmap, a_varNames);
-        first_time = false;
-    }
-
-    m_geoms.resize(a_nlevels);
-    m_refRatio.resize(a_nlevels-1);
-    for (int lev = 0; lev < a_nlevels; lev++) {
-        m_geoms[lev] = a_geoms[lev];
-        if (lev > 0) {
-            m_refRatio[lev-1] = amrex::IntVect(static_cast<int>(a_geoms[lev-1].CellSize(0)/a_geoms[lev].CellSize(0)));
-        }
-    }
-
-}
-
-void
-DiagPDF::processDiag(int a_nstep,
-                     const amrex::Real &a_time,
-                     const amrex::Vector<const amrex::MultiFab*> &a_state,
-                     const amrex::Vector<std::string> &a_stateVar)
-{
-    // Set PDF range
-    int fieldIdx = getFieldIndex(m_fieldName,a_stateVar);
-    if (m_useFieldMinMax) {
-       m_lowBnd = MFVecMin(a_state,fieldIdx);
-       m_highBnd = MFVecMax(a_state,fieldIdx);
-    }
-    amrex::Real binWidth = (m_highBnd - m_lowBnd) / m_nBins;
-
-    // Data holders
-    amrex::Gpu::DeviceVector<amrex::Real> pdf_d(m_nBins,0.0);
-    amrex::Vector<amrex::Real> pdf(m_nBins,0.0);
-
-    // Populate the data from each level on each proc
-    for (int lev = 0; lev < a_state.size(); ++lev) {
-
-        // Make mask tagging fine-covered and filtered cells
-        amrex::iMultiFab mask;
-        if (lev == a_state.size()-1) {
-           mask.define(a_state[lev]->boxArray(), a_state[lev]->DistributionMap(),
-                       1, amrex::IntVect(0));
-           mask.setVal(1);
-        } else {
-           mask = amrex::makeFineMask(*a_state[lev], *a_state[lev+1], amrex::IntVect(0),
-                                      m_refRatio[lev],amrex::Periodicity::NonPeriodic(),
-                                      1, 0);
-        }
[…]
