+++ Source/PeleLMUtils.cpp
@@ -799,7 +799,7 @@ PeleLM::derive(const std::string &a_name,
       std::unique_ptr<MultiFab> statemf = fillPatchState(lev, a_time, m_nGrowState);
       // Get pressure: TODO no fillpatch for pressure just yet, simply get new state
       auto ldata_p = getLevelDataPtr(lev,AmrNewTime);
-      auto ldataR_p = getLevelDataReactPtr(lev);
+      std::unique_ptr<MultiFab> reactmf = fillPatchReact(lev, a_time, nGrow);
       auto stateBCs = fetchBCRecArray(VELX,NVAR);
 #ifdef AMREX_USE_OMP
 #pragma omp parallel if (Gpu::notInLaunchRegion())
@@ -809,7 +809,7 @@ PeleLM::derive(const std::string &a_name,
           const Box& bx = mfi.growntilebox(nGrow);
           FArrayBox& derfab = (*mf)[mfi];
           FArrayBox const& statefab = (*statemf)[mfi];
-          FArrayBox const& reactfab = (m_incompressible) ? ldata_p->press[mfi] : ldataR_p->I_R[mfi];
+          FArrayBox const& reactfab = (m_incompressible) ? ldata_p->press[mfi] : (*reactmf)[mfi];
           FArrayBox const& pressfab = ldata_p->press[mfi];
           rec->derFunc()(this, bx, derfab, 0, rec->numDerive(), statefab, reactfab, pressfab, geom[lev], a_time, stateBCs, lev);
       }
@@ -818,10 +818,11 @@ PeleLM::derive(const std::string &a_name,
       int idx = stateVariableIndex(a_name);
       std::unique_ptr<MultiFab> statemf = fillPatchState(lev, a_time, nGrow);
       MultiFab::Copy(*mf,*statemf,idx,0,1,nGrow);
-   } else {                                       // This is a reaction variable, just alias, no time interpolation
-      AMREX_ASSERT(nGrow <= getLevelDataReactPtr(lev)->I_R.nGrow());
+   } else {                                       // This is a reaction variable
+      mf.reset(new MultiFab(grids[lev], dmap[lev], 1, nGrow, MFInfo(), Factory(lev)));
       int idx = reactVariableIndex(a_name);
-      mf.reset(new MultiFab(getLevelDataReactPtr(lev)->I_R, amrex::make_alias, idx, 1));
+      std::unique_ptr<MultiFab> reactmf = fillPatchReact(lev, a_time, nGrow);
+      MultiFab::Copy(*mf,*reactmf,idx,0,1,nGrow);
    }
 
    return mf;
@@ -854,7 +855,7 @@ PeleLM::deriveComp(const std::string &a_name,
       std::unique_ptr<MultiFab> statemf = fillPatchState(lev, a_time, m_nGrowState);
       // Get pressure: TODO no fillpatch for pressure just yet, simply get new state
       auto ldata_p = getLevelDataPtr(lev,AmrNewTime);
-      auto ldataR_p = getLevelDataReactPtr(lev);
+      std::unique_ptr<MultiFab> reactmf = fillPatchReact(lev, a_time, nGrow);
       auto stateBCs = fetchBCRecArray(VELX,NVAR);
 
       // Temp MF for all the derive components
@@ -867,7 +868,7 @@ PeleLM::deriveComp(const std::string &a_name,
           const Box& bx = mfi.growntilebox(nGrow);
           FArrayBox& derfab = derTemp[mfi];
           FArrayBox const& statefab = (*statemf)[mfi];
-          FArrayBox const& reactfab = (m_incompressible) ? ldata_p->press[mfi] : ldataR_p->I_R[mfi];
[…]
