### 2023-05-10 051d58f
Resolve/remove TODOs throughout the code base (#204)
* Remove unecessary TODO in Eos.

* Extra ;

* Assert in low level projection function.

* Remove unused scale from DiffFluxes function.

* Minor tweaks in Plot functions.

* Consistent divU number of ghost cells.

* Fix warnings and TODO in diffusion.

* Make sure to use Factory when defining temporary MFs for derive var
and such.

* Make vorticity, Qcrit and enstropy derived EB-aware

* Switch to Array4<const Real>.

* Remove deprecated TODO.

* Odd paste typo.

* Fix warning in Eos.

* Restore alternative method to estimate divU dt and ad
[…]
