+++ Source/Diagnostics/DiagFramePlane.cpp
@@ -1,724 +1,0 @@
-#include "DiagFramePlane.H"
-#include "AMReX_VisMF.H"
-#include <AMReX_FPC.H>
-#include "AMReX_PlotFileUtil.H"
-#include <regex>
-#include <stdio.h>
-
-
-void
-printLowerDimIntVect(std::ostream &a_File,
-                     const amrex::IntVect &a_IntVect,
-                     int skipDim)
-{
-    int doneDim = 0;
-    a_File << '(';
-    for (int idim = 0; idim < AMREX_SPACEDIM; ++idim) {
-        if (idim != skipDim) {
-            a_File << a_IntVect[idim];
-            doneDim ++;
-            if (doneDim < AMREX_SPACEDIM-1) a_File << ",";
-        }
-    }
-    a_File << ')';
-}
-
-void
-printLowerDimBox(std::ostream &a_File,
-                 const amrex::Box &a_box,
-                 int skipDim)
-{
-    a_File << '(';
-    printLowerDimIntVect(a_File, a_box.smallEnd(),skipDim);
-    a_File << ' ';
-    printLowerDimIntVect(a_File, a_box.bigEnd(),skipDim);
-    a_File << ' ';
-    printLowerDimIntVect(a_File, a_box.type(),skipDim);
-    a_File << ')';
-}
-
-void
-DiagFramePlane::init(const std::string &a_prefix,
-                     std::string_view a_diagName)
-{
-    DiagBase::init(a_prefix, a_diagName);
-
-    if (m_filters.size() != 0) {
-        amrex::Print() << " Filters are not available on DiagFramePlane and will be discarded \n";
-    }
-
-    amrex::ParmParse pp(a_prefix);
-
-    // Outputed variables
-    int nOutFields = pp.countval("field_names");
-    AMREX_ASSERT(nOutFields > 0);
-    m_fieldNames.resize(nOutFields);
-    m_fieldIndices_d.resize(nOutFields);
-    for (int f{0}; f<nOutFields; ++f) {
-        pp.get("field_names", m_fieldNames[f],f);
-    }
-
-    // Plane normal
-    pp.get("normal", m_normal);
-    AMREX_ASSERT(m_normal>=0 && m_normal<AMREX_SPACEDIM);
-
-    // Plane center
-    amrex::Vector<amrex::Real> center;
-    pp.getarr("center",center,0,pp.countval("center"));
-    if (center.size() == AMREX_SPACEDIM) {
-        for (int idim = 0; idim < AMREX_SPACEDIM; idim++) {
-            m_center[idim] = center[idim];
-        }
-    } else if (center.size() == 1) {
-        m_center[m_normal] = center[0];
-    }
-
-    // Interpolation
-    std::string intType = "Quadratic";
-    pp.query("interpolation", intType);
-    if (intType == "Linear") {
-        m_interpType = Linear;
-    } else if (intType == "Quadratic") {
-        m_interpType = Quadratic;
-    } else {
-        amrex::Abort("Unknown interpolation type for "+a_prefix);
-    }
-}
-
-void
-DiagFramePlane::addVars(amrex::Vector<std::string> &a_varList) {
-    DiagBase::addVars(a_varList);
-    for (const auto& v : m_fieldNames) {
-        a_varList.push_back(v);
-    }
-}
-
-void
-DiagFramePlane::prepare(int a_nlevels,
-                        const amrex::Vector<amrex::Geometry> &a_geoms,
-                        const amrex::Vector<amrex::BoxArray> &a_grids,
-                        const amrex::Vector<amrex::DistributionMapping> &a_dmap,
-                        const amrex::Vector<std::string> &a_varNames)
-{
-    if (first_time) {
-        // Store the level0 geometry
-        auto initDomain  = a_geoms[0].Domain();
[…]
