### 2023-05-11 bbed4be
A set of small fixes to address warnings from DEBUG/CodeQL
