### 2023-05-11 7b457a6
Tidy and fixes (#205)
* Remove AMReX_buildinfo from CodeQL.

* A set of small fixes to address warnings from DEBUG/CodeQL
