+++ Source/PeleLMSetup.cpp
@@ -17,7 +17,6 @@
 using namespace amrex;
 
 static Box the_same_box (const Box& b)    { return b;                }
-static Box grow_box_by_one (const Box& b) { return amrex::grow(b,1); }
 static Box grow_box_by_two (const Box& b) { return amrex::grow(b,2); }
 
 void PeleLM::Setup() {
