+++ .github/workflows/codeql.yml
@@ -52,6 +52,7 @@ jobs:
             -**/Submodules/sundials/**
             -**/Submodules/amrex/**
             -**/Submodules/AMReX-Hydro/**
+            -**/pelelmex/**
           input: sarif-results/cpp.sarif
           output: sarif-results/cpp.sarif
       - name: Upload CodeQL sarif
