### 2023-05-16 51a7027
Generalize functions to get state at the EB.
