+++ Source/PeleLMEB.cpp
@@ -1,7 +1,8 @@
 #include <PeleLM.H>
+#include <PeleLM_K.H>
+#include "EBUserDefined.H"
 #include <PeleLMUtils.H>
 #include <pelelm_prob.H>
-#include <EBUserDefined.H>
 
 #ifdef AMREX_USE_EB
 #include <AMReX_EB2.H>
@@ -429,6 +430,79 @@ void PeleLM::getEBDistance(int a_lev,
     }
 }
 
+Vector<std::unique_ptr<MultiFab>>
+PeleLM::getEBState(int first_comp, int ncomp,
+                   const PeleLM::TimeStamp &a_time)
+{
+    AMREX_ASSERT(first_comp >= VELX);
+    AMREX_ASSERT(first_comp+ncomp <= NVAR);
+    Vector<std::unique_ptr<MultiFab> > r;
+    r.reserve(finest_level+1);
+    for (int lev = 0; lev <= finest_level; ++lev) {
+        r.push_back(std::make_unique<MultiFab> (grids[lev], dmap[lev], ncomp, 0, MFInfo(), Factory(lev)));
+        getEBState(lev, getTime(lev,a_time), *r[lev], first_comp, ncomp);
+    }
+    return r;
+}
+
+std::unique_ptr<MultiFab>
+PeleLM::getEBState(int a_lev,
+                   int first_comp, int ncomp,
+                   const PeleLM::TimeStamp &a_time)
+{
+    AMREX_ASSERT(first_comp >= VELX);
+    AMREX_ASSERT(first_comp+ncomp <= NVAR);
+    std::unique_ptr<MultiFab> r = std::make_unique<MultiFab> (grids[a_lev], dmap[a_lev], ncomp, 0, MFInfo(), Factory(a_lev));
+    getEBState(a_lev, getTime(a_lev,a_time), *r, first_comp, ncomp);
+
+    return r;
+}
+
+FArrayBox
+PeleLM::getEBState(MFIter const& a_mfi,
+                   int a_lev,
+                   int first_comp, int ncomp,
+                   const PeleLM::TimeStamp &a_time)
+{
+    AMREX_ASSERT(first_comp >= VELX);
+    AMREX_ASSERT(first_comp+ncomp <= NVAR);
+
+    ProbParm const* lprobparm = prob_parm_d;
+    const auto geomdata = geom[a_lev].data();
+    auto time = getTime(a_lev,a_time);
+
+    const auto bx = a_mfi.tilebox();
+    FArrayBox r(bx,ncomp,The_Async_Arena());
+    auto ebscal_arr = r.array();
+    const auto &ebfact = EBFactory(a_lev);
+    auto const& flagfab = ebfact.getMultiEBCellFlagFab()[a_mfi];
+    if ( flagfab.getType(bx) == FabType::singlevalued ) {
+         auto const& flag    = flagfab.const_array();
+         AMREX_D_TERM( const auto& ebfc_x = ebfact.getFaceCent()[0]->const_array(a_mfi);,
+                       const auto& ebfc_y = ebfact.getFaceCent()[1]->const_array(a_mfi);,
+                       const auto& ebfc_z = ebfact.getFaceCent()[2]->const_array(a_mfi););
+         amrex::ParallelFor(bx, [=]
+         AMREX_GPU_DEVICE (int i, int j, int k) noexcept
+         {
+            // Regular/covered cells -> 0.0
+            if ( flag(i,j,k).isCovered() ||
+                 flag(i,j,k).isRegular()) {
+               for (int n = 0; n < ncomp; n++) {
+                  ebscal_arr(i,j,k,n) = 0.0;
+               }
+            } else { // cut-cells
[…]
