+++ Source/PeleLMUtils.cpp
@@ -558,6 +558,111 @@ void PeleLM::advFluxDivergence(int a_lev,
     }
 }
 
+#ifdef AMREX_USE_EB
+void PeleLM::advFluxDivergence(int a_lev,
+                               MultiFab &a_divergence, int div_comp,
+                               MultiFab &a_divu,
+                               const Array<const MultiFab*,AMREX_SPACEDIM> &a_fluxes, int flux_comp,
+                               const Array<const MultiFab*,AMREX_SPACEDIM> &a_faceState, int face_comp,
+                               const MultiFab *a_EBvelocity,
+                               const MultiFab *a_EBvalue,
+                               int ncomp,
+                               int const* l_conserv_d,
+                               const Geometry &a_geom,
+                               amrex::Real scale,
+                               bool fluxes_are_area_weighted)
+{
+    BL_PROFILE("PeleLM::advFluxDivergence()");
+
+    AMREX_ASSERT(a_divergence.nComp() >= div_comp+ncomp);
+    AMREX_ASSERT(a_fluxes[0]->nComp() >= flux_comp+ncomp);
+    AMREX_ASSERT(a_faceState[0]->nComp() >= face_comp+ncomp);
+
+    auto const& ebfact = EBFactory(a_lev);
+
+#ifdef AMREX_USE_OMP
+#pragma omp parallel if (Gpu::notInLaunchRegion())
+#endif
+    for (MFIter mfi(a_divergence,TilingIfNotGPU()); mfi.isValid(); ++mfi) {
+
+        Box const& bx = mfi.tilebox();
+
+        // Get the divergence
+        auto const& div_arr  = a_divergence.array(mfi,div_comp);
+        AMREX_D_TERM(auto const& fx = a_fluxes[0]->const_array(mfi,flux_comp);,
+                     auto const& fy = a_fluxes[1]->const_array(mfi,flux_comp);,
+                     auto const& fz = a_fluxes[2]->const_array(mfi,flux_comp);)
+
+        auto const& flagfab = ebfact.getMultiEBCellFlagFab()[mfi];
+        auto const& vfrac_arr = ebfact.getVolFrac().const_array(mfi);
+        auto const& ebvel_arr = a_EBvelocity->const_array(mfi);
+        auto const& ebval_arr = a_EBvalue->const_array(mfi);
+        if (flagfab.getType(bx) == FabType::singlevalued) {
+           HydroUtils::EB_ComputeDivergence(bx, div_arr,
+                                            AMREX_D_DECL(fx,fy,fz),
+                                            vfrac_arr,
+                                            ncomp, a_geom,
+                                            scale, fluxes_are_area_weighted,
+                                            ebvel_arr, ebval_arr,
+                                            ebfact.getMultiEBCellFlagFab().const_array(mfi),
+                                            ebfact.getBndryArea().const_array(mfi),
+                                            ebfact.getBndryNormal().const_array(mfi));
+        } else if (flagfab.getType(bx) == FabType::regular)
+        {
+           HydroUtils::ComputeDivergence(bx, div_arr,
+                                         AMREX_D_DECL(fx,fy,fz),
+                                         ncomp, a_geom,
+                                         scale, fluxes_are_area_weighted);
+        }
+
+        // If convective, we define u dot grad q = div (u q) - q div(u)
+        // averaging face and t^{n+1/2} q to the cell center
+        auto const& divu_arr    = a_divu.const_array(mfi);
+        AMREX_D_TERM(auto const& facex = a_faceState[0]->const_array(mfi,face_comp);,
[…]
