+++ Source/PeleLMAdvection.cpp
@@ -95,6 +95,7 @@ void PeleLM::computeVelocityAdvTerm(std::unique_ptr<AdvanceAdvData> &advData)
                                                  bcRecVel, bcRecVel_d.dataPtr(), AdvTypeVel_d.dataPtr(),
 #ifdef AMREX_USE_EB
                                                  ebfact,
+                                                 m_useEBinflow ? getEBState(mfi,lev,VELX,AMREX_SPACEDIM,AmrOldTime).const_array() : Array4<Real const>{},
 #endif
                                                  m_Godunov_ppm, m_Godunov_ForceInTrans,
                                                  is_velocity, fluxes_are_area_weighted,
@@ -147,14 +148,27 @@ void PeleLM::computeVelocityAdvTerm(std::unique_ptr<AdvanceAdvData> &advData)
       int nGrow_divT = 3;
       MultiFab divTmp(grids[lev],dmap[lev],AMREX_SPACEDIM,nGrow_divT,MFInfo(),EBFactory(lev));
       divTmp.setVal(0.0);
-      advFluxDivergence(lev, divTmp, 0,
-                        divu,
-                        GetArrOfConstPtrs(fluxes[lev]), 0,
-                        GetArrOfConstPtrs(faces[lev]), 0,
-                        AMREX_SPACEDIM,
-                        AdvTypeVel_d.dataPtr(),
-                        geom[lev], -1.0,
-                        fluxes_are_area_weighted);
+      if (m_useEBinflow) {
+          advFluxDivergence(lev, divTmp, 0,
+                            divu,
+                            GetArrOfConstPtrs(fluxes[lev]), 0,
+                            GetArrOfConstPtrs(faces[lev]), 0,
+                            getEBState(lev,VELX,AMREX_SPACEDIM,AmrOldTime).get(),
+                            getEBState(lev,VELX,AMREX_SPACEDIM,AmrOldTime).get(),
+                            AMREX_SPACEDIM,
+                            AdvTypeVel_d.dataPtr(),
+                            geom[lev], -1.0,
+                            fluxes_are_area_weighted);
+      } else {
+          advFluxDivergence(lev, divTmp, 0,
+                            divu,
+                            GetArrOfConstPtrs(fluxes[lev]), 0,
+                            GetArrOfConstPtrs(faces[lev]), 0,
+                            AMREX_SPACEDIM,
+                            AdvTypeVel_d.dataPtr(),
+                            geom[lev], -1.0,
+                            fluxes_are_area_weighted);
+      }
 
       divTmp.FillBoundary(geom[lev].periodicity());
 
@@ -374,6 +388,7 @@ void PeleLM::computeScalarAdvTerms(std::unique_ptr<AdvanceAdvData> &advData)
                                                  bcRecSpec, bcRecSpec_d.dataPtr(), AdvTypeSpec_d.dataPtr(),
 #ifdef AMREX_USE_EB
                                                  ebfact,
+                                                 m_useEBinflow ? getEBState(mfi,lev,FIRSTSPEC,NUM_SPECIES-NUM_IONS,AmrOldTime).const_array() : Array4<Real const>{},
 #endif
                                                  m_Godunov_ppm, m_Godunov_ForceInTrans,
                                                  is_velocity, fluxes_are_area_weighted,
@@ -382,21 +397,22 @@ void PeleLM::computeScalarAdvTerms(std::unique_ptr<AdvanceAdvData> &advData)
 
          // Ions one by one
          for ( int n = 0; n < NUM_IONS; n++) {
-            auto bcRecIons = fetchBCRecArray(FIRSTSPEC+NUM_SPECIES-NUM_IONS+n,1);
[…]
