### 2023-05-17 b71b1ac
Missing number of boxes in init grid.
