+++ Docs/source/manual/Model.rst
@@ -21,12 +21,12 @@ In a nutshell, `PeleLMeX` features include:
 
 * Software :
    * Written in C++
-   * Parallelization using MPI+X appraoch, with X one of OpenMP, CUDA, HIP or SYCL
+   * Parallelization using MPI+X approach, with X one of OpenMP, CUDA, HIP or SYCL
    * Parallel I/O
    * Built-in profiling tools
    * Plotfile format supported by `Amrvis <https://github.com/AMReX-Codes/Amrvis/>`_, `yt <http://yt-project.org/>`_, `Paraview <https://www.paraview.org/>`_
 * Physics & numerics :
-   * Finite volume, block-structured AMR appraoch
+   * Finite volume, block-structured AMR approach
    * 2D-Cartesian, 2D-Axisymmetric and 3D support
    * Combustion (transport, kinetics, thermodynamics) models based on Cantera and EGLib through `PelePhysics <https://github.com/AMReX-Combustion/PelePhysics>`_
    * Complex geometries using Embedded Boundaries (EB)
@@ -182,7 +182,7 @@ An overview of `PeleLMeX` time-advance function is provided in :numref:`LMeX_Alg
    : Flowchart of the *PeleLMeX* advance function.
 
 The three steps of the low Mach number projection scheme described :ref:`below <ssec:projScheme>` are referenced to better
-emphasize how the thermodynamic solve is closely weaved into the fractional step appraoch. Striped boxes indicate where the
+emphasize how the thermodynamic solve is closely weaved into the fractional step approach. Striped boxes indicate where the
 :ref:`Godunov procedure <ssec:advScheme>` is employed while the four different linear solves are highlighted.
 
 Low Mach number projection scheme
@@ -192,7 +192,7 @@ Low Mach number projection scheme
 `PeleLMeX` implements a finite-volume, Cartesian grid discretization approach with constant grid spacing, where
 :math:`U`, :math:`\rho`, :math:`\rho Y_m`, :math:`\rho h`, and :math:`T` represent cell averages, and the pressure field, :math:`\pi`, is defined on the nodes
 of the grid, and is temporally constant on the intervals over the time step.
-The projection scheme is based on a fractional step appraoch where, for purely incompressible flow, the velocity is first advanced in time
+The projection scheme is based on a fractional step approach where, for purely incompressible flow, the velocity is first advanced in time
 using the momentum equation (**Step 1**) and subsequently projected to enforce the divergence constraint (**Step 3**). When considering variable density flows,
 the scalar thermodynamic advance is performed between these two steps (**Step 2**), but within the SDC context, **Step 1** and **Step 2** are effectively interlaced.
 The three major steps of the algorithm (Almgren *et al.* 1998, Day and Bell, 2000, Nonaka *et al.* 2012):
@@ -223,7 +223,7 @@ field at :math:`t^{n+1/2}` that discretely satisfies the constraint. This field 
 the time-explicit advective fluxes for :math:`U`, :math:`\rho h`, and :math:`\rho Y_m`.
 
 
-**Step 2**: (*Advance thermodynamic variables*) Integrate :math:`(\rho Y_m,\rho h)` over the full time step using a spectral deferred correction (SDC) appraoch, the details of which can be found in `PeleLM documentation <https://amrex-combustion.github.io/PeleLM/manual/html/Model.html#sdc-preliminaries>`_. An even more detailed version of the algorithm is available in Nonaka *et al.*, 2018.
[…]
