### 2023-05-19 0691c0d
First version of the NTNU bluff-body experiment used for the TNF workshop
