+++ Exec/Cases/NTNU_TNF_flame/pelelm_prob.H
@@ -1,0 +1,184 @@
+#ifndef PELELM_PROB_H_
+#define PELELM_PROB_H_
+
+#include <AMReX_Geometry.H>
+#include <AMReX_FArrayBox.H>
+#include <AMReX_ParmParse.H>
+#include <AMReX_SPACE.H>
+
+#include <PeleLM_Index.H>
+#include <pelelm_prob_parm.H>
+#include <PMF.H>
+#include <PMFData.H>
+#include <PelePhysics.H>
+
+
+
+AMREX_GPU_DEVICE
+AMREX_FORCE_INLINE
+void
+pelelm_initdata (int i, int j, int k,
+                 int is_incompressible,
+                 amrex::Array4<amrex::Real> const& state,
+                 amrex::Array4<amrex::Real> const& aux,
+                 amrex::GeometryData const& geomdata,
+                 ProbParm const& prob_parm,
+                 pele::physics::PMF::PmfData::DataContainer const *pmf_data)
+{
+    const amrex::Real* prob_lo = geomdata.ProbLo();
+    const amrex::Real* prob_hi = geomdata.ProbHi();
+    const amrex::Real* dx      = geomdata.CellSize();
+
+    AMREX_D_TERM(const amrex::Real x = prob_lo[0] + (i+0.5)*dx[0];,
+                 const amrex::Real y = prob_lo[1] + (j+0.5)*dx[1];,
+                 const amrex::Real z = prob_lo[2] + (k+0.5)*dx[2];);
+
+    AMREX_D_TERM(,
+                 const amrex::Real Lx = prob_hi[0] - prob_lo[0];,
+                 const amrex::Real Ly = prob_hi[1] - prob_lo[1]);
+
+    auto eos = pele::physics::PhysicsType::eos();
+    amrex::GpuArray<amrex::Real, NUM_SPECIES + 4> pmf_vals = {0.0};
+    amrex::Real pert = 0.0;
+    amrex::Real y1, y2, z1, z2;
+
+    amrex::Real massfrac[NUM_SPECIES]      = {0.0};
+    amrex::Real molefrac[NUM_SPECIES]      = {0.0};
+      
+    state(i,j,k,TEMP) = prob_parm.T_in;
+
+    for (int n = 0; n < NUM_SPECIES; n++)
+      massfrac[n] = prob_parm.Ys[n];
+
+    state(i,j,k,VELX) = 0;
+    state(i,j,k,VELY) = 0;
+    state(i,j,k,VELZ) = prob_parm.U_b;
+
+    amrex::Real rho_cgs, P_cgs;
+    P_cgs = prob_parm.P_mean * 10.0;
+
+    eos.PYT2R(P_cgs, massfrac, state(i,j,k,TEMP), rho_cgs);
+    state(i,j,k,DENSITY) = rho_cgs * 1.0e3;            // CGS -> MKS conversion
+
+    eos.TY2H(state(i,j,k,TEMP), massfrac, state(i,j,k,RHOH));
+    state(i,j,k,RHOH) *= 1.0e-4 * state(i,j,k,DENSITY);   // CGS -> MKS conversion
+
+    for (int n = 0; n < NUM_SPECIES; n++) {
+        state(i,j,k,FIRSTSPEC+n) = massfrac[n] * state(i,j,k,DENSITY);
+    }
+}
+
+AMREX_GPU_DEVICE
+AMREX_FORCE_INLINE
+void
+bcnormal(
+  const amrex::Real x[AMREX_SPACEDIM],
+  const int m_nAux,
+  amrex::Real s_ext[NVAR],
+  const int idir,
+  const int sgn,
+  const amrex::Real time,
[…]
