+++ Exec/Cases/NTNU_TNF_flame/EBUserDefined.H
@@ -1,0 +1,123 @@
+#ifndef _EBUSERDEFINED_H_
+#define _EBUSERDEFINED_H_
+
+using namespace amrex;
+
+#ifdef AMREX_USE_EB
+#include <AMReX_EB2.H>
+#include <AMReX_EB2_IF.H>
+void
+EBUserDefined(const Geometry& geom,
+              const int required_coarsening_level,
+              const int max_coarsening_level)
+{
+    // ParmParse your geometry parameters
+    // --> fixed geometry of the challenge problem
+
+    const amrex::Real *problo = geom.ProbLo();
+    const amrex::Real *probhi = geom.ProbHi();
+    amrex::Real dy = geom.CellSize()[1]; //* pow(2.0,max_coarsening_level);
+    amrex::Real dx = geom.CellSize()[0]; //* pow(2.0,max_coarsening_level);
+
+    amrex::Print() << "** Initialising bluff-body geometry... \n";
+    
+    amrex::ParmParse pp("EB");
+    amrex::Real in_diam = 0.0;
+    pp.query("in_diam",in_diam);
+    amrex::Real out_diam = 0.0;
+    pp.query("out_diam",out_diam);
+    amrex::Real plane = 0.0;
+    pp.query("horizontal_plane" ,plane);
+    amrex::Real bluff_diam = 0.0;
+    pp.query("bluff_diam"       ,bluff_diam);
+    amrex::Real bluff_height = 0.0;
+    pp.query("bluff_height"     ,bluff_height);
+    // The cell_id step is done to have the vertical and horizontal planes 
+    // exactly at the cell face and avoid numerical problems with small volume cells
+
+    EB2::PlaneIF out_vert_right({AMREX_D_DECL(out_diam/2.0,0.,0.)},
+                                   {AMREX_D_DECL(1.,0.,0.)});
+
+    EB2::PlaneIF in_vert_right({AMREX_D_DECL(in_diam/2.0,0.,0.)},
+                                   {AMREX_D_DECL(-1.,0.,0.)});
+    
+    int cell_id = (plane-problo[1])/dy;
+    EB2::PlaneIF horizontal_plane({AMREX_D_DECL(0., plane,0.)},
+                                  {AMREX_D_DECL(0.,-1.,0.)});
+
+
+    amrex::Array<amrex::Real,AMREX_SPACEDIM> point0_cat;
+    amrex::Array<amrex::Real,AMREX_SPACEDIM> point1_cat;
+    
+    point1_cat[0] =   in_diam/2.0; //x-coordinate
+    point1_cat[1] =   bluff_height; //y-coordinate
+
+    point0_cat[0] =  bluff_diam/2.; //x-coordinate
+    point0_cat[1] =   0.0; //y-coordinate
+
+    amrex::Array<amrex::Real,AMREX_SPACEDIM> norm0;
+
+    norm0[0] =  (point0_cat[0]-point1_cat[0]);
+    norm0[1] =  (point0_cat[1]-point1_cat[1]); 
+    norm0[2] = 0.0;
+
+    amrex::Real norm = sqrt(norm0[0]*norm0[0]+norm0[1]*norm0[1]);
+    norm0[0] = norm0[0]/norm;
+    norm0[1] = norm0[1]/norm;
+    
+    amrex::Real loc_x;
+    pp.query("loc_x", loc_x );
+    amrex::Real loc_y;
+    pp.query("loc_y", loc_y );    
+    EB2::PlaneIF bluff_body({AMREX_D_DECL(loc_x,loc_y,0)},{AMREX_D_DECL(-norm0[0],-norm0[1],0)});
+
[…]
