### 2023-05-22 4ad588b
Minor fix in boundary condition velocity profile calculation
