### 2023-05-23 cf693dd
Use std::void_t
