### 2023-05-24 cbd0034
Remove EB kernel.
