### 2023-05-24 7befa01
Remove couple of EB kernels now defined elsewhere.
