### 2023-05-24 #5450 @mameehan5
Hi @esclapez ,

I finished my first pass at a review. Overall, the paper and codebase look great! Below, I have a few comments with a few minor revisions before acceptance.

General checks:
- All of the required checks here pass, no major concerns or questions.

Functionality:
- There is a small bug installation instructions. I think there is a typo in the README where `cd PeleLMeX/Exec/RegTest/HotBubble` should be `cd PeleLMeX/Exec/RegTests/HotBubble`, right? Otherwise, all works well.
- I have personally confirmed key functionality and performance claims, at least at a superficial level (e.g., it runs combustion simulations, highly scalable, etc.). Of course, given the enormous scope of combustion scenarios, I cannot test everything, and would recommend the authors continue adding tests that integrate different functionalities beyond this publication.

Documentation:
- A statement of need is needed in the repository, likely a couple of sentences in the README.
- Installation, examples, and functionality are all great.
- Automated tests are also good, but refer to previous comments.
- A small piece of documentation is needed specifically stating where to report issues and seek support. 1-2 sentences in the README should do the trick.

Software paper:
- The authors provide a great summary.
- The statement of need is in good shape but it should be expanded, particularly the last paragraph. Why do we need, specifically, PeleLMeX to study these fine scale interactions? What next-generation combustion devices are you referring to? Who is the target audience? While I think most readers may be able to infer these (e.g., AMR is required for dynamic, spatially-local flow features which sets it apart), these goals should be more clearly stated.
- The authors do a great job explaining the underlying software that supports PeleLMeX, but I think a little more context would be good, specifically including a couple other pieces of software that do similar things (e.g., PeleLM, OpenFOAM, Sierra/Fuego, Chord, etc.). Obviously this does not need to be a complete list, but there should be enough so the reader has a rough idea where PeleLMeX fits into combustion codes.
