### 2023-05-31 0a5f680
Set default droplets CFL to 0.7 (same as gaseous default).
