### 2023-05-31 ce6bd98
Simplify spray data to make plotting spray sources easier
