+++ Source/Spray/PeleLMSprayParticles.cpp
@@ -8,13 +8,11 @@
 using namespace amrex;
 
 namespace {
-SprayData sprayData;
 // Indices for spray source MultiFab
 int sprayMomSrcIndx = VELX;
 int sprayRhoSrcIndx = DENSITY;
 int spraySpecSrcIndx = FIRSTSPEC;
 int sprayEngSrcIndx = FIRSTSPEC + SPRAY_FUEL_NUM;
-SprayComps scomps;
 Vector<Real> spray_cfl;
 Vector<int> spray_state_ghosts;
 Vector<int> spray_source_ghosts;
@@ -23,10 +21,7 @@ Vector<int> prev_state;
 Vector<int> prev_source;
 
 bool mesh_regrid = true;
-std::string init_file;
-int init_function = 1;
 int spray_verbose = 0;
-Real max_spray_cfl = 0.7;
 } // namespace
 
 std::unique_ptr<SprayParticleContainer> PeleLM::SprayPC = nullptr;
@@ -35,9 +30,6 @@ std::unique_ptr<SprayParticleContainer> PeleLM::GhostPC = nullptr;
 bool PeleLM::do_spray_particles = true;
 // momentum + density + fuel species + enthalpy
 int PeleLM::num_spray_src = AMREX_SPACEDIM + 2 + SPRAY_FUEL_NUM;
-
-int PeleLM::write_spray_ascii_files = 0;
-int PeleLM::plot_spray_src = 0;
 
 Real
 PeleLM::SprayEstDt()
@@ -48,7 +40,7 @@ PeleLM::SprayEstDt()
   }
   BL_PROFILE("PeleLM::SprayEstDt()");
   for (int lev = 0; lev <= finest_level; ++lev) {
-    Real estdt_lev = SprayPC->estTimestep(lev, max_spray_cfl);
+    Real estdt_lev = SprayPC->estTimestep(lev);
     if (estdt_lev > 0. && estdt_lev < estdt) {
       estdt = estdt_lev;
     } else if (lev > 0) {
@@ -68,12 +60,9 @@ PeleLM::SprayReadParameters()
   if (!do_spray_particles) {
     return;
   }
-  const Real temp_cfl = max_spray_cfl;
   // Mush change dtmod to 1 since we only do MKD
   sprayData.dtmod = 1.;
-  SprayParticleContainer::readSprayParams(
-    spray_verbose, max_spray_cfl, write_spray_ascii_files,
-    plot_spray_src, init_function, init_file, sprayData, temp_cfl);
+  SprayParticleContainer::readSprayParams(spray_verbose);
 }
 
 void
@@ -87,7 +76,8 @@ PeleLM::SpraySetup()
   if (SPRAY_FUEL_NUM > NUM_SPECIES) {
     Abort("Cannot have more spray fuel species than fluid species");
   }
-  SprayParticleContainer::spraySetup(sprayData, m_gravity.data());
+  SprayParticleContainer::spraySetup(m_gravity.data());
+  SprayData scomps;
   // Component indices for conservative variables
   scomps.rhoIndx = DENSITY;
   scomps.momIndx = VELX;
@@ -99,6 +89,7 @@ PeleLM::SpraySetup()
   scomps.momSrcIndx = sprayMomSrcIndx;
   scomps.specSrcIndx = spraySpecSrcIndx;
   scomps.engSrcIndx = sprayEngSrcIndx;
+  SprayParticleContainer::AssignSprayComps(scomps);
 }
 
 void
@@ -149,13 +140,10 @@ PeleLM::removeGhostParticles(const int level)
 void
 PeleLM::SprayCreateData()
 {
-  SprayPC = std::make_unique<SprayParticleContainer>(
[…]
