+++ Source/PeleLM.cpp
@@ -22,6 +22,9 @@ PeleLM::~PeleLM()
    The_Arena()->free(prob_parm_d);
    m_initial_ba.clear();
    m_regrid_ba.clear();
+#ifdef PELELM_USE_SPRAY
+   SprayParticleContainer::SprayCleanUp();
+#endif
 #ifdef PELELM_USE_SOOT
    cleanupSootModel();
 #endif
