### 2023-05-31 6af8705
Set dt mod after reading params
