+++ Source/PeleLMSetup.cpp
@@ -1,8 +1,9 @@
-#include <PeleLM.H>
 #include <AMReX_ParmParse.H>
-#include <PeleLMDeriveFunc.H>
-#include "PelePhysics.H"
 #include <AMReX_buildInfo.H>
+#include <PeleLM.H>
+#include <PeleLMDeriveFunc.H>
+
+#include "PelePhysics.H"
 #ifdef PELE_USE_EFIELD
 #include "EOS_Extension.H"
 #endif
@@ -21,1131 +22,1261 @@
 
 using namespace amrex;
 
-static Box the_same_box (const Box& b)    { return b;                }
-static Box grow_box_by_two (const Box& b) { return amrex::grow(b,2); }
+static Box the_same_box(const Box& b) { return b; }
+static Box grow_box_by_two(const Box& b) { return amrex::grow(b, 2); }
 
 void PeleLM::Setup() {
-   BL_PROFILE("PeleLM::Setup()");
+	BL_PROFILE("PeleLM::Setup()");
 
-   m_wall_start = amrex::ParallelDescriptor::second();
+	m_wall_start = amrex::ParallelDescriptor::second();
 
-   // Ensure grid is isotropic
-   {
-     auto const dx = geom[0].CellSizeArray();
-     AMREX_ALWAYS_ASSERT(AMREX_D_TERM(,amrex::almostEqual(dx[0], dx[1]), && amrex::almostEqual(dx[1], dx[2])));
-   }
-   // Print build info to screen
-   const char* githash1 = buildInfoGetGitHash(1);
-   const char* githash2 = buildInfoGetGitHash(2);
-   const char* githash3 = buildInfoGetGitHash(3);
-   const char* githash4 = buildInfoGetGitHash(4);
-   amrex::Print() << "\n ================= Build infos =================\n";
-   amrex::Print() << " PeleLMeX    git hash: " << githash1 << "\n";
-   amrex::Print() << " AMReX       git hash: " << githash2 << "\n";
-   amrex::Print() << " PelePhysics git hash: " << githash3 << "\n";
-   amrex::Print() << " AMReX-Hydro git hash: " << githash4 << "\n";
-   amrex::Print() << " ===============================================\n";
+	// Ensure grid is isotropic
+	{
+		auto const dx = geom[0].CellSizeArray();
+		AMREX_ALWAYS_ASSERT(
+		    AMREX_D_TERM(, amrex::almostEqual(dx[0], dx[1]),
+				 &&amrex::almostEqual(dx[1], dx[2])));
+	}
+	// Print build info to screen
+	const char* githash1 = buildInfoGetGitHash(1);
+	const char* githash2 = buildInfoGetGitHash(2);
+	const char* githash3 = buildInfoGetGitHash(3);
+	const char* githash4 = buildInfoGetGitHash(4);
+	amrex::Print()
+	    << "\n ================= Build infos =================\n";
+	amrex::Print() << " PeleLMeX    git hash: " << githash1 << "\n";
+	amrex::Print() << " AMReX       git hash: " << githash2 << "\n";
+	amrex::Print() << " PelePhysics git hash: " << githash3 << "\n";
+	amrex::Print() << " AMReX-Hydro git hash: " << githash4 << "\n";
+	amrex::Print() << " ===============================================\n";
 
 #ifdef PELELM_USE_SOOT
-   soot_model = new SootModel{};
+	soot_model = new SootModel{};
 #endif
 
-   // Read PeleLM parameters
-   readParameters();
+	// Read PeleLM parameters
+	readParameters();
 
[…]
