+++ Source/PeleLMPlot.cpp
@@ -1,11 +1,12 @@
-#include <PeleLM.H>
+#include <AMReX_FillPatchUtil.H>
+#include <AMReX_ParmParse.H>
 #include <AMReX_PlotFileUtil.H>
 #include <AMReX_buildInfo.H>
-#include "PelePhysics.H"
-#include <PltFileManager.H>
-#include <AMReX_ParmParse.H>
+#include <PeleLM.H>
 #include <PeleLMBCfill.H>
-#include <AMReX_FillPatchUtil.H>
+#include <PltFileManager.H>
+
+#include "PelePhysics.H"
 #ifdef AMREX_USE_EB
 #include <AMReX_EBInterpolater.H>
 #endif
@@ -22,997 +23,1135 @@
 
 using namespace amrex;
 
-namespace { const std::string level_prefix{"Level_"}; }
-
-void GotoNextLine(std::istream& is)
-{
-       constexpr std::streamsize bl_ignore_max{100000};
-           is.ignore(bl_ignore_max, '\n');
+namespace {
+const std::string level_prefix{"Level_"};
 }
 
-void PeleLM::WriteDebugPlotFile(const Vector<const MultiFab*> &a_MF,
-                                const std::string &pltname)
-{
-   int nComp = a_MF[0]->nComp();
-   Vector<std::string> names(nComp);
-   for (int n = 0; n < nComp; n++) {
-      names[n] = "comp"+std::to_string(n);
-   }
-   Vector<int> istep(finest_level + 1, m_nstep);
+void GotoNextLine(std::istream& is) {
+	constexpr std::streamsize bl_ignore_max{100000};
+	is.ignore(bl_ignore_max, '\n');
+}
+
+void PeleLM::WriteDebugPlotFile(const Vector<const MultiFab*>& a_MF,
+				const std::string& pltname) {
+	int nComp = a_MF[0]->nComp();
+	Vector<std::string> names(nComp);
+	for (int n = 0; n < nComp; n++) {
+		names[n] = "comp" + std::to_string(n);
+	}
+	Vector<int> istep(finest_level + 1, m_nstep);
 #ifdef AMREX_USE_HDF5
-   if (m_write_hdf5_pltfile) {
-       amrex::WriteMultiLevelPlotfileHDF5(pltname, finest_level + 1, a_MF,
-                                          names, Geom(), m_cur_time, istep, refRatio());
-   } else
+	if (m_write_hdf5_pltfile) {
+		amrex::WriteMultiLevelPlotfileHDF5(
+		    pltname, finest_level + 1, a_MF, names, Geom(), m_cur_time,
+		    istep, refRatio());
+	} else
 #endif
-   {
-       amrex::WriteMultiLevelPlotfile(pltname, finest_level + 1, a_MF,
-                                      names, Geom(), m_cur_time, istep, refRatio());
-   }
+	{
+		amrex::WriteMultiLevelPlotfile(pltname, finest_level + 1, a_MF,
+					       names, Geom(), m_cur_time, istep,
+					       refRatio());
+	}
 }
 
 void PeleLM::WritePlotFile() {
-   BL_PROFILE("PeleLM::WritePlotFile()");
+	BL_PROFILE("PeleLM::WritePlotFile()");
 
-   const std::string& plotfilename = amrex::Concatenate(m_plot_file, m_nstep, m_ioDigits);
+	const std::string& plotfilename =
+	    amrex::Concatenate(m_plot_file, m_nstep, m_ioDigits);
 
-   if (m_verbose) {
-      amrex::Print() << "\n Writing plotfile: " << plotfilename << "\n";
-   }
+	if (m_verbose) {
+		amrex::Print()
[…]

[…]
