+++ Source/PeleLMPlot.cpp
@@ -27,1131 +27,1062 @@ namespace {
 const std::string level_prefix{"Level_"};
 }
 
-void GotoNextLine(std::istream& is) {
-	constexpr std::streamsize bl_ignore_max{100000};
-	is.ignore(bl_ignore_max, '\n');
+void
+GotoNextLine(std::istream& is)
+{
+  constexpr std::streamsize bl_ignore_max{100000};
+  is.ignore(bl_ignore_max, '\n');
 }
 
-void PeleLM::WriteDebugPlotFile(const Vector<const MultiFab*>& a_MF,
-				const std::string& pltname) {
-	int nComp = a_MF[0]->nComp();
-	Vector<std::string> names(nComp);
-	for (int n = 0; n < nComp; n++) {
-		names[n] = "comp" + std::to_string(n);
-	}
-	Vector<int> istep(finest_level + 1, m_nstep);
+void
+PeleLM::WriteDebugPlotFile(
+  const Vector<const MultiFab*>& a_MF, const std::string& pltname)
+{
+  int nComp = a_MF[0]->nComp();
+  Vector<std::string> names(nComp);
+  for (int n = 0; n < nComp; n++) {
+    names[n] = "comp" + std::to_string(n);
+  }
+  Vector<int> istep(finest_level + 1, m_nstep);
 #ifdef AMREX_USE_HDF5
-	if (m_write_hdf5_pltfile) {
-		amrex::WriteMultiLevelPlotfileHDF5(
-		    pltname, finest_level + 1, a_MF, names, Geom(), m_cur_time,
-		    istep, refRatio());
-	} else
+  if (m_write_hdf5_pltfile) {
+    amrex::WriteMultiLevelPlotfileHDF5(
+      pltname, finest_level + 1, a_MF, names, Geom(), m_cur_time, istep,
+      refRatio());
+  } else
 #endif
-	{
-		amrex::WriteMultiLevelPlotfile(pltname, finest_level + 1, a_MF,
-					       names, Geom(), m_cur_time, istep,
-					       refRatio());
-	}
+  {
+    amrex::WriteMultiLevelPlotfile(
+      pltname, finest_level + 1, a_MF, names, Geom(), m_cur_time, istep,
+      refRatio());
+  }
 }
 
-void PeleLM::WritePlotFile() {
-	BL_PROFILE("PeleLM::WritePlotFile()");
+void
+PeleLM::WritePlotFile()
+{
+  BL_PROFILE("PeleLM::WritePlotFile()");
 
-	const std::string& plotfilename =
-	    amrex::Concatenate(m_plot_file, m_nstep, m_ioDigits);
+  const std::string& plotfilename =
+    amrex::Concatenate(m_plot_file, m_nstep, m_ioDigits);
 
-	if (m_verbose) {
-		amrex::Print()
-		    << "\n Writing plotfile: " << plotfilename << "\n";
-	}
+  if (m_verbose) {
+    amrex::Print() << "\n Writing plotfile: " << plotfilename << "\n";
+  }
 
-	//----------------------------------------------------------------
-	// Average down the state
-	averageDownState(AmrNewTime);
+  //----------------------------------------------------------------
+  // Average down the state
+  averageDownState(AmrNewTime);
 
-	// Get consistent reaction data across level
-	if (m_do_react && !m_skipInstantRR && m_plot_react) {
-		averageDownReaction();
-	}
+  // Get consistent reaction data across level
[…]
