+++ Exec/Cases/NTNU_TNF_flame/pelelm_prob.H
@@ -86,16 +86,15 @@ bcnormal(
   const amrex::Real* prob_hi = geomdata.ProbHi();
 
 
-  amrex::Real massfrac[NUM_SPECIES] = {0.0};
-  //load premixed composition to massfrac array
-  for (int n = 0; n < NUM_SPECIES; n++) {
-    massfrac[n] = prob_parm.Ys[n];
-  }
-
-
   auto eos = pele::physics::PhysicsType::eos();
 
   if(idir == 2 and sgn == 1){
+
+    amrex::Real massfrac[NUM_SPECIES] = {0.0};
+    //load premixed composition to massfrac array
+    for (int n = 0; n < NUM_SPECIES; n++) {
+      massfrac[n] = prob_parm.Ys[n];
+    }
 
     // amrex::GpuArray<amrex::Real, NUM_SPECIES+4> vel_from_file = {0.0};
 
@@ -147,7 +146,37 @@ bcnormal(
       s_ext[FIRSTSPEC+n] = massfrac[n] * s_ext[DENSITY];
     }
   }
+
+  // else if(idir == 0 or idir == 1){
+  //   amrex::Real profile = 1.0 - 0.5*(1.0+std::tanh((x[2]-prob_parm.z_height_tanh)/prob_parm.delta));
+
+  //   amrex::Real rho_cgs, P_cgs, RhoH_temp;
+  //   P_cgs = prob_parm.P_mean * 10.0;
+
+  //   amrex::Real theta  = atan(x[1]/x[0]);
+
+  //   s_ext[VELX] = -prob_parm.U_crossflow*cos(theta);
+  //   s_ext[VELY] = -prob_parm.U_crossflow*sin(theta);
+  //   s_ext[VELZ] = 0.0;      
+
+  //   s_ext[TEMP] = prob_parm.T_in;
+
+  //   eos.PYT2R(P_cgs, massfrac_air, s_ext[TEMP], rho_cgs);
+  //   s_ext[DENSITY] = rho_cgs * 1.0e3;
+
+  //   eos.TY2H(s_ext[TEMP], massfrac_air, RhoH_temp);
+  //   s_ext[RHOH] = RhoH_temp * 1.0e-4 * s_ext[DENSITY];   // CGS -> MKS conversion
+
+  //   for (int n = 0; n < NUM_SPECIES; n++) {
+  //     s_ext[FIRSTSPEC+n] = massfrac_air[n] * s_ext[DENSITY];
+  //   }
+  // }
+
   else{
+
+    amrex::Real massfrac_air[NUM_SPECIES] = {0.0};
+    massfrac_air[N2_ID] = 0.767;
+    massfrac_air[O2_ID] = 0.233;
 
     amrex::Real rho_cgs, P_cgs, RhoH_temp;
     P_cgs = prob_parm.P_mean * 10.0;
@@ -158,14 +187,15 @@ bcnormal(
 
     s_ext[TEMP] = prob_parm.T_in;
 
-    eos.PYT2R(P_cgs, massfrac, s_ext[TEMP], rho_cgs);
+    eos.PYT2R(P_cgs, massfrac_air, s_ext[TEMP], rho_cgs);
     s_ext[DENSITY] = rho_cgs * 1.0e3;
 
-    eos.TY2H(s_ext[TEMP], massfrac, RhoH_temp);
+    eos.TY2H(s_ext[TEMP], massfrac_air, RhoH_temp);
     s_ext[RHOH] = RhoH_temp * 1.0e-4 * s_ext[DENSITY];   // CGS -> MKS conversion
 
     for (int n = 0; n < NUM_SPECIES; n++) {
[…]

[…]
