### 2023-06-12 #5450 @xzz105
Hi @esclapez

I have gone through the paper and software packages. Everything looks good, and the user experience is smooth. I have a few minor suggestions below for your consideration.

Most of the general review points check off.

Documentation:

* Installation instructions: it would be helpful to specify the minimal requirement of environments up front. I tried to install it on a new personal MacBook, and realized I need to have mpi installed; I tried another older cluster to install the software package, and realized gcc 8 above is required. It would be good to have these information up front so that user would get prepared.
*  Running instructions are available for a few cases on the tutorial pages of the website. It would be helpful if there is a README for each of the test folders in the RegTests folder so users have everything in one place. Also, the README.md in the RegTests folder seem to be a bit outdated, for example, the EB_BackwardStepFlame case was not introduced in the README.md file.
* Otherwise the tutorial cases ran fine. As an OpenFOAM user, I am excited to find that it is easy to switch between serial run and parallel run with PeleLMeX.

Software paper

* As someone who followed Pele development a while ago, my biggest question when I saw the title is "what's the difference between PeleLMeX and PeleLM". I couldn't find the answer from the software paper, which may be due to me being unfamiliar with either softwares by now. It would be very helpful if the authors can explicitly mention the differences between PeleLM and PeleLMex, and guidance on how to choose one versus another in the software paper.

* I also echo @mameehan5 on his last point; I also think a little more context involving comparison with other similar softwares would be helpful for the users, especially new ones to choose their direction. Discussions such as pros and cons of PeleLMeX would be very helpful.
