### 2023-06-13 fc74630
Add weak scaling scripts and instructions.
