+++ Utils/RunScripts/Scaling/WeakScaling/WeakScaling.py
@@ -1,0 +1,233 @@
+#!/usr/bin/env python3
+
+import sys
+import os
+import shutil
+import argparse
+import socket
+import random
+import numpy as np
+
+# A set of known host is listed in getMachines
+# This info is then used to generate the batch_script in setBatchScript
+# and in launchRun.
+
+USAGE = """
+    A script to setup weak scaling of PeleLMeX.
+    Given an initial problem provided by the input file, an increasing number of nodes
+    is tested (2^n nodes) while increasing the problem size in x and y
+"""
+
+def parse_args(arg_string=None):
+    parser = argparse.ArgumentParser(description=USAGE)
+
+    parser.add_argument("-n","--test_name", type=str, default="Case1", metavar="test-name",
+                        help="Name of the test. Default = Case1")
+
+    parser.add_argument("-e","--exec", type=str, default="None", metavar="Pele-exec",
+                        help="Pele executable name. Default = first Pele*.ex in current directory")
+
+    parser.add_argument("-i","--input_file", type=str, default="None", metavar="Pele-input",
+                        help="Pele input file name. Default = first inputs.* in current directory")
+
+    parser.add_argument("-b","--batch_script", type=str, default="None", metavar="Cluster-batch",
+                        help="The batch script name.", required=True)
+
+    parser.add_argument("--extra_files", nargs='+', type=str, default="None", metavar="Extras",
+                        help="Additional files reqs. for the run. Default = None")
+
+    if not arg_string is None:
+        args, unknown = parser.parse_known_args(arg_string)
+    else:
+        args, unknown = parser.parse_known_args()
+
+    # Checks
+    if (not args.exec== "None" ):
+        if (not os.path.exists(args.exec)):
+            sys.exit("Provided executable '{}' does not exists!".format(args.exec))
+
+    if (not args.input_file== "None" ):
+        if (not os.path.exists(args.input_file)):
+            sys.exit("Provided inputs file '{}' does not exists!".format(args.input_file))
+
+    if (not os.path.exists(args.batch_script)):
+        sys.exit("Provided batch_script '{}' does not exists!".format(args.batch_script))
+
+    if (not args.extra_files== "None" ):
+        for files in args.extra_files:
+            if (not os.path.exists(files)):
+                sys.exit("Provided extra_files '{}' does not exists!".format(files))
+
+    return args 
+
+def getMachines():
+    host = socket.getfqdn()
+
+    site = "unknown"
+    machine = "unknown"
+
+    # handle NERSC
+    nersc_host = os.environ.get('NERSC_HOST')
+    if nersc_host is not None:
+        site = "NERSC"
+        if "cori" in nersc_host:
+            machine = "Cori"
+        if "perlmutter" in nersc_host:
+            machine = "Perlmutter"
+    
+    if "olcf" in host:
+        site = "OLCF"
+        if "ascent" in host:
+            machine = "Ascent"
+        if "summit" in host:
+            machine = "Summit"
+        if "crusher" in host:
+            machine = "Crusher"
+        if "frontier" in host:
+            machine = "Frontier"
+    
+    return [site,machine]
+
+def getInitialSize(fileIn):
+    ofile=open(fileIn,'r')
+    sizeLine = "" 
+    for line in ofile:
+        if 'geometry.prob_hi' in line:
+            sizeLine=line
+
+    sizesStr=sizeLine.split("=")[1]
+    sizesStr=sizesStr.split("#")[0]
+    sizes=[float(x) for x in sizesStr.strip().split(" ")]
+    
[…]
