### 2023-06-21 bed87aa
Add weak scaling scripts and instructions. (#220)
* Add weak scaling scripts and instructions.

* fix style

---------

Co-authored-by: Bruce Perry <bruce.perry@nrel.gov>
