### 2023-06-28 f8a23e8
Changes in NTNU TNF set up
