+++ Exec/Cases/NTNU_TNF_flame/pelelm_prob.H
@@ -147,30 +147,34 @@ bcnormal(
     }
   }
 
-  // else if(idir == 0 or idir == 1){
-  //   amrex::Real profile = 1.0 - 0.5*(1.0+std::tanh((x[2]-prob_parm.z_height_tanh)/prob_parm.delta));
+  else if(idir == 0 or idir == 1){
+    amrex::Real profile = 0.5*(1.0+std::tanh((x[2]-prob_parm.z_height_tanh)/prob_parm.delta));
 
-  //   amrex::Real rho_cgs, P_cgs, RhoH_temp;
-  //   P_cgs = prob_parm.P_mean * 10.0;
+    amrex::Real massfrac_air[NUM_SPECIES] = {0.0};
+    massfrac_air[N2_ID] = 0.767;
+    massfrac_air[O2_ID] = 0.233;
 
-  //   amrex::Real theta  = atan(x[1]/x[0]);
+    amrex::Real rho_cgs, P_cgs, RhoH_temp;
+    P_cgs = prob_parm.P_mean * 10.0;
 
-  //   s_ext[VELX] = -prob_parm.U_crossflow*cos(theta);
-  //   s_ext[VELY] = -prob_parm.U_crossflow*sin(theta);
-  //   s_ext[VELZ] = 0.0;      
+    amrex::Real theta  = atan(x[1]/x[0]);
 
-  //   s_ext[TEMP] = prob_parm.T_in;
+    s_ext[VELX] = -prob_parm.U_crossflow*cos(theta);
+    s_ext[VELY] = -prob_parm.U_crossflow*sin(theta);
+    s_ext[VELZ] = 0.0;      
 
-  //   eos.PYT2R(P_cgs, massfrac_air, s_ext[TEMP], rho_cgs);
-  //   s_ext[DENSITY] = rho_cgs * 1.0e3;
+    s_ext[TEMP] = prob_parm.T_in;
 
-  //   eos.TY2H(s_ext[TEMP], massfrac_air, RhoH_temp);
-  //   s_ext[RHOH] = RhoH_temp * 1.0e-4 * s_ext[DENSITY];   // CGS -> MKS conversion
+    eos.PYT2R(P_cgs, massfrac_air, s_ext[TEMP], rho_cgs);
+    s_ext[DENSITY] = rho_cgs * 1.0e3;
 
-  //   for (int n = 0; n < NUM_SPECIES; n++) {
-  //     s_ext[FIRSTSPEC+n] = massfrac_air[n] * s_ext[DENSITY];
-  //   }
-  // }
+    eos.TY2H(s_ext[TEMP], massfrac_air, RhoH_temp);
+    s_ext[RHOH] = RhoH_temp * 1.0e-4 * s_ext[DENSITY];   // CGS -> MKS conversion
+
+    for (int n = 0; n < NUM_SPECIES; n++) {
+      s_ext[FIRSTSPEC+n] = massfrac_air[n] * s_ext[DENSITY];
+    }
+  }
 
   else{
 
