+++ Source/PeleLMUMac.cpp
@@ -60,7 +60,6 @@ void PeleLM::predictVelocity(std::unique_ptr<AdvanceAdvData>  &advData)
                                    geom[lev], m_dt,
 #ifdef AMREX_USE_EB
                                    ebfact,
-                                   m_useEBinflow ? getEBState(lev,VELX,AMREX_SPACEDIM,AmrOldTime).get() : nullptr,
 #endif
                                    m_Godunov_ppm, m_Godunov_ForceInTrans,
                                    m_predict_advection_type,
@@ -176,15 +175,6 @@ void PeleLM::macProject(const TimeStamp &a_time,
    macproj->getLinOp().setMaxOrder(m_mac_max_order);
    macproj->setUMAC(GetVecOfArrOfPtrs(advData->umac));
    if (has_divu) macproj->setDivU(GetVecOfConstPtrs(a_divu));
-
-#ifdef AMREX_USE_EB
-    if (m_useEBinflow) {
-       for (int lev=0; lev <= finest_level; ++lev)
-       {
-          macproj->setEBInflowVelocity(lev, *getEBState(lev,VELX,AMREX_SPACEDIM,a_time));
-       }
-    }
-#endif
 
    // Project
    macproj->project(m_mac_mg_rtol,m_mac_mg_atol);
