+++ Source/PeleLMProjection.cpp
@@ -399,14 +399,6 @@ void PeleLM::doNodalProject(const Vector<MultiFab*> &a_vel,
 
    nodal_projector->setDomainBC(lobc, hibc);
 
-#ifdef AMREX_USE_EB
-    if (m_useEBinflow) {
-       for(int lev = 0; lev <= finest_level; ++lev) {
-          nodal_projector->getLinOp().setEBInflowVelocity(lev, *getEBState(lev,VELX,AMREX_SPACEDIM,AmrNewTime));
-       }
-    }
-#endif
-
 #ifdef AMREX_USE_HYPRE
    nodal_projector->getMLMG().setHypreOptionsNamespace(m_hypre_namespace_nodal);
 #endif
