### 2023-06-29 e511266
Untag cells after a given z height
