### 2023-06-29 26d3c8b
Change in BC for x and y. Air blowing from the sides
