### 2023-06-29 7d310ea
Fill EB file deleted
