### 2023-06-29 8d378a6
Isothermal wall included in EBUserDefined.H
