### 2023-07-14 d92e96d
Remove a couple of typos in FlameSheet turorial.
