+++ Source/PeleLMReactions.cpp
@@ -8,7 +8,7 @@ using namespace amrex;
 
 void PeleLM::advanceChemistry(std::unique_ptr<AdvanceAdvData> &advData)
 {
-   BL_PROFILE("PeleLM::advanceChemistry()");
+   BL_PROFILE("PeleLMeX::advanceChemistry()");
 
    for (int lev = finest_level; lev >= 0; --lev) {
       if (lev != finest_level) {
@@ -30,7 +30,7 @@ void PeleLM::advanceChemistry(int lev,
                               const Real &a_dt,
                               MultiFab &a_extForcing)
 {
-   BL_PROFILE("PeleLM::advanceChemistry_Lev"+std::to_string(lev)+"()");
+   BL_PROFILE("PeleLMeX::advanceChemistry_Lev"+std::to_string(lev)+"()");
 
    auto ldataOld_p = getLevelDataPtr(lev,AmrOldTime);
    auto ldataNew_p = getLevelDataPtr(lev,AmrNewTime);
@@ -175,7 +175,7 @@ void PeleLM::advanceChemistryBAChem(int lev,
                               const Real &a_dt,
                               MultiFab &a_extForcing)
 {
-   BL_PROFILE("PeleLM::advanceChemistry_Lev"+std::to_string(lev)+"()");
+   BL_PROFILE("PeleLMeX::advanceChemistry_Lev"+std::to_string(lev)+"()");
 
    auto ldataOld_p = getLevelDataPtr(lev,AmrOldTime);
    auto ldataNew_p = getLevelDataPtr(lev,AmrNewTime);
@@ -374,7 +374,7 @@ void PeleLM::computeInstantaneousReactionRate(int lev,
                                               const TimeStamp &a_time,
                                               MultiFab* a_I_R)
 {
-   BL_PROFILE("PeleLM::computeInstantaneousReactionRate()");
+   BL_PROFILE("PeleLMeX::computeInstantaneousReactionRate()");
    auto ldata_p = getLevelDataPtr(lev,a_time);
 
 #ifdef AMREX_USE_EB
