### 2023-07-14 fe5e12a
Couple of updates in the FS tuto.
