### 2023-07-14 fc1b143
Explicitly point to open source MPI distributions in the README.
