### 2023-08-21 f2d23fe
Only restrict Vnew if velMax > 0
