### 2023-08-22 fbbe5a9
Active Control max velocity default value fix (#233)
* Set active_control.velMax default value to -1.0

* Only restrict Vnew if velMax > 0

* Update manual to match changes
