### 2023-08-28 #5450 @drummerdoc
Hey Kyle, Lucas has recently left NREL, so others on the LMeX team are addressing the remaining issue identified above this week.
