### 2023-08-29 b0a252c
Update the RegTests README.
