+++ Exec/RegTests/EB_EnclosedFlame/README.md
@@ -1,0 +1,2 @@
+## EB\_EnclosedFlame
+A 2D (or 3D) cylindrical flame enclosed in a cylindrical EB chamber. Testing the closed chamber algorithm with EB.
