+++ Exec/RegTests/PeriodicCases/README.md
@@ -1,0 +1,4 @@
+## PeriodicCases
+A set of 2D/3D periodic cases: convected vortex in different directions, convected temperature/mixture Gaussian bump,
+pure diffusion of species/temperature. This is the base case to test PeleLMeX accuracy and order of convergence of
+the spatial and temporal schemes.
