+++ Exec/RegTests/TurbInflow/README.md
@@ -1,0 +1,3 @@
+## TurbInflow
+Injection of a 3D precomputed turbulent velocity field at the boundary of a PeleLMeX simulation. Testing the 
+turbulence injection functionality.
