+++ Exec/RegTests/EB_PipeFlow/README.md
@@ -1,0 +1,5 @@
+## EB\_PipeFlow
+2D and 3D setup of both channel and pipe flow, either in an inflow/outflow setup or a 
+periodc setup driven by a background pressure gradient.
+
+Poiseuille flow results with PeleLMeX can be found [here](https://amrex-combustion.github.io/PeleLMeX/manual/html/Validation.html#laminar-poiseuille-flow).
