+++ Exec/RegTests/EnclosedFlame/README.md
@@ -1,0 +1,2 @@
+## EnclosedFlame
+A 2D (or 3D) cylindrical flame enclosed in a square (rectangular) domain. Testing the closed chamber algorithm.
