+++ Exec/RegTests/HITDecay/README.md
@@ -1,0 +1,3 @@
+## HITDecay
+A 3D decaying HIT case, where the initial solution is generated with a Passot-Pouquet spectrum. Testing basic incompressible
+integration algorithm and Large Eddy Simulation implementation.
