+++ Exec/RegTests/FlameSheet/README.md
@@ -1,0 +1,7 @@
+## FlameSheet
+A 2D (3D) harmonically perturbed flame sheet, initial solution from a Cantera simulation provided for 3 mechanisms
+(drm19, dodecane\_lu and dodecane\_lu\_qss). This is the basis for [weak scaling studies](https://amrex-combustion.github.io/PeleLMeX/manual/html/Performances.html) in PeleLMeX and test all the 
+reactive pieces of the algorithm as well as transport options (Unity Lewis number, Soret effect, ...).
+More detqils on the cqse setup qnd step-by-step instructions cqn be found in this [tutoriql](https://amrex-combustion.github.io/PeleLMeX/manual/html/Tutorials_FlameSheet.html).
+
+Comparisons of PeleLMeX results on a methane/air flame against Cantera at several resolution are reported [here](https://amrex-combustion.github.io/PeleLMeX/manual/html/Validation.html#laminar-premixed-flame).
