+++ Exec/RegTests/EB_BackwardStepFlame/README.md
@@ -1,0 +1,3 @@
+## EB\_BackwardStepFlame
+
+This folder implement the case of a laminar flame stabilized behind a backward facing step. It is intended to test integrated EB representation with reactive flow. More details on this case and step-by-step instructions can be found in this [tutorial](https://amrex-combustion.github.io/PeleLMeX/manual/html/Tutorials_BFSFlame.html)
