+++ Exec/RegTests/TaylorGreen/Readme.md
@@ -17,3 +17,5 @@ script.
 
 Note that these commands provide adimentional results, with t\* = t/(L\_0/V\_0),
 E\_k\* = E\_k / ( rho\_0 * V\_0 * V\_0) and psi\* = psi / (V\_0/L\_0)^2
+
+PeleLMeX results can be found in [here](https://amrex-combustion.github.io/PeleLMeX/manual/html/Validation.html#taylor-green-vortex-breakdown).
