### 2023-08-29 1c6951e
Add a short HotBubble tutorial.
