### 2023-08-29 df37020
Update PelePhysics - Use Sundials submod in gmake (#235)
* update pelephysics for sundials submod gmake

* default submodule paths

* CI fixes hopefully

* linux workflow add SUNDIALS variable

* update the rest of the workflows

* update PelePhysics
