+++ .github/workflows/linux.yml
@@ -30,6 +30,7 @@ jobs:
          PELE_PHYSICS_HOME: ${GITHUB_WORKSPACE}/build/PelePhysics
          PELELMEX_HOME: ${GITHUB_WORKSPACE}
          AMREX_HYDRO_HOME: ${GITHUB_WORKSPACE}/build/AMReX-Hydro
+         SUNDIALS_HOME : ${GITHUB_WORKSPACE}/build/sundials
       run: |
         cd Exec/RegTests/FlameSheet/
         make help
@@ -58,6 +59,7 @@ jobs:
          PELE_PHYSICS_HOME: ${GITHUB_WORKSPACE}/build/PelePhysics
          PELELMEX_HOME: ${GITHUB_WORKSPACE}
          AMREX_HYDRO_HOME: ${GITHUB_WORKSPACE}/build/AMReX-Hydro
+         SUNDIALS_HOME : ${GITHUB_WORKSPACE}/build/sundials
       run: |
         cd Exec/RegTests/FlameSheet/
         make TPL COMP=gnu
@@ -93,6 +95,7 @@ jobs:
          PELE_PHYSICS_HOME: ${GITHUB_WORKSPACE}/build/PelePhysics
          PELELMEX_HOME: ${GITHUB_WORKSPACE}
          AMREX_HYDRO_HOME: ${GITHUB_WORKSPACE}/build/AMReX-Hydro
+         SUNDIALS_HOME : ${GITHUB_WORKSPACE}/build/sundials
       run: |
         cd Exec/RegTests/FlameSheet/
         make TPL COMP=gnu USE_OMP=TRUE
@@ -116,6 +119,7 @@ jobs:
          PELE_PHYSICS_HOME: ${GITHUB_WORKSPACE}/build/PelePhysics
          PELELMEX_HOME: ${GITHUB_WORKSPACE}
          AMREX_HYDRO_HOME: ${GITHUB_WORKSPACE}/build/AMReX-Hydro
+         SUNDIALS_HOME : ${GITHUB_WORKSPACE}/build/sundials
       working-directory: ./Exec/RegTests/PeriodicCases/
       run: |
         make TPL COMP=gnu USE_MPI=FALSE
@@ -143,6 +147,7 @@ jobs:
          PELE_PHYSICS_HOME: ${GITHUB_WORKSPACE}/build/PelePhysics
          PELELMEX_HOME: ${GITHUB_WORKSPACE}
          AMREX_HYDRO_HOME: ${GITHUB_WORKSPACE}/build/AMReX-Hydro
+         SUNDIALS_HOME : ${GITHUB_WORKSPACE}/build/sundials
       working-directory: ./Exec/RegTests/HotBubble/
       run: |
         make TPL COMP=gnu USE_MPI=TRUE
@@ -170,6 +175,7 @@ jobs:
          PELE_PHYSICS_HOME: ${GITHUB_WORKSPACE}/build/PelePhysics
          PELELMEX_HOME: ${GITHUB_WORKSPACE}
          AMREX_HYDRO_HOME: ${GITHUB_WORKSPACE}/build/AMReX-Hydro
+         SUNDIALS_HOME : ${GITHUB_WORKSPACE}/build/sundials
       working-directory: ./Exec/RegTests/HotBubble/
       run: |
         make TPL COMP=gnu DEBUG=TRUE USE_MPI=FALSE
@@ -197,6 +203,7 @@ jobs:
          PELE_PHYSICS_HOME: ${GITHUB_WORKSPACE}/build/PelePhysics
          PELELMEX_HOME: ${GITHUB_WORKSPACE}
[…]
