### 2023-08-29 00bb8e1
Add SUNDIALS path to readme and tutorials
