+++ Docs/source/manual/Tutorials_BFSFlame.rst
@@ -34,13 +34,14 @@ Follow the steps listed below to get to this point:
     cd PeleLMeX/Exec/RegTests/EB_BackwardStepFlame
 
 Note that the makefile system is set up such that default paths are automatically set to the
-submodules obtained with the recursive *git clone*, however the user can set its own dependencies
+submodules obtained with the recursive *git clone*, however the user can set their own dependencies
 in the `GNUmakefile` by updating the top-most lines as follows: ::
 
        PELELMEX_HOME     = <path_to_PeleLMeX>
        AMREX_HOME        = <path_to_MyAMReX>
        AMREX_HYDRO_HOME  = <path_to_MyAMReXHydro>
        PELE_PHYSICS_HOME = <path_to_MyPelePhysics>
+       SUNDIALS_HOME     = <path_to_MySUNDIALS>
 
 or directly through shell environement variables (using *bash* for instance): ::
 
@@ -48,6 +49,10 @@ or directly through shell environement variables (using *bash* for instance): ::
        export AMREX_HOME=<path_to_MyAMReX>
        export AMREX_HYDRO_HOME=<path_to_MyAMReXHydro>
        export PELE_PHYSICS_HOME=<path_to_MyPelePhysics>
+       export SUNDIALS_HOME=<path_to_MySUNDIALS>
+
+Note that using the first option will overwrite any
+environement variables you might have previously defined when using this `GNUmakefile`.
 
 You're good to go !
 

[…]
