### 2023-08-29 4e8164e
typo fixes and minor changes in Docs and Readmes
