+++ README.md
@@ -12,7 +12,8 @@
 ## Overview
 
 *PeleLMeX* is a solver for high fidelity reactive flow simulations, namely direct numerical simulation (DNS) and large eddy simulation (LES).
-The solver combines a low Mach number approach, adaptive mesh refinement (AMR) and high performance computing (HPC) to provide
+The solver combines a low Mach number approach, adaptive mesh refinement (AMR), embedded boundary (EB) geometry treatment and high performance
+computing (HPC) to provide
 a flexible tool to address research questions on platforms ranging from small workstations to the world's largest GPU-accelerated supercomputers.
 
 *PeleLMeX* is part of the [Pele combustion Suite](https://amrex-combustion.github.io/).
@@ -26,6 +27,9 @@ a flexible tool to address research questions on platforms ranging from small wo
 solver [incflo](https://github.com/AMReX-Codes/incflo). It solves of the multispecies reactive Navier-Stokes equations
 in the low Mach number limit as described in the [documentation](https://amrex-combustion.github.io/PeleLMeX/manual/html/index.html).
 It inherits most of *PeleLM* algorithmic features, but differs significantly in its implementation stemming from the non-subcycling approach.
+*PeleLM* is no longer under active development; *PeleLMeX* should be used for simulations of low Mach number reacting flows and
+[*PeleC*](https://github.com/AMReX-Combustion/PeleC) for simulations of flows with higher Mach numbers where compressibility effects are
+significant.
 
 A overview of *PeleLMeX* controls is provided in the [documentation](https://amrex-combustion.github.io/PeleLMeX/manual/html/LMeXControls.html).
 
@@ -59,16 +63,16 @@ Finally, when building with GPU support, CUDA >= 11 is required with NVIDIA GPUs
 
 ### Download
 
-The prefered method consist in cloning the *PeleLMeX* and its submodules using a recursive `git clone`:
-([amrex](https://github.com/AMReX-Codes/amrex), [AMReX-Hydro](https://github.com/AMReX-Codes/AMReX-Hydro), [PelePhysics](https://github.com/AMReX-Combustion/PelePhysics), [PeleMP](https://github.com/AMReX-Combustion/PeleMP)), and [SUNDIALS](https://github.com/LLNL/sundials) using:
+The prefered method consists of cloning *PeleLMeX* and its submodules
+([amrex](https://github.com/AMReX-Codes/amrex), [AMReX-Hydro](https://github.com/AMReX-Codes/AMReX-Hydro), [PelePhysics](https://github.com/AMReX-Combustion/PelePhysics), [PeleMP](https://github.com/AMReX-Combustion/PeleMP)), and [SUNDIALS](https://github.com/LLNL/sundials) using a recursive `git clone`:
 
 ```
 git clone --recursive https://github.com/AMReX-Combustion/PeleLMeX.git
 ```
 
-Alternatively, you can use separate `git clone` of the each of the submodules.
+Alternatively, you can use a separate `git clone` of the each of the submodules.
 The default location for *PeleLMeX* dependencies is the `Submodules` folder but you optionnally
-setup the following environment variables (e.g. using bash) to an other location:
+setup the following environment variables (e.g. using bash) to any other location:
 
 ```
 export PELELMEX_HOME=<path_to_PeleLMeX>
@@ -94,7 +98,7 @@ cd PeleLMeX/Exec/RegTests/HotBubble
 
 If this is a clean install, you will need to make the third party libraries with: `make TPL` (note: if on macOS, you might need to specify `COMP=llvm` in the `make` statements).
 
[…]
