+++ Exec/RegTests/README.md
@@ -1,6 +1,6 @@
 # PeleLMeX testing
 
 This folder contains a set of small-scale cases testing the various functionalities of PeleLMeX.
-A number of these cases are automatically compiled or compiled & run upon pull requests into the development branch,
-A short description of all the case is provided in the README file of each folder, stating the tested PeleLMeX capabilities and 
-,when available, point towards a tutorial or documentation page.
+A number of these cases are automatically compiled or compiled & run upon pull requests into the development branch.
+A short description of all the case is provided in the README file in each folder, stating the tested PeleLMeX capabilities and,
+when available, pointing towards a tutorial or documentation page.
