+++ Exec/RegTests/HITDecay/README.md
@@ -1,3 +1,3 @@
 ## HITDecay
-A 3D decaying HIT case, where the initial solution is generated with a Passot-Pouquet spectrum. Testing basic incompressible
+A 3D decaying Homogeneous Isotropic Turbulence (HIT) case, where the initial solution is generated with a Passot-Pouquet spectrum. Testing the basic incompressible
 integration algorithm and Large Eddy Simulation implementation.

[…]
