### 2023-09-07 3358b7e
use proper dx in flux divergence
