+++ Source/PeleLMUtils.cpp
@@ -334,8 +334,7 @@ void PeleLM::intFluxDivergenceLevelEB(int lev,
    geom[lev].GetVolume(volume);
 
    // Get area
-   const Real* dx = Geom(lev).CellSize();
-   const Real tmp_dx = dx[0];
+   const GpuArray<Real, AMREX_SPACEDIM> dx = Geom(lev).CellSizeArray();
 #if ( AMREX_SPACEDIM == 2 )
    Real areax = dx[1];
    Real areay = dx[0];
@@ -403,7 +402,7 @@ void PeleLM::intFluxDivergenceLevelEB(int lev,
                                        AMREX_D_DECL(afrac_x, afrac_y, afrac_z),
                                        AMREX_D_DECL(areax, areay, areaz),
                                        ebflux, ebarea,
-                                       vol, tmp_dx, scale, divergence);
+                                       vol, dx[0], scale, divergence);
                for (int n = 0; n < ncomp; n++) {
                   divergence(i,j,k,n) *= vfracinv;
                }
