### 2023-09-08 a7364cb
Fix bug for isothermal EB on GPU (#240)
* use proper dx in flux divergence
