### 2023-09-13 d1b97ad
Add a couple of examples of PeleLMeX application to README
