+++ Exec/RegTests/EB_FlowPastCylinder/EBUserDefined.H
@@ -7,9 +7,10 @@ using namespace amrex;
 #include <AMReX_EB2.H>
 #include <AMReX_EB2_IF.H>
 void
-EBUserDefined(const Geometry& /*geom*/,
-              const int /*required_coarsening_level*/,
-              const int /*max_coarsening_level*/)
+EBUserDefined(
+  const Geometry& /*geom*/,
+  const int /*required_coarsening_level*/,
+  const int /*max_coarsening_level*/)
 {
   // ParmParse your geometry parameters
 
@@ -21,32 +22,35 @@ EBUserDefined(const Geometry& /*geom*/,
 
   // We shoulnd't be here, copy this file in your run folder
   // and implement your geometry
-  Abort("In default EBUserDefined function! Shouldn't be here. Copy and edit this file for your needs");
+  Abort("In default EBUserDefined function! Shouldn't be here. Copy and edit "
+        "this file for your needs");
 }
 
 AMREX_GPU_DEVICE
 AMREX_FORCE_INLINE
 void
-setEBState(const amrex::Real xEBface[AMREX_SPACEDIM],
-           amrex::Real s_ext[NVAR],
-           const amrex::Real /*time*/,
-           amrex::GeometryData const& /*geomdata*/,
-           ProbParm const& /*prob_parm*/)
+setEBState(
+  const amrex::Real xEBface[AMREX_SPACEDIM],
+  amrex::Real s_ext[NVAR],
+  const amrex::Real /*time*/,
+  amrex::GeometryData const& /*geomdata*/,
+  ProbParm const& /*prob_parm*/)
 {
-  if ( xEBface[1] > 0.02 ) {
-     s_ext[TEMP] = 500.0;
+  if (xEBface[1] > 0.02) {
+    s_ext[TEMP] = 500.0;
   } else {
-     s_ext[TEMP] = 300.0;
+    s_ext[TEMP] = 300.0;
   }
 }
 
 AMREX_GPU_DEVICE
 AMREX_FORCE_INLINE
 void
-setEBType(const amrex::Real xEBface[AMREX_SPACEDIM],
-          amrex::Real & EBflagType,
-          amrex::GeometryData const& /*geomdata*/,
-          ProbParm const& /*prob_parm*/)
+setEBType(
+  const amrex::Real xEBface[AMREX_SPACEDIM],
+  amrex::Real& EBflagType,
+  amrex::GeometryData const& /*geomdata*/,
+  ProbParm const& /*prob_parm*/)
 {
   EBflagType = 1.0;
 }
