+++ Source/Efield/LinOps/AMReX_MLABecCecLap_1D_K.H
@@ -5,429 +5,537 @@ namespace amrex {
 
 constexpr Real eps = 1.0e-6;
 
-inline Real getFaceStateUpwind ( Real const &edgeVel, Real const &leftState, Real const &rightState ) {
-   Real edstate = 0.0;
-   if ( edgeVel > eps ) {
-      edstate = leftState;
-   }
-   if ( edgeVel < -eps ) {
-      edstate = rightState;
-   }
-   if ( fabs(edgeVel) <= eps ) {
-      edstate =  0.5 * ( leftState + rightState );
-   }
-   return edstate;
-}
-
-inline Real getCellUpwindCoeff ( Real const &leftedgeVel, Real const &rightedgeVel ) {
-
-   Real coeff = 0.0;
-   if ( leftedgeVel*rightedgeVel < 0.0 ) { // Case where velocity change sign, use centered + art. visc
-      Real maxvel = std::max(fabs(leftedgeVel),fabs(rightedgeVel));
-         coeff += 0.5*(rightedgeVel-leftedgeVel); // centered part
-         coeff += maxvel; // art. visc.
-   } else { // same sign velocity, use upwind
-      if ( leftedgeVel < -eps )
-         coeff -= leftedgeVel;
-      else if ( fabs(leftedgeVel) <= eps )
-         coeff -= 0.5*leftedgeVel;
-      if ( rightedgeVel > eps )
-         coeff += rightedgeVel;
-      else if ( fabs(rightedgeVel) <= eps )
-         coeff += 0.5*rightedgeVel;
-   }
-
-   return coeff;
-}
-
-inline Real getLeftCellUpwindCoeff ( Real const &leftedgeVel, Real const &rightedgeVel ) {
-
-   Real coeff = 0.0;
-   if ( leftedgeVel*rightedgeVel < 0.0 ) {
-      Real maxvel = std::max(fabs(leftedgeVel),fabs(rightedgeVel));
-      coeff -= 0.5*leftedgeVel;
-      coeff -= 0.5*maxvel;
-   } else {
-      if ( leftedgeVel > eps )
-         coeff -= leftedgeVel;
-      else if ( fabs(leftedgeVel) < eps )
-         coeff -= 0.5*leftedgeVel;
-   }
-   return coeff;
-}
-
-inline Real getRightCellUpwindCoeff ( Real const &leftedgeVel, Real const &rightedgeVel ) {
-
-   Real coeff = 0.0;
-   if ( leftedgeVel*rightedgeVel < 0.0 ) {
-      Real maxvel = std::max(fabs(leftedgeVel),fabs(rightedgeVel));
-      coeff += 0.5*rightedgeVel;
-      coeff += 0.5*maxvel;
-   } else {
-      if ( rightedgeVel < -eps )
-         coeff += rightedgeVel;
-      else if ( fabs(rightedgeVel) < eps )
-         coeff += 0.5*rightedgeVel;
-   }
-   return coeff;
-}
-
-inline Real getCellUpwindCoeff ( Real const &leftedgeVel, Real const &rightedgeVel,
-                                 Real const &relaxleft  , Real const &relaxright ) {
-
-   Real coeff = 0.0;
-   if ( leftedgeVel*relaxleft*rightedgeVel*relaxright < 0.0 ) { // Case where velocity change sign, use centered + art. visc
-      Real maxvel = std::max(fabs(leftedgeVel*relaxleft),fabs(rightedgeVel*relaxright));
-         coeff += 0.5*(rightedgeVel*relaxright-leftedgeVel*relaxleft); // centered part
[…]

[…]
