+++ Exec/Cases/DiffBunsen2D/pelelm_prob_parm.H
@@ -11,9 +11,9 @@ struct ProbParm
   // Definition of the geometry (in meters)
   amrex::Real Xmax = 0.015; // Maximum size in x direction of the calcuation
                             // zone
-  amrex::Real Xf = 0.0011; // Radius of the fuel pipe
-  amrex::Real Xe = 0.0001; // Thickness of the extruded tube
-  amrex::Real Xc = 0.0113; // Size of the oxidizer fuel
+  amrex::Real Xf = 0.0011;  // Radius of the fuel pipe
+  amrex::Real Xe = 0.0001;  // Thickness of the extruded tube
+  amrex::Real Xc = 0.0113;  // Size of the oxidizer fuel
   amrex::Real Ymax =
     0.037; // Maximum size in y direction of the calculation zone == Position of
            // the electrode ine the experiment
