+++ .github/workflows/docs.yml
@@ -38,7 +38,7 @@ jobs:
            echo "::remove-matcher owner=sphinx-problem-matcher-loose-no-severity::"
            echo "::remove-matcher owner=sphinx-problem-matcher-loose::"
            echo "::remove-matcher owner=sphinx-problem-matcher::"
-           touch ${{runner.workspace}}/build-docs/sphinx/html/.nojekyll
+           touch ${{runner.workspace}}/build-docs/html/.nojekyll
       - name: Deploy
         if: github.event_name == 'push' && github.ref == 'refs/heads/development'
         uses: JamesIves/github-pages-deploy-action@releases/v3
