### 2023-09-14 3e9f585
Joss review update (#236)
* Typo on RegTests

* Update README with better overview, explicit requirements and clearer
get help instructions.

* Remove a couple of typos in FlameSheet turorial.

* Reduce duality between the use of PeleLM and PeleLMeX. Profiling
now use PeleLMeX.

* Update the scaling script with new TPROF region names.

* Couple of updates in the FS tuto.

* Explicitly point to open source MPI distributions in the README.

* Small comment update.

* Update the RegTests README.

* Update the HotBubble input files for the tutorial.

* Remove deprecated line in GNUmakefile.

* Rem
[…]
