+++ Exec/RegTests/HotBubble/README.md
@@ -1,0 +1,3 @@
+## HotBubble
+Case of a 2D/2D-RZ/3D bubble of light gases (either hotter or lighter mixture composition) lifted under the effect of 
+a gravity field. Testing of RZ algorthim and gravity forces.

[…]
