+++ manual/html/_sources/Tutorials_HotBubble.rst.txt
@@ -1,0 +1,356 @@
+.. highlight:: rst
+
+.. _sec:tutorialHB:
+
+Rising light bubble
+===================
+
+.. _sec:TUTO_HB::Intro:
+
+Introduction
+------------
+
+At the core of `PeleLMeX` algorithm is a variable-density, low Mach number Navier-Stokes solver based
+on a fractional step approach. This short tutorial presents the case of a rising light bubble under
+the influence of a gravity field, which exercice this algorithm without the added complexity
+of chemical reactions or embedded boundaries.
+
+..  _sec:TUTO_HB::PrepStep:
+
+Setting-up your environment
+---------------------------
+
+Getting a functioning environment in which to compile and run `PeleLMeX` is the first step of this tutorial.
+Follow the steps listed below to get to this point:
+
+#. The first step is to get `PeleLMeX` and its dependencies. To do so, use a recursive *git clone*: ::
+
+    git clone --recursive https://github.com/AMReX-Combustion/PeleLMeX.git
+
+#. Move into the Exec folder containing the ``HotBubble``. To do so: ::
+
+    cd PeleLMeX/Exec/RegTests/HotBubble
+
+Note that the makefile system is set up such that default paths are automatically set to the
+submodules obtained with the recursive *git clone*, however the user can set their own dependencies
+in the `GNUmakefile` by updating the top-most lines as follows: ::
+
+       PELELMEX_HOME     = <path_to_PeleLMeX>
+       AMREX_HOME        = <path_to_MyAMReX>
+       AMREX_HYDRO_HOME  = <path_to_MyAMReXHydro>
+       PELE_PHYSICS_HOME = <path_to_MyPelePhysics>
+       SUNDIALS_HOME     = <path_to_MySUNDIALS>
+
+or directly through shell environment variables (using *bash* for instance): ::
+
+       export PELELMEX_HOME=<path_to_PeleLMeX>
+       export AMREX_HOME=<path_to_MyAMReX>
+       export AMREX_HYDRO_HOME=<path_to_MyAMReXHydro>
+       export PELE_PHYSICS_HOME=<path_to_MyPelePhysics>
+       export SUNDIALS_HOME=<path_to_MySUNDIALS>
+
+Note that using the first option will overwrite any
+environment variables you might have previously defined when using this `GNUmakefile`.
+
+You're good to go !
+
+Case setup
+----------
+
+A `PeleLMeX` case folder contains a minimal set of files to enable compilation,
+and the reader is referred to the FlameSheet tutorial :doc:`Tutorials_FlameSheet` for a
+more detailed description of `PeleLMeX` case setup. The case of interest for this
+tutorial can be found in `PeleLMeX` Exec folder: ::
+
+    Exec/RegTests/HotBubble
+
+Geometry, grid and boundary conditions
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+This simulation is performed on a 0.016x0.032 :math:`m^2` 2D computational domain,
+with the bottom left corner located at (0.0:0.0) and the top right corner at (0.016:0.032).
+Two versions of that domain will be considered: 1) a 2D-cartesian case, 2) a 2D-RZ
+case where the :math:`x`-low is the symmetry axis of the azimuthally homogeneous case. In both cases, the
+:math:`x`-low boundary is treated as symmetry BC while :math:`x`-high is treated as an adiabatic slip wall,
+the :math:`y`-low boundary is treated as an adiabatic slip wall, while the :math:`y`-high boundary is an ``Outflow``.
+
+All of the geometrical information can be specified the first two blocks of the input file (`input.2d-regt_sym`): ::
+
[…]
