+++ manual/html/_sources/Tutorials_FlowPastCyl.rst.txt
@@ -35,26 +35,26 @@ Follow the steps listed below to get to this point:
 
     cd PeleLMeX/Exec/RegTests/EB_FlowPastCylinder
 
-#. Finally, setup the environment variables providing paths to `PeleLMeX` and its dependencies. This can done in
-   one of two ways:
-
-   #. Directly into the `GNUmakefile` by updating the top-most lines as follows: ::
+Note that the makefile system is set up such that default paths are automatically set to the
+submodules obtained with the recursive *git clone*, however the user can set their own dependencies
+in the `GNUmakefile` by updating the top-most lines as follows: ::
 
        PELELMEX_HOME     = <path_to_PeleLMeX>
-       AMREX_HOME        =${PELELMEX_HOME}/Submodules/amrex
-       AMREX_HYDRO_HOME  =${PELELMEX_HOME}/Submodules/AMReX-Hydro
-       PELE_PHYSICS_HOME =${PELELMEX_HOME}/Submodules/PelePhysics
+       AMREX_HOME        = <path_to_MyAMReX>
+       AMREX_HYDRO_HOME  = <path_to_MyAMReXHydro>
+       PELE_PHYSICS_HOME = <path_to_MyPelePhysics>
+       SUNDIALS_HOME     = <path_to_MySUNDIALS>
 
-
-   #. Exporting shell environement variables (using *bash* for instance): ::
+or directly through shell environment variables (using *bash* for instance): ::
 
        export PELELMEX_HOME=<path_to_PeleLMeX>
-       export AMREX_HOME=${PELELMEX_HOME}/Submodules/amrex
-       export AMREX_HYDRO_HOME=${PELELMEX_HOME}/Submodules/AMReX-Hydro
-       export PELE_PHYSICS_HOME=${PELELMEX_HOME}/Submodules/PelePhysics
+       export AMREX_HOME=<path_to_MyAMReX>
+       export AMREX_HYDRO_HOME=<path_to_MyAMReXHydro>
+       export PELE_PHYSICS_HOME=<path_to_MyPelePhysics>
+       export SUNDIALS_HOME=<path_to_MySUNDIALS>
 
-   Both options require to provide the path to where you cloned `PeleLMeX`. Note that using the first option will overwrite any
-   environement variables you might have previously defined when using this `GNUmakefile`.
+Note that using the first option will overwrite any
+environment variables you might have previously defined when using this `GNUmakefile`.
 
 You're good to go !
 
@@ -212,11 +212,13 @@ Here, the model ``air``, only contains 2 species (O2 and N2). The user is referr
     Eos_Model       := Fuego
     Transport_Model := Constant
 
-Finally, `PeleLMeX` utilizes the chemical kinetic ODE integrator `CVODE <https://computing.llnl.gov/projects/sundials/cvode>`_. This Third Party Librabry (TPL) is not shipped with the `PeleLMeX` distribution but can be readily installed through the makefile system of `PeleLMeX`. Note that compiling Sundials is necessary even if the simualtion do not involve reactions. To do so, type in the following command: ::
+Finally, `PeleLMeX` utilizes the chemical kinetic ODE integrator `CVODE <https://computing.llnl.gov/projects/sundials/cvode>`_. This
+Third Party Librabry (TPL) is shipped as a submodule of the `PeleLMeX` distribution and can be readily installed through the makefile system
+of `PeleLMeX`. To do so, type in the following command: ::
 
     make -j4 TPL
 
-Note that the installation of `CVODE` requires CMake 3.12.1 or higher.
+Note that the installation of `CVODE` requires CMake 3.23.1 or higher.
 
 You are now ready to build your first `PeleLMeX` executable !! Type in: ::
 

[…]
