+++ Source/PeleLMTransportProp.cpp
@@ -8,8 +8,10 @@
 
 using namespace amrex;
 
-void PeleLM::calcTurbViscosity(const TimeStamp &a_time) {
-   BL_PROFILE("PeleLMeX::calcTurbViscosity()");
+void
+PeleLM::calcTurbViscosity(const TimeStamp& a_time)
+{
+  BL_PROFILE("PeleLMeX::calcTurbViscosity()");
 
   // We shouldn't be here unless we're doing LES
   AMREX_ALWAYS_ASSERT(m_do_les);
@@ -43,210 +45,21 @@ void PeleLM::calcTurbViscosity(const TimeStamp &a_time) {
   for (int lev = 0; lev <= finest_level; ++lev) {
     auto ldata_p = getLevelDataPtr(lev, a_time);
 
-     // Get density and cp (if needed) at faces for computing turbulent transport properties
-     amrex::Array<amrex::MultiFab, AMREX_SPACEDIM> dens_fc;
-     amrex::Array<amrex::MultiFab, AMREX_SPACEDIM> cp_fc;
-     amrex::MultiFab cp_cc;
-     const auto& ba = ldata_p->state.boxArray();
-     const auto& dm = ldata_p->state.DistributionMap();
-     const auto& factory = ldata_p->state.Factory();
-     if (m_incompressible) {
-       // just set density as a constant; don't need to worry about cp
-       for (int idim = 0; idim < AMREX_SPACEDIM; ++idim) {
-         dens_fc[idim].define(amrex::convert(ba,IntVect::TheDimensionVector(idim)), dm,
-                              1, 0, MFInfo(), factory);
-         dens_fc[idim].setVal(m_rho);
-       }
-     } else {
-       // get cp_cc (valid in 1 grow cell for interpolation to FCs)
-       int ngrow = 1;
-       cp_cc.define(ba, dm, 1, ngrow, MFInfo(), factory);
-       auto const& state_arr      = ldata_p->state.const_arrays();
-       auto const& cp_arr       = cp_cc.arrays();
-       amrex::ParallelFor(cp_cc, cp_cc.nGrowVect(), [=]
-                          AMREX_GPU_DEVICE (int box_no, int i, int j, int k) noexcept
-                          {
-                            getCpmixGivenRYT( i, j, k,
-                                              Array4<Real const>(state_arr[box_no], DENSITY),
-                                              Array4<Real const>(state_arr[box_no], FIRSTSPEC),
-                                              Array4<Real const>(state_arr[box_no], TEMP),
-                                              Array4<Real      >(cp_arr[box_no]) );
-                          });
-       Gpu::streamSynchronize();
-
-       // this function really just interpolates CCs to FCs in this case
-       int doZeroVisc = 0;
-       auto bcRec = fetchBCRecArray(DENSITY,1);
-       dens_fc = getDiffusivity(lev,DENSITY,1,doZeroVisc,{bcRec},ldata_p->state);
-       cp_fc   = getDiffusivity(lev,0      ,1,doZeroVisc,{bcRec},cp_cc);
-     }
-
-     // Now we compute the turbulent viscosity at faces
-     for (int idim = 0; idim < AMREX_SPACEDIM; ++idim) {
-       auto const& velgrad_arr = GradVel[lev][idim].const_arrays();
-       auto const& dens_arr = dens_fc[idim].const_arrays();
-       auto const& mut_arr = ldata_p->visc_turb_fc[idim].arrays();
-       const amrex::Real vol = AMREX_D_TERM(  geom[lev].CellSize(0),
-                                            * geom[lev].CellSize(1),
-                                            * geom[lev].CellSize(2));
[…]
