+++ .github/workflows/docs.yml
@@ -21,7 +21,7 @@ jobs:
       - name: Clone
         uses: actions/checkout@v3
       - name: Python
-        uses: actions/setup-python@v2
+        uses: actions/setup-python@v3
         with:
           python-version: '3.x'
       - name: Dependencies
@@ -41,7 +41,7 @@ jobs:
            touch ${{runner.workspace}}/build-docs/html/.nojekyll
       - name: Deploy
         if: github.event_name == 'push' && github.ref == 'refs/heads/development'
-        uses: JamesIves/github-pages-deploy-action@releases/v3
+        uses: JamesIves/github-pages-deploy-action@releases/v4
         with:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           BRANCH: gh-pages
