+++ Source/PeleLMSetup.cpp
@@ -792,779 +792,6 @@ void
 PeleLM::readGridFile(
   std::string grid_file, amrex::Vector<amrex::BoxArray>& input_ba)
 {
-  return b;
-}
-#ifdef PELE_USE_EFIELD
-static Box
-grow_box_by_one(const Box& b)
-{
-  return amrex::grow(b, 1);
-}
-#endif
-static Box
-grow_box_by_two(const Box& b)
-{
-  return amrex::grow(b, 2);
-}
-
-void
-PeleLM::Setup()
-{
-  BL_PROFILE("PeleLM::Setup()");
-
-  m_wall_start = amrex::ParallelDescriptor::second();
-
-  // Ensure grid is isotropic
-  {
-    auto const dx = geom[0].CellSizeArray();
-    AMREX_ALWAYS_ASSERT(AMREX_D_TERM(
-      , amrex::almostEqual(dx[0], dx[1]), &&amrex::almostEqual(dx[1], dx[2])));
-  }
-  // Print build info to screen
-  const char* githash1 = buildInfoGetGitHash(1);
-  const char* githash2 = buildInfoGetGitHash(2);
-  const char* githash3 = buildInfoGetGitHash(3);
-  const char* githash4 = buildInfoGetGitHash(4);
-  amrex::Print() << "\n ================= Build infos =================\n";
-  amrex::Print() << " PeleLMeX    git hash: " << githash1 << "\n";
-  amrex::Print() << " AMReX       git hash: " << githash2 << "\n";
-  amrex::Print() << " PelePhysics git hash: " << githash3 << "\n";
-  amrex::Print() << " AMReX-Hydro git hash: " << githash4 << "\n";
-  amrex::Print() << " ===============================================\n";
-
-#ifdef PELELM_USE_SOOT
-  soot_model = new SootModel{};
-#endif
-
-  // Read PeleLM parameters
-  readParameters();
-
-#ifdef AMREX_USE_EB
-  makeEBGeometry();
-#endif
-
-  // Setup the state variables
-  variablesSetup();
-
-  // Derived variables
-  derivedSetup();
-
-  // Evaluate variables
-  evaluateSetup();
-
-  // Tagging setup
-  taggingSetup();
-
-#ifdef PELELM_USE_SPRAY
-  SpraySetup();
-#endif
-#ifdef PELELM_USE_SOOT
-  if (do_soot_solve) {
-    soot_model->define();
-  }
-#endif
-  // Diagnostics setup
-  createDiagnostics();
-
-  // Initialize Level Hierarchy data
-  resizeArray();
-
-  // Initialize EOS and others
-  if (!m_incompressible) {
-    amrex::Print() << " Initialization of Transport ... \n";
-    trans_parms.allocate();
-    if (m_les_verbose and m_do_les)
-      amrex::Print() << "    Using LES in transport with Sc = "
-                     << 1.0 / m_Schmidt_inv
-                     << " and Pr = " << 1.0 / m_Prandtl_inv << std::endl;
-    if (m_verbose and m_unity_Le)
-      amrex::Print() << "    Using Le = 1 transport with Sc = "
-                     << 1.0 / m_Schmidt_inv
-                     << " and Pr = " << 1.0 / m_Prandtl_inv << std::endl;
-    if (m_do_react) {
-      int reactor_type = 2;
-      int ncells_chem = 1;
-      amrex::Print() << " Initialization of chemical reactor ... \n";
-      m_chem_integrator = "ReactorNull";
-      ParmParse pp("peleLM");
-      pp.query("chem_integrator", m_chem_integrator);
[…]
