+++ CMake/SetAmrexOptions.cmake
@@ -2,7 +2,7 @@ set(USE_XSDK_DEFAULTS OFF)
 set(AMReX_SPACEDIM "${PELELMEX_DIM}" CACHE STRING "Number of physical dimensions" FORCE)
 set(AMReX_MPI ${PELELMEX_ENABLE_MPI})
 set(AMReX_OMP ${PELELMEX_ENABLE_OPENMP})
-set(AMReX_EB ON)
+set(AMReX_EB ${PELELMEX_ENABLE_EB})
 set(AMReX_PARTICLES ${PELELMEX_ENABLE_AMREX_PARTICLES})
 set(AMReX_PLOTFILE_TOOLS ${PELELMEX_ENABLE_FCOMPARE})
 set(AMReX_SUNDIALS OFF)
