### 2023-09-15 8f61bc6
Assimilate PeleLMeX to look more like PeleC (#245)
PeleC build system framework added to PeleLMeX.
