+++ Exec/Efield/PremBunsen3DKuhl/pelelm_prob.H
@@ -13,78 +13,84 @@
 
 AMREX_GPU_DEVICE
 AMREX_FORCE_INLINE
-void pelelm_initdata(int i, int j, int k,
-                     int is_incompressible,
-                     amrex::Array4<amrex::Real> const& state,
-                     amrex::Array4<amrex::Real> const& aux,
-                     amrex::GeometryData const& geomdata,
-                     ProbParm const& prob_parm,
-                     pele::physics::PMF::PmfData::DataContainer const *pmf_data)
+void
+pelelm_initdata(
+  int i,
+  int j,
+  int k,
+  int is_incompressible,
+  amrex::Array4<amrex::Real> const& state,
+  amrex::Array4<amrex::Real> const& aux,
+  amrex::GeometryData const& geomdata,
+  ProbParm const& prob_parm,
+  pele::physics::PMF::PmfData::DataContainer const* pmf_data)
 {
-    const amrex::Real* prob_lo = geomdata.ProbLo();
-    const amrex::Real* prob_hi = geomdata.ProbHi();
-    const amrex::Real* dx      = geomdata.CellSize();
+  const amrex::Real* prob_lo = geomdata.ProbLo();
+  const amrex::Real* prob_hi = geomdata.ProbHi();
+  const amrex::Real* dx = geomdata.CellSize();
 
-    AMREX_D_TERM(const amrex::Real x = prob_lo[0] + (i+0.5)*dx[0];,
-                 const amrex::Real y = prob_lo[1] + (j+0.5)*dx[1];,
-                 const amrex::Real z = prob_lo[2] + (k+0.5)*dx[2];);
+  AMREX_D_TERM(const amrex::Real x = prob_lo[0] + (i + 0.5) * dx[0];
+               , const amrex::Real y = prob_lo[1] + (j + 0.5) * dx[1];
+               , const amrex::Real z = prob_lo[2] + (k + 0.5) * dx[2];);
 
-    AMREX_D_TERM(const amrex::Real Lx = prob_hi[0] - prob_lo[0];,
-                 const amrex::Real Ly = prob_hi[1] - prob_lo[1];,
-                 const amrex::Real Lz = prob_hi[2] - prob_lo[2]);
+  AMREX_D_TERM(const amrex::Real Lx = prob_hi[0] - prob_lo[0];
+               , const amrex::Real Ly = prob_hi[1] - prob_lo[1];
+               , const amrex::Real Lz = prob_hi[2] - prob_lo[2]);
 
-    constexpr amrex::Real Pi = 3.14159265358979323846264338327950288;
+  constexpr amrex::Real Pi = 3.14159265358979323846264338327950288;
 
-    auto eos = pele::physics::PhysicsType::eos();
-    amrex::GpuArray<amrex::Real, NUM_SPECIES + 4> pmf_vals = {0.0};
-    amrex::Real u[3] = {0.0};
-    amrex::Real molefrac[NUM_SPECIES] = {0.0};
-    amrex::Real massfrac[NUM_SPECIES] = {0.0};
-    amrex::Real y1, y2;
-    amrex::Real splitx = prob_lo[0];
-    amrex::Real splity = prob_lo[1];
-    amrex::Real rad_xy = std::sqrt((x-splitx)*(x-splitx) + (y-splity)*(y-splity));
+  auto eos = pele::physics::PhysicsType::eos();
[…]
