+++ Exec/RegTests/HotBubble/pelelm_prob.H
@@ -12,75 +12,76 @@
 
 AMREX_GPU_DEVICE
 AMREX_FORCE_INLINE
-void pelelm_initdata(int i, int j, int k,
-                     int is_incompressible,
-                     amrex::Array4<amrex::Real> const& state,
-                     amrex::Array4<amrex::Real> const& aux,
-                     amrex::GeometryData const& geomdata,
-                     ProbParm const& prob_parm,
-                     pele::physics::PMF::PmfData::DataContainer const * pmf_data)
+void
+pelelm_initdata(
+  int i,
+  int j,
+  int k,
+  int is_incompressible,
+  amrex::Array4<amrex::Real> const& state,
+  amrex::Array4<amrex::Real> const& aux,
+  amrex::GeometryData const& geomdata,
+  ProbParm const& prob_parm,
+  pele::physics::PMF::PmfData::DataContainer const* pmf_data)
 {
-    const amrex::Real* prob_lo = geomdata.ProbLo();
-    const amrex::Real* prob_hi = geomdata.ProbHi();
-    const amrex::Real* dx      = geomdata.CellSize();
+  const amrex::Real* prob_lo = geomdata.ProbLo();
+  const amrex::Real* prob_hi = geomdata.ProbHi();
+  const amrex::Real* dx = geomdata.CellSize();
 
-    AMREX_D_TERM(const amrex::Real x = prob_lo[0] + (i+0.5)*dx[0];,
-                 const amrex::Real y = prob_lo[1] + (j+0.5)*dx[1];,
-                 const amrex::Real z = prob_lo[2] + (k+0.5)*dx[2];);
+  AMREX_D_TERM(const amrex::Real x = prob_lo[0] + (i + 0.5) * dx[0];
+               , const amrex::Real y = prob_lo[1] + (j + 0.5) * dx[1];
+               , const amrex::Real z = prob_lo[2] + (k + 0.5) * dx[2];);
 
-    AMREX_D_TERM(const amrex::Real Lx = prob_hi[0] - prob_lo[0];,
-                 const amrex::Real Ly = prob_hi[1] - prob_lo[1];,
-                 const amrex::Real Lz = prob_hi[2] - prob_lo[2]);
+  AMREX_D_TERM(const amrex::Real Lx = prob_hi[0] - prob_lo[0];
+               , const amrex::Real Ly = prob_hi[1] - prob_lo[1];
+               , const amrex::Real Lz = prob_hi[2] - prob_lo[2]);
 
-    constexpr amrex::Real Pi = 3.14159265358979323846264338327950288;
+  constexpr amrex::Real Pi = 3.14159265358979323846264338327950288;
 
-    auto eos = pele::physics::PhysicsType::eos();
-    amrex::Real massfrac[NUM_SPECIES] = {0.0};
+  auto eos = pele::physics::PhysicsType::eos();
+  amrex::Real massfrac[NUM_SPECIES] = {0.0};
 
-    massfrac[O2_ID] = 0.233;
-    massfrac[N2_ID] = 0.767;
+  massfrac[O2_ID] = 0.233;
+  massfrac[N2_ID] = 0.767;
 
-    state(i,j,k,TEMP) = prob_parm.T_mean;
-    amrex::Real x_c = prob_parm.is_sym ? prob_lo[0]
-                                       : prob_lo[0]+0.5*Lx;
+  state(i, j, k, TEMP) = prob_parm.T_mean;
[…]
