### 2023-09-15 dcc9d1d
bugprone-macro-parentheses
