### 2023-09-18 c97c160
Fix readability-braces-around-statements.
