### 2023-09-18 792f9a4
Fix readability-redundant-string-init.
