### 2023-09-18 15fc3be
Fix warns (#248)
* bugprone-macro-parentheses

* Turn on the ci

* modernize-make-unique

* minor thing

* minor

* header stuff

* dont check amrex-hydro

* Revert "modernize-make-unique"

This reverts commit 1f9bddf8df46e0760fe9b667dd978d8b732b389c.

* Fix quotes.

* Fix readability-qualified-auto.

* Fix readability-braces-around-statements.

* Fix modernize-use-override.

* Fix readability-redundant-string-init.

---------

Co-authored-by: Jon Rood <jon.rood@nrel.gov>
