+++ Source/PeleLMProjection.cpp
@@ -29,7 +29,7 @@ PeleLM::initialProjection()
       sigma[lev].reset(new MultiFab(
         grids[lev], dmap[lev], 1, nGhost, MFInfo(), *m_factory[lev]));
 
-      auto ldata_p = getLevelDataPtr(lev, AmrNewTime);
+      auto* ldata_p = getLevelDataPtr(lev, AmrNewTime);
 
 #ifdef AMREX_USE_OMP
 #pragma omp parallel if (Gpu::notInLaunchRegion())
@@ -88,7 +88,7 @@ PeleLM::initialProjection()
   // Set back press and gpress to zero and restore divu
   // and rescale velocity if 2D-RZ
   for (int lev = 0; lev <= finest_level; lev++) {
-    auto ldata_p = getLevelDataPtr(lev, AmrNewTime);
+    auto* ldata_p = getLevelDataPtr(lev, AmrNewTime);
     ldata_p->press.setVal(0.0);
     ldata_p->gp.setVal(0.0);
     if (!m_incompressible && m_has_divu) {
@@ -140,7 +140,7 @@ PeleLM::initialPressProjection()
       sigma[lev].reset(new MultiFab(
         grids[lev], dmap[lev], 1, nGhost, MFInfo(), *m_factory[lev]));
 
-      auto ldata_p = getLevelDataPtr(lev, AmrNewTime);
+      auto* ldata_p = getLevelDataPtr(lev, AmrNewTime);
 
 #ifdef AMREX_USE_OMP
 #pragma omp parallel if (Gpu::notInLaunchRegion())
@@ -218,12 +218,13 @@ PeleLM::velocityProjection(
 
   if (!incremental) {
     Vector<std::unique_ptr<MultiFab>> rhoHalf(finest_level + 1);
-    if (!m_incompressible)
+    if (!m_incompressible) {
       rhoHalf = getDensityVect(a_rhoTime);
+    }
     for (int lev = 0; lev <= finest_level; ++lev) {
 
-      auto ldataOld_p = getLevelDataPtr(lev, AmrOldTime);
-      auto ldataNew_p = getLevelDataPtr(lev, AmrNewTime);
+      auto* ldataOld_p = getLevelDataPtr(lev, AmrOldTime);
+      auto* ldataNew_p = getLevelDataPtr(lev, AmrNewTime);
 
 #ifdef AMREX_USE_OMP
 #pragma omp parallel if (Gpu::notInLaunchRegion())
@@ -255,8 +256,8 @@ PeleLM::velocityProjection(
   // define "vel" to be U^{np1*} - U^{n} rather than U^{np1*}
   if (incremental) {
     for (int lev = 0; lev <= finest_level; ++lev) {
-      auto ldataOld_p = getLevelDataPtr(lev, AmrOldTime);
-      auto ldataNew_p = getLevelDataPtr(lev, AmrNewTime);
+      auto* ldataOld_p = getLevelDataPtr(lev, AmrOldTime);
+      auto* ldataNew_p = getLevelDataPtr(lev, AmrNewTime);
       MultiFab::Subtract(
         ldataNew_p->state, ldataOld_p->state, VELX, VELX, AMREX_SPACEDIM, 0);
     }
@@ -271,8 +272,9 @@ PeleLM::velocityProjection(
     EB_set_covered(*vel[lev], 0.0);
 #endif
     vel[lev]->setBndry(0.0);
-    if (!incremental)
+    if (!incremental) {
       setInflowBoundaryVel(*vel[lev], lev, AmrNewTime);
+    }
     scaleProj_RZ(lev, *vel[lev]);
   }
 
[…]
