+++ Source/PeleLMDeriveFunc.H
@@ -1,5 +1,5 @@
-#ifndef PELELM_DERIVEFUNC_H_
-#define PELELM_DERIVEFUNC_H_
+#ifndef PELELM_DERIVEFUNC_H
+#define PELELM_DERIVEFUNC_H
 
 #include <AMReX_FArrayBox.H>
 #include <AMReX_Geometry.H>
