+++ Source/Efield/LinOps/AMReX_MLABecCecLap_1D_K.H
@@ -1,5 +1,5 @@
-#ifndef AMREX_MLABECCECLAP_1D_K_H_
-#define AMREX_MLABECCECLAP_1D_K_H_
+#ifndef AMREX_MLABECCECLAP_1D_K_H
+#define AMREX_MLABECCECLAP_1D_K_H
 
 namespace amrex {
 
