+++ Source/Efield/LinOps/AMReX_MLCellABecCecLap_1D_K.H
@@ -1,5 +1,5 @@
-#ifndef AMREX_MLCELLABECCECLAP_1D_K_H_
-#define AMREX_MLCELLABECCECLAP_1D_K_H_
+#ifndef AMREX_MLCELLABECCECLAP_1D_K_H
+#define AMREX_MLCELLABECCECLAP_1D_K_H
 #include <AMReX_Config.H>
 
 namespace amrex {

[…]
