+++ .github/workflows/ci.yml
@@ -61,7 +61,7 @@ jobs:
           echo "::add-matcher::.github/problem-matchers/gcc.json"
           set -eu -o pipefail
           ccache -z
-          make -j ${{env.NPROCS}} DIM=3 KEEP_BUILDINFO_CPP=TRUE USE_CCACHE=TRUE COMP=gnu USE_MPI=FALSE DEBUG=TRUE 2>&1 | tee -a build-output.txt
+          make -j ${{env.NPROCS}} DIM=2 KEEP_BUILDINFO_CPP=TRUE USE_CCACHE=TRUE COMP=gnu USE_MPI=FALSE DEBUG=TRUE 2>&1 | tee -a build-output.txt
       - name: FlameSheet ccache report
         working-directory: ./Exec/RegTests/FlameSheet
         run: |
@@ -69,12 +69,12 @@ jobs:
           du -hs ${HOME}/.cache/ccache
       - name: FlameSheet debug test
         working-directory: ./Exec/RegTests/FlameSheet
-        run: ./PeleLMeX3d.gnu.DEBUG.ex flamesheet-drm19-3d.inp amr.max_step=1 amr.plot_int=-1 amr.check_int=-1
+        run: ./PeleLMeX3d.gnu.DEBUG.ex flamesheet-drm19-2d.inp amr.max_step=1 amr.plot_int=-1 amr.check_int=-1
       - name: FlameSheet debug report
         working-directory: ./Exec/RegTests/FlameSheet
         run: |
           egrep "warning:|error:" build-output.txt \
-            | egrep -v "Submodules/amrex|Submodules/sundials|GNUmakefile" \
+            | egrep -v "Submodules/amrex|Submodules/sundials|Submodules/AMReX-Hydro|GNUmakefile" \
             | egrep -v "ld: warning:" | sort | uniq \
             | awk 'BEGIN{i=0}{print $0}{i++}END{print "Warnings: "i}' > warnings.txt
           cat warnings.txt
@@ -99,7 +99,7 @@ jobs:
       #  working-directory: ./Exec/RegTests/EB-C10
       #  run: |
       #    egrep "warning:|error:" build-output.txt \
-      #      | egrep -v "Submodules/amrex|Submodules/sundials|GNUmakefile" \
+      #      | egrep -v "Submodules/amrex|Submodules/sundials|Submodules/AMReX-Hydro|GNUmakefile" \
       #      | egrep -v "ld: warning:" | sort | uniq \
       #      | awk 'BEGIN{i=0}{print $0}{i++}END{print "Warnings: "i}' > warnings.txt
       #    cat warnings.txt
@@ -124,7 +124,7 @@ jobs:
       #  working-directory: ./Exec/RegTests/TG
       #  run: |
       #    egrep "warning:|error:" build-output.txt \
-      #      | egrep -v "Submodules/amrex|Submodules/sundials|GNUmakefile" \
+      #      | egrep -v "Submodules/amrex|Submodules/sundials|Submodules/AMReX-Hydro|GNUmakefile" \
       #      | egrep -v "ld: warning:" | sort | uniq \
       #      | awk 'BEGIN{i=0}{print $0}{i++}END{print "Warnings: "i}' > warnings.txt
       #    cat warnings.txt
@@ -146,7 +146,7 @@ jobs:
       #  working-directory: ./Exec/RegTests/Soot-Flame
       #  run: |
       #    egrep "warning:|error:" build-output.txt \
-      #      | egrep -v "Submodules/amrex|Submodules/sundials|GNUmakefile|Submodules/PeleMP" \
+      #      | egrep -v "Submodules/amrex|Submodules/sundials|GNUmakefile|Submodules/PeleMP|Submodules/AMReX-Hydro" \
[…]
