### 2023-09-18 7147044
Fix readability-implicit-bool-conversion
