+++ Source/PeleLM.cpp
@@ -14,7 +14,7 @@ PeleLM::PeleLM() = default;
 
 PeleLM::~PeleLM()
 {
-  if (!m_incompressible) {
+  if (m_incompressible == 0) {
     trans_parms.deallocate();
     m_reactor->close();
   }
@@ -47,7 +47,7 @@ PeleLM::getLevelDataPtr(
   } else {
     m_leveldata_floating.reset(new LevelData(
       grids[lev], dmap[lev], *m_factory[lev], m_incompressible, m_has_divu,
-      m_nAux, m_nGrowState, m_use_soret, m_do_les));
+      m_nAux, m_nGrowState, m_use_soret, static_cast<int>(m_do_les)));
     Real time = getTime(lev, a_time);
     fillpatch_state(lev, time, m_leveldata_floating->state, m_nGrowState);
     // if (useUMac) {
@@ -64,7 +64,7 @@ PeleLM::getLevelDataPtr(
 PeleLM::LevelDataReact*
 PeleLM::getLevelDataReactPtr(int lev)
 {
-  if (m_do_react) {
+  if (m_do_react != 0) {
     return m_leveldatareact[lev].get();
   } else {
     return nullptr;
@@ -77,7 +77,7 @@ PeleLM::getStateVect(const TimeStamp& a_time)
   Vector<std::unique_ptr<MultiFab>> r;
   r.reserve(finest_level + 1);
   if (a_time == AmrOldTime) {
-    if (m_incompressible) {
+    if (m_incompressible != 0) {
       for (int lev = 0; lev <= finest_level; ++lev) {
         r.push_back(std::make_unique<MultiFab>(
           m_leveldata_old[lev]->state, amrex::make_alias, 0, AMREX_SPACEDIM));
@@ -89,7 +89,7 @@ PeleLM::getStateVect(const TimeStamp& a_time)
       }
     }
   } else {
-    if (m_incompressible) {
+    if (m_incompressible != 0) {
       for (int lev = 0; lev <= finest_level; ++lev) {
         r.push_back(std::make_unique<MultiFab>(
           m_leveldata_new[lev]->state, amrex::make_alias, 0, AMREX_SPACEDIM));
@@ -279,7 +279,7 @@ PeleLM::getIRVect()
 void
 PeleLM::averageDownState(const PeleLM::TimeStamp& a_time)
 {
-  int nCompState = (m_incompressible) ? AMREX_SPACEDIM : NVAR;
+  int nCompState = (m_incompressible) != 0 ? AMREX_SPACEDIM : NVAR;
   for (int lev = finest_level; lev > 0; --lev) {
     auto* ldataFine_p = getLevelDataPtr(lev, a_time);
     auto* ldataCrse_p = getLevelDataPtr(lev - 1, a_time);
