+++ Source/PeleLMBCfill.H
@@ -66,7 +66,7 @@ struct PeleLMCCFillExtDirState
 
           // If using TurbInflow, pass in the turbulent data. User can overwrite
           // if needed
-          if (m_do_turbInflow) {
+          if (m_do_turbInflow != 0) {
             if (n >= VELX && n < DENSITY) {
               if (!isnan(
                     state(iv, VELX + n))) { // During fillcoarsepatch, get here
@@ -97,7 +97,7 @@ struct PeleLMCCFillExtDirState
 
           // If using TurbInflow, pass in the turbulent data. User can overwrite
           // if needed
-          if (m_do_turbInflow) {
+          if (m_do_turbInflow != 0) {
             if (n >= VELX && n < DENSITY) {
               if (!isnan(
                     state(iv, VELX + n))) { // During fillcoarsepatch, get here

[…]
