+++ Exec/RegTests/CMakeLists.txt
@@ -3,14 +3,14 @@
 #add_subdirectory(EB_EnclosedVortex)
 #add_subdirectory(EB_FlowPastCylinder)
 #add_subdirectory(EB_PipeFlow)
-#add_subdirectory(EnclosedFlame)
-#add_subdirectory(EnclosedInjection)
+add_subdirectory(EnclosedFlame)
+add_subdirectory(EnclosedInjection)
 add_subdirectory(FlameSheet)
-#add_subdirectory(HITDecay)
-#add_subdirectory(HotBubble)
-#add_subdirectory(PeriodicCases)
+add_subdirectory(HITDecay)
+add_subdirectory(HotBubble)
+add_subdirectory(PeriodicCases)
 #add_subdirectory(SprayTest)
-#add_subdirectory(TaylorGreen)
-#add_subdirectory(TripleFlame)
-#add_subdirectory(TurbInflow)
+add_subdirectory(TaylorGreen)
+add_subdirectory(TripleFlame)
+add_subdirectory(TurbInflow)
 #add_subdirectory(Unit)
