+++ Exec/RegTests/EB_EnclosedVortex/CMakeLists.txt
@@ -1,4 +1,3 @@
-set(PELELMEX_ENABLE_PARTICLES OFF)
 set(PELELMEX_EOS_MODEL Fuego)
 set(PELELMEX_CHEMISTRY_MODEL air)
 set(PELELMEX_TRANSPORT_MODEL Constant)
