+++ Exec/RegTests/CMakeLists.txt
@@ -12,5 +12,7 @@ add_subdirectory(PeriodicCases)
 #add_subdirectory(SprayTest)
 add_subdirectory(TaylorGreen)
 add_subdirectory(TripleFlame)
-add_subdirectory(TurbInflow)
+if(PELELMEX_DIM GREATER 2)
+  add_subdirectory(TurbInflow)
+endif()
 #add_subdirectory(Unit)
