+++ .github/workflows/ci.yml
@@ -416,11 +416,11 @@ jobs:
         run: |
           source /etc/profile.d/rocm.sh
           which clang
-          which hipcc
+          which clang++
           cmake -B${{runner.workspace}}/build-hip \
           -DCMAKE_BUILD_TYPE:STRING=Release \
           -DCMAKE_VERBOSE_MAKEFILE:BOOL=ON \
-          -DCMAKE_CXX_COMPILER:STRING=$(which hipcc) \
+          -DCMAKE_CXX_COMPILER:STRING=$(which clang++) \
           -DCMAKE_C_COMPILER:STRING=$(which clang) \
           -DCMAKE_CXX_STANDARD:STRING=17 \
           -DPELELMEX_ENABLE_MPI:BOOL=OFF \
