### 2023-09-18 e245adf
Fix readability-implicit-bool-conversion (#251)
