+++ Source/PeleLMData.cpp
@@ -13,7 +13,7 @@ PeleLM::LevelData::LevelData(
   int a_use_soret,
   int a_do_les)
 {
-  if (a_incompressible) {
+  if (a_incompressible != 0) {
     state.define(ba, dm, AMREX_SPACEDIM, a_nGrowState, MFInfo(), factory);
   } else {
     state.define(ba, dm, NVAR, a_nGrowState, MFInfo(), factory);
@@ -22,23 +22,23 @@ PeleLM::LevelData::LevelData(
   press.define(
     amrex::convert(ba, IntVect::TheNodeVector()), dm, 1, 1, MFInfo(), factory);
   visc_cc.define(ba, dm, 1, 1, MFInfo(), factory);
-  if (a_do_les) {
+  if (a_do_les != 0) {
     for (int i = 0; i < AMREX_SPACEDIM; ++i) {
       visc_turb_fc[i].define(
         amrex::convert(ba, IntVect::TheDimensionVector(i)), dm, 1, 0, MFInfo(),
         factory);
-      if (!a_incompressible) {
+      if (a_incompressible == 0) {
         lambda_turb_fc[i].define(
           amrex::convert(ba, IntVect::TheDimensionVector(i)), dm, 1, 0,
           MFInfo(), factory);
       }
     }
   }
-  if (!a_incompressible) {
-    if (a_has_divu) {
+  if (a_incompressible == 0) {
+    if (a_has_divu != 0) {
       divu.define(ba, dm, 1, 1, MFInfo(), factory);
     }
-    if (a_use_soret) {
+    if (a_use_soret != 0) {
       diff_cc.define(ba, dm, 2 * NUM_SPECIES + 2, 1, MFInfo(), factory);
     } else {
       diff_cc.define(ba, dm, NUM_SPECIES + 2, 1, MFInfo(), factory);
@@ -98,7 +98,7 @@ PeleLM::AdvanceDiffData::AdvanceDiffData(
   int a_use_soret,
   int is_init)
 {
-  if (is_init) { // All I need is a container for a single diffusion term
+  if (is_init != 0) { // All I need is a container for a single diffusion term
     // Resize Vectors
     Dnp1.resize(a_finestLevel + 1);
 
@@ -112,11 +112,11 @@ PeleLM::AdvanceDiffData::AdvanceDiffData(
     Dn.resize(a_finestLevel + 1);
     Dnp1.resize(a_finestLevel + 1);
     Dhat.resize(a_finestLevel + 1);
-    if (a_use_wbar) {
+    if (a_use_wbar != 0) {
       Dwbar.resize(a_finestLevel + 1);
       wbar_fluxes.resize(a_finestLevel + 1);
     }
-    if (a_use_soret) {
+    if (a_use_soret != 0) {
       DT.resize(a_finestLevel + 1);
       soret_fluxes.resize(a_finestLevel + 1);
     }
@@ -129,7 +129,7 @@ PeleLM::AdvanceDiffData::AdvanceDiffData(
         ba[lev], dm[lev], NUM_SPECIES + 2, nGrowAdv, MFInfo(), *factory[lev]);
       Dhat[lev].define(
         ba[lev], dm[lev], NUM_SPECIES + 2, nGrowAdv, MFInfo(), *factory[lev]);
-      if (a_use_wbar) {
+      if (a_use_wbar != 0) {
         Dwbar[lev].define(
           ba[lev], dm[lev], NUM_SPECIES, nGrowAdv, MFInfo(), *factory[lev]);
         for (int idim = 0; idim < AMREX_SPACEDIM; ++idim) {
[…]
