### 2023-09-18 ce0b7bc
Try installing less HIP libraries.
