+++ Exec/RegTests/CMakeLists.txt
@@ -11,7 +11,7 @@ add_subdirectory(HotBubble)
 add_subdirectory(PeriodicCases)
 #add_subdirectory(SprayTest)
 add_subdirectory(TaylorGreen)
-add_subdirectory(TripleFlame)
+#add_subdirectory(TripleFlame)
 if(PELELMEX_DIM GREATER 2)
   add_subdirectory(TurbInflow)
 endif()
