### 2023-09-19 1cf98b8
Try updating sundials for HIP.
