### 2023-09-19 52acdbe
fix readability-convert-member-functions-to-static
