### 2023-09-19 b437926
Fix more
fix readability-inconsistent-declaration-parameter-name, readability-else-after-return, readability-redundant-smartptr-get, readability-simplify-boolean-expr, readability-container-data-pointer'
