+++ Source/PeleLM.H
@@ -1564,11 +1564,10 @@ public:
       a_time == AmrOldTime || a_time == AmrNewTime || a_time == AmrHalfTime);
     if (a_time == AmrOldTime) {
       return m_t_old[lev];
-    } else if (a_time == AmrNewTime) {
+    } if (a_time == AmrNewTime) {
       return m_t_new[lev];
-    } else {
-      return 0.5 * (m_t_old[lev] + m_t_new[lev]);
-    }
+    }       return 0.5 * (m_t_old[lev] + m_t_new[lev]);
+   
   }
 
   amrex::Vector<std::unique_ptr<amrex::MultiFab>>
