+++ Source/PeleLMDerive.cpp
@@ -128,13 +128,12 @@ PeleLMDeriveRec::variableComp(const std::string& a_name) const noexcept
 {
   if (n_derive == 1) {
     return 0;
-  } else {
-    for (int comp = 0; comp < n_derive; comp++) {
+  }     for (int comp = 0; comp < n_derive; comp++) {
       if (variable_names[comp] == a_name) {
         return comp;
       }
     }
-  }
+ 
   return -1;
 }
 
