### 2023-09-19 0ac5479
fix modernize-use-auto, modernize-use-nullptr, modernize-pass-by-value, modernize-use-default-member-init
