### 2023-09-19 cbd5f07
Turn on eb for clang-tidy
