+++ Source/PeleLMUtils.cpp
@@ -18,7 +18,8 @@ PeleLM::fluxDivergence(
   Real scale)
 {
   BL_PROFILE("PeleLMeX::fluxDivergence()");
-  if (intensiveFluxes) { // Fluxes are intensive -> need area scaling in div
+  if (intensiveFluxes != 0) { // Fluxes are intensive -> need area scaling in
+                              // div
     for (int lev = 0; lev <= finest_level; ++lev) {
       intFluxDivergenceLevel(
         lev, *a_divergence[lev], div_comp, a_fluxes[lev], flux_comp, ncomp,
@@ -47,7 +48,8 @@ PeleLM::fluxDivergence(
 {
 
   BL_PROFILE("PeleLMeX::fluxDivergence()");
-  if (intensiveFluxes) { // Fluxes are intensive -> need area scaling in div
+  if (intensiveFluxes != 0) { // Fluxes are intensive -> need area scaling in
+                              // div
     for (int lev = 0; lev <= finest_level; ++lev) {
       intFluxDivergenceLevelEB(
         lev, *a_divergence[lev], div_comp, a_fluxes[lev], flux_comp,
@@ -540,7 +542,7 @@ PeleLM::advFluxDivergence(
         Array4<Real const> const& ay = mf_ay.const_array(mfi);
         ParallelFor(
           bx, ncomp, [=] AMREX_GPU_DEVICE(int i, int j, int k, int n) noexcept {
-            if (!l_conserv_d[n]) {
+            if (l_conserv_d[n] == 0) {
               Real qavg = AMREX_D_TERM(
                 ax(i, j, k) * facex(i, j, k, n) +
                   ax(i + 1, j, k) * facex(i + 1, j, k, n),
@@ -560,7 +562,7 @@ PeleLM::advFluxDivergence(
       {
         ParallelFor(
           bx, ncomp, [=] AMREX_GPU_DEVICE(int i, int j, int k, int n) noexcept {
-            if (!l_conserv_d[n]) {
+            if (l_conserv_d[n] == 0) {
               Real qavg = AMREX_D_TERM(
                 facex(i, j, k, n) + facex(i + 1, j, k, n),
                 +facey(i, j, k, n) + facey(i, j + 1, k, n),
@@ -581,7 +583,7 @@ PeleLM::floorSpecies(const TimeStamp& a_time)
 {
   BL_PROFILE("PeleLMeX::floorSpecies()");
   AMREX_ASSERT(a_time == AmrOldTime || a_time == AmrNewTime);
-  if (!m_floor_species) {
+  if (m_floor_species == 0) {
     return;
   }
 
@@ -626,9 +628,9 @@ void
 PeleLM::resetCoveredMask()
 {
   BL_PROFILE("PeleLMeX::resetCoveredMask()");
-  if (m_resetCoveredMask) {
+  if (m_resetCoveredMask != 0) {
 
-    if (m_verbose) {
+    if (m_verbose != 0) {
       Print() << " Resetting fine-covered cells mask \n";
     }
 
@@ -686,7 +688,7 @@ PeleLM::resetCoveredMask()
       m_dmapChem[lev].reset(new DistributionMapping(*m_baChem[lev]));
 
       // Load balancing of the chemistry DMap
-      if (m_doLoadBalance) {
+      if (m_doLoadBalance != 0) {
         loadBalanceChemLev(lev);
       }
     }
@@ -702,7 +704,7 @@ PeleLM::resetCoveredMask()
     m_dmapChem[finest_level].reset(
       new DistributionMapping(*m_baChem[finest_level]));
 
-    if (m_doLoadBalance && m_max_grid_size_chem.min() > 0) {
[…]
