+++ Exec/CMakeLists.txt
@@ -1,1 +1,4 @@
+#add_subdirectory(Efield)
+#add_subdirectory(Production)
 add_subdirectory(RegTests)
+#add_subdirectory(UnitTests)

[…]
