### 2023-09-19 eea58f2
Fix readability-implicit-bool-conversion
Fix readability-string-compare
