+++ .github/workflows/ci.yml
@@ -376,7 +376,7 @@ jobs:
           du -hs ${HOME}/.cache/ccache
   GPU-AMD:
     name: GPU-HIP
-    needs: [Formatting, CPU-GNUmake]
+    #needs: [Formatting, CPU-GNUmake]
     runs-on: ubuntu-20.04
     steps:
       - name: Clone
@@ -399,7 +399,7 @@ jobs:
           echo 'export PATH=/opt/rocm/llvm/bin:/opt/rocm/bin:/opt/rocm/profiler/bin:/opt/rocm/opencl/bin:$PATH' \
               | sudo tee -a /etc/profile.d/rocm.sh
           sudo apt-get update
-          export TMPDIR=${{runner.workspace}}/.tmp && mkdir -p $TMPDIR && sudo apt-get install -y rocm-dev rocrand-dev rocprim-dev rocm-libs
+          export TMPDIR=${{runner.workspace}}/.tmp && mkdir -p $TMPDIR && sudo apt-get -o dir::cache::archives=$TMPDIR install -y rocm-dev rocrand-dev rocprim-dev rocm-libs
       - name: Install Ccache
         run: |
           wget https://github.com/ccache/ccache/releases/download/v4.8/ccache-4.8-linux-x86_64.tar.xz
