+++ .github/workflows/ci.yml
@@ -23,63 +23,63 @@ jobs:
         exclude: '.'
         extensions: 'H,h,cpp'
         clangFormatVersion: 16
-  CPU-GNUmake:
-    needs: Formatting
-    runs-on: ubuntu-latest
-    steps:
-      - name: Clone
-        uses: actions/checkout@v3
-        with:
-          submodules: true
-      - name: Setup
-        working-directory: ./Exec/RegTests/FlameSheet
-        run: |
-          echo "NPROCS=$(nproc)" >> $GITHUB_ENV
-          echo "CCACHE_COMPRESS=1" >> $GITHUB_ENV
-          echo "CCACHE_COMPRESSLEVEL=5" >> $GITHUB_ENV
-          echo "CCACHE_LOGFILE=${{github.workspace}}/ccache.log.txt" >> $GITHUB_ENV
-          echo "CCACHE_MAXSIZE=1G" >> $GITHUB_ENV
-          #printf "CC=gcc-10\nCXX=g++-10\nFC=gfortran-10\nF90=gfortran-10" \
-          # >> ${{github.workspace}}/Submodules/amrex/Tools/GNUMake/Make.local
-          make COMP=gnu USE_MPI=FALSE DEBUG=TRUE TPL
-          make COMP=llvm USE_MPI=FALSE DEBUG=TRUE PRECISION=FLOAT TPL
-      - name: Install Ccache
-        run: |
-          wget https://github.com/ccache/ccache/releases/download/v4.8/ccache-4.8-linux-x86_64.tar.xz
-          tar xvf ccache-4.8-linux-x86_64.tar.xz
-          sudo cp -f ccache-4.8-linux-x86_64/ccache /usr/local/bin/
-      - name: Set Up Ccache
-        uses: actions/cache@v3
-        with:
-          path: ~/.cache/ccache
-          key: ccache-${{github.workflow}}-${{github.job}}-git-${{github.sha}}
-          restore-keys: |
-               ccache-${{github.workflow}}-${{github.job}}-git-
-      - name: FlameSheet debug make
-        working-directory: ./Exec/RegTests/FlameSheet
-        run: |
-          echo "::add-matcher::.github/problem-matchers/gcc.json"
-          set -eu -o pipefail
-          ccache -z
-          make -j ${{env.NPROCS}} DIM=2 KEEP_BUILDINFO_CPP=TRUE USE_CCACHE=TRUE COMP=gnu USE_MPI=FALSE DEBUG=TRUE 2>&1 | tee -a build-output.txt
-      - name: FlameSheet ccache report
-        working-directory: ./Exec/RegTests/FlameSheet
-        run: |
-          ccache -s
-          du -hs ${HOME}/.cache/ccache
-      - name: FlameSheet debug test
-        working-directory: ./Exec/RegTests/FlameSheet
-        run: ./PeleLMeX2d.gnu.DEBUG.ex flamesheet-drm19-2d.inp amr.max_step=1 amr.plot_int=-1 amr.check_int=-1
-      - name: FlameSheet debug report
-        working-directory: ./Exec/RegTests/FlameSheet
-        run: |
-          egrep "warning:|error:" build-output.txt \
-            | egrep -v "Submodules/amrex|Submodules/sundials|Submodules/AMReX-Hydro|GNUmakefile" \
-            | egrep -v "ld: warning:" | sort | uniq \
-            | awk 'BEGIN{i=0}{print $0}{i++}END{print "Warnings: "i}' > warnings.txt
-          cat warnings.txt
-          export return=$(tail -n 1 warnings.txt | awk '{print $2}')
-          #exit ${return}
+  #CPU-GNUmake:
+  #  needs: Formatting
+  #  runs-on: ubuntu-latest
+  #  steps:
[…]
