+++ .github/workflows/ci.yml
@@ -394,12 +394,12 @@ jobs:
         run: |
           curl -O https://repo.radeon.com/rocm/rocm.gpg.key
           sudo apt-key add rocm.gpg.key
-          echo 'deb [arch=amd64] https://repo.radeon.com/rocm/apt/debian/ ubuntu main' \
+          echo 'deb [arch=amd64] https://repo.radeon.com/rocm/apt/5.6.1 ubuntu main' \
               | sudo tee /etc/apt/sources.list.d/rocm.list
           echo 'export PATH=/opt/rocm/llvm/bin:/opt/rocm/bin:/opt/rocm/profiler/bin:/opt/rocm/opencl/bin:$PATH' \
               | sudo tee -a /etc/profile.d/rocm.sh
           sudo apt-get update
-          export TMPDIR=${{runner.workspace}}/.tmp && mkdir -p $TMPDIR && sudo apt-get -o dir::cache::archives=$TMPDIR install -y rocm-dev rocrand-dev rocprim-dev rocm-libs
+          sudo apt-get install -y rocm-dev rocrand-dev rocprim-dev rocm-libs
       - name: Install Ccache
         run: |
           wget https://github.com/ccache/ccache/releases/download/v4.8/ccache-4.8-linux-x86_64.tar.xz
