+++ Exec/RegTests/EB_FlowPastCylinder/pelelm_prob.H
@@ -1,178 +1,0 @@
-#ifndef PELELM_PROB_H
-#define PELELM_PROB_H
-
-#include <AMReX_Geometry.H>
-#include <AMReX_FArrayBox.H>
-#include <AMReX_SPACE.H>
-
-#include <PeleLM_Index.H>
-#include <pelelm_prob_parm.H>
-#include <PMFData.H>
-#include <PelePhysics.H>
-
-AMREX_GPU_DEVICE
-AMREX_FORCE_INLINE
-void
-pelelm_initdata(
-  int i,
-  int j,
-  int k,
-  int is_incompressible,
-  amrex::Array4<amrex::Real> const& state,
-  amrex::Array4<amrex::Real> const& aux,
-  amrex::GeometryData const& geomdata,
-  ProbParm const& prob_parm,
-  pele::physics::PMF::PmfData::DataContainer const* /*pmf_data*/)
-{
-  const amrex::Real* prob_lo = geomdata.ProbLo();
-  const amrex::Real* prob_hi = geomdata.ProbHi();
-  const amrex::Real* dx = geomdata.CellSize();
-
-  AMREX_D_TERM(const amrex::Real x = prob_lo[0] + (i + 0.5) * dx[0];
-               , const amrex::Real y = prob_lo[1] + (j + 0.5) * dx[1];
-               , const amrex::Real z = prob_lo[2] + (k + 0.5) * dx[2];);
-
-  AMREX_D_TERM(const amrex::Real Lx = prob_hi[0] - prob_lo[0];
-               , const amrex::Real Ly = prob_hi[1] - prob_lo[1];
-               , const amrex::Real Lz = prob_hi[2] - prob_lo[2]);
-
-  AMREX_D_TERM(const amrex::Real x_c = prob_lo[0] + 0.5 * Lx;
-               , const amrex::Real y_c = prob_lo[1] + 0.5 * Ly;
-               , const amrex::Real z_c = prob_lo[2] + 0.5 * Lz);
-
-  auto eos = pele::physics::PhysicsType::eos();
-  constexpr amrex::Real Pi = 3.14159265358979323846264338327950288;
-
-  amrex::Real massfrac[NUM_SPECIES] = {0.0};
-  massfrac[O2_ID] = 0.233;
-  massfrac[N2_ID] = 0.767;
-
-  switch (prob_parm.meanFlowDir) {
-  case 1:
-    AMREX_D_TERM(state(i, j, k, VELX) = prob_parm.meanFlowMag;
-                 , state(i, j, k, VELY) = 0.0;, state(i, j, k, VELZ) = 0.0);
-    break;
-  case -1:
-    AMREX_D_TERM(state(i, j, k, VELX) = -prob_parm.meanFlowMag;
-                 , state(i, j, k, VELY) = 0.0;, state(i, j, k, VELZ) = 0.0);
-    break;
-  case 2:
-    AMREX_D_TERM(state(i, j, k, VELX) = 0.0;
-                 , state(i, j, k, VELY) = prob_parm.meanFlowMag;
-                 , state(i, j, k, VELZ) = 0.0);
-    break;
-  case -2:
-    AMREX_D_TERM(state(i, j, k, VELX) = 0.0;
-                 , state(i, j, k, VELY) = -prob_parm.meanFlowMag;
-                 , state(i, j, k, VELZ) = 0.0);
-    break;
-  case 3:
-    AMREX_D_TERM(state(i, j, k, VELX) = prob_parm.meanFlowMag;
[…]

[…]
