### 2023-09-19 78e2284
Fix some readability (#254)
fix readability-convert-member-functions-to-static, readability-make-member-function-const, readability-named-parameter, readability-inconsistent-declaration-parameter-name, readability-else-after-return, readability-redundant-smartptr-get, readability-simplify-boolean-expr, readability-container-data-pointer'
