+++ Source/PeleLMUtils.cpp
@@ -1593,7 +1593,7 @@ PeleLM::MLmin(const Vector<const MultiFab*>& a_MF, int scomp, int ncomp)
 }
 
 void
-PeleLM::checkMemory(const std::string& a_message)
+PeleLM::checkMemory(const std::string& a_message) const
 {
   if (m_checkMem == 0) {
     return;
