+++ Source/PeleLMEvolve.cpp
@@ -119,7 +119,7 @@ PeleLM::Evolve()
 }
 
 bool
-PeleLM::writePlotNow()
+PeleLM::writePlotNow() const
 {
   bool write_now = false;
 
@@ -167,7 +167,7 @@ PeleLM::writePlotNow()
 }
 
 bool
-PeleLM::writeCheckNow()
+PeleLM::writeCheckNow() const
 {
   bool write_now = false;
 
@@ -209,7 +209,7 @@ PeleLM::writeCheckNow()
 }
 
 bool
-PeleLM::doTemporalsNow()
+PeleLM::doTemporalsNow() const
 {
   bool write_now = false;
 
@@ -221,7 +221,7 @@ PeleLM::doTemporalsNow()
 }
 
 bool
-PeleLM::checkMessage(const std::string& a_action)
+PeleLM::checkMessage(const std::string& a_action) const
 {
   bool take_action = false;
 
