+++ Source/PeleLMSetup.cpp
@@ -1148,7 +1148,7 @@ PeleLM::taggingSetup()
       Vector<Real> box_hi(AMREX_SPACEDIM);
       ppr.getarr("in_box_lo", box_lo, 0, box_lo.size());
       ppr.getarr("in_box_hi", box_hi, 0, box_hi.size());
-      realbox = RealBox(&(box_lo[0]), &(box_hi[0]));
+      realbox = RealBox(box_lo.data(), box_hi.data());
     }
 
     AMRErrorTagInfo info;
