### 2023-09-19 12f1d1f
Fix HIP CI (#256)
- Also convert some CI tests to CTest.
