+++ Exec/RegTests/HITDecay/pelelm_prob.cpp
@@ -101,11 +101,11 @@ PeleLM::readProbParm()
 {
   amrex::ParmParse pp("prob");
 
-  ProbParm prob_parm;
+  ProbParm local_prob_parm;
 
   // Read HIT user-defined options from the input file
-  pp.query("input_resolution", prob_parm.input_resolution);
-  if (prob_parm.input_resolution == 0) {
+  pp.query("input_resolution", local_prob_parm.input_resolution);
+  if (local_prob_parm.input_resolution == 0) {
     amrex::Abort("for HIT, the input resolution cannot be 0 !");
   }
 
@@ -113,22 +113,22 @@ PeleLM::readProbParm()
   pp.query("input_name", datafile);
   int binfmt = 0; // Default is ASCII format
   pp.query("input_binaryformat", binfmt);
-  pp.query("urms0", prob_parm.urms0);
+  pp.query("urms0", local_prob_parm.urms0);
 
   amrex::Real lambda0 = 0.5;
-  amrex::Real tau = lambda0 / prob_parm.urms0;
+  amrex::Real tau = lambda0 / local_prob_parm.urms0;
 
   // Output initial conditions
   std::ofstream ofs("initialConditions.txt", std::ofstream::out);
   amrex::Print(ofs) << "lambda0, urms0, tau \n";
   amrex::Print(ofs).SetPrecision(17)
-    << lambda0 << "," << prob_parm.urms0 << "," << tau << std::endl;
+    << lambda0 << "," << local_prob_parm.urms0 << "," << tau << std::endl;
   ofs.close();
 
   // Read initial velocity field
-  const size_t nx = prob_parm.input_resolution;
-  const size_t ny = prob_parm.input_resolution;
-  const size_t nz = prob_parm.input_resolution;
+  const int nx = local_prob_parm.input_resolution;
+  const int ny = local_prob_parm.input_resolution;
+  const int nz = local_prob_parm.input_resolution;
   amrex::Vector<amrex::Real> data(
     nx * ny * nz * 6); /* this needs to be double */
   if (binfmt != 0) {
@@ -152,9 +152,12 @@ PeleLM::readProbParm()
 
   for (long i = 0; i < xinput.size(); i++) {
     xinput[i] = data[0 + i * 6];
-    uinput[i] = data[3 + i * 6] * prob_parm.urms0 / prob_parm.uin_norm;
-    vinput[i] = data[4 + i * 6] * prob_parm.urms0 / prob_parm.uin_norm;
-    winput[i] = data[5 + i * 6] * prob_parm.urms0 / prob_parm.uin_norm;
+    uinput[i] =
+      data[3 + i * 6] * local_prob_parm.urms0 / local_prob_parm.uin_norm;
+    vinput[i] =
+      data[4 + i * 6] * local_prob_parm.urms0 / local_prob_parm.uin_norm;
+    winput[i] =
+      data[5 + i * 6] * local_prob_parm.urms0 / local_prob_parm.uin_norm;
   }
 
   // Get the xarray table and the differences.
@@ -171,28 +174,28 @@ PeleLM::readProbParm()
     amrex::Abort("Error: non ascending x-coordinate array.");
   }
 
-  // Pass data to the prob_parm
-  prob_parm.Linput = xarray[nx - 1] + 0.5 * xdiff[nx - 1];
+  // Pass data to the local_prob_parm
[…]
