+++ Exec/RegTests/PeriodicCases/pelelm_prob.H
@@ -19,27 +19,25 @@ pelelm_initdata(
   int k,
   int is_incompressible,
   amrex::Array4<amrex::Real> const& state,
-  amrex::Array4<amrex::Real> const& aux,
+  amrex::Array4<amrex::Real> const& /*aux*/,
   amrex::GeometryData const& geomdata,
   ProbParm const& prob_parm,
   pele::physics::PMF::PmfData::DataContainer const* /*pmf_data*/)
 {
-  const amrex::Real* prob_lo = geomdata.ProbLo();
-  const amrex::Real* prob_hi = geomdata.ProbHi();
+  // const amrex::Real* prob_lo = geomdata.ProbLo();
+  // const amrex::Real* prob_hi = geomdata.ProbHi();
   const amrex::Real* dx = geomdata.CellSize();
-
-  AMREX_D_TERM(const amrex::Real x = prob_lo[0] + (i + 0.5) * dx[0];
-               , const amrex::Real y = prob_lo[1] + (j + 0.5) * dx[1];
-               , const amrex::Real z = prob_lo[2] + (k + 0.5) * dx[2];);
-
-  AMREX_D_TERM(const amrex::Real Lx = prob_hi[0] - prob_lo[0];
-               , const amrex::Real Ly = prob_hi[1] - prob_lo[1];
-               , const amrex::Real Lz = prob_hi[2] - prob_lo[2]);
-
-  AMREX_D_TERM(const amrex::Real x_c = prob_lo[0] + 0.5 * Lx;
-               , const amrex::Real y_c = prob_lo[1] + 0.5 * Ly;
-               , const amrex::Real z_c = prob_lo[2] + 0.5 * Lz);
-
+  /*
+    AMREX_D_TERM(const amrex::Real x = prob_lo[0] + (i + 0.5) * dx[0];
+                 , const amrex::Real y = prob_lo[1] + (j + 0.5) * dx[1];
+                 , const amrex::Real z = prob_lo[2] + (k + 0.5) * dx[2];);
+    AMREX_D_TERM(const amrex::Real Lx = prob_hi[0] - prob_lo[0];
+                 , const amrex::Real Ly = prob_hi[1] - prob_lo[1];
+                 , const amrex::Real Lz = prob_hi[2] - prob_lo[2]);
+    AMREX_D_TERM(const amrex::Real x_c = prob_lo[0] + 0.5 * Lx;
+                 , const amrex::Real y_c = prob_lo[1] + 0.5 * Ly;
+                 , const amrex::Real z_c = prob_lo[2] + 0.5 * Lz);
+  */
   auto eos = pele::physics::PhysicsType::eos();
   constexpr amrex::Real Pi = 3.14159265358979323846264338327950288;
 
