### 2023-09-21 cbc4036
Fix some more warnings.
