### 2023-09-22 7ad9a3e
Turn on eb for clang-tidy (#255)
* Turn on eb for clang-tidy

* Maybe this works

* Don't fail the EB matrix.

---------

Co-authored-by: Jon Rood <jon.rood@nrel.gov>
