+++ Source/PeleLMeX_DiffusionOp.cpp
@@ -332,10 +332,10 @@ DiffusionOp::diffuse_scalar(
   AMREX_ASSERT(m_ncomp == 1 || m_ncomp == ncomp);
   AMREX_ASSERT(a_phi[0]->nComp() >= phi_comp + ncomp);
   AMREX_ASSERT(a_rhs[0]->nComp() >= rhs_comp + ncomp);
-  if (have_fluxes) {
+  if (have_fluxes != 0) {
     AMREX_ASSERT(a_flux[0][0]->nComp() >= flux_comp + ncomp);
   }
-  if (have_bcoeff) {
+  if (have_bcoeff != 0) {
     AMREX_ASSERT(a_bcoeff[0]->nComp() >= bcoeff_comp + ncomp);
     AMREX_ASSERT(a_bcrec.size() >= ncomp);
   }
@@ -361,7 +361,7 @@ DiffusionOp::diffuse_scalar(
       const Box& gbx = mfi.growntilebox();
       auto const& a_phi_arr = a_phi[lev]->const_array(mfi, phi_comp);
       auto const& a_rho_arr =
-        (have_density)
+        (have_density) != 0
           ? a_density[lev]->const_array(mfi)
           : a_phi[lev]->const_array(mfi); // Get dummy Array4 if no density
       auto const& phi_arr = phi[lev].array(mfi);
@@ -369,7 +369,7 @@ DiffusionOp::diffuse_scalar(
         gbx, ncomp,
         [a_phi_arr, a_rho_arr, phi_arr,
          have_density] AMREX_GPU_DEVICE(int i, int j, int k, int n) noexcept {
-          if (have_density) {
+          if (have_density != 0) {
             phi_arr(i, j, k, n) = a_phi_arr(i, j, k, n) / a_rho_arr(i, j, k);
           } else {
             phi_arr(i, j, k, n) = a_phi_arr(i, j, k, n);
@@ -384,11 +384,11 @@ DiffusionOp::diffuse_scalar(
   // => \alpha = 1.0, A is a_acoeff if provided, 1.0 otherwise
   // => \beta = a_dt, B face centered diffusivity bcoeff^{np1,k}
 
-  Real alpha = (isPoissonSolve) ? 0.0 : 1.0;
+  Real alpha = (isPoissonSolve) != 0 ? 0.0 : 1.0;
   Real beta = a_dt;
   m_scal_solve_op->setScalars(alpha, beta);
   for (int lev = 0; lev <= finest_level; ++lev) {
-    if (have_acoeff) {
+    if (have_acoeff != 0) {
       m_scal_solve_op->setACoeffs(lev, *a_acoeff[lev]);
     } else {
       m_scal_solve_op->setACoeffs(lev, 1.0);
@@ -411,14 +411,14 @@ DiffusionOp::diffuse_scalar(
 
     // Set aliases and bcoeff comp
     for (int lev = 0; lev <= finest_level; ++lev) {
-      if (have_fluxes) {
+      if (have_fluxes != 0) {
         for (int idim = 0; idim < AMREX_SPACEDIM; idim++) {
           fluxes[lev][idim] = new MultiFab(
             *a_flux[lev][idim], amrex::make_alias, flux_comp + comp, m_ncomp);
         }
       }
 
-      if (have_bcoeff) {
+      if (have_bcoeff != 0) {
         int doZeroVisc = 1;
         Vector<BCRec> subBCRec = {
           a_bcrec.begin() + comp, a_bcrec.begin() + comp + m_ncomp};
@@ -462,7 +462,7 @@ DiffusionOp::diffuse_scalar(
[…]
