+++ Source/PeleLMeX_DiffusionOp.cpp
@@ -302,7 +302,7 @@ DiffusionOp::diffuse_scalar(
   Vector<MultiFab*> const& a_phi,
   int phi_comp,
   Vector<MultiFab const*> const& a_phiEB,
-  int phiEB_comp,
+  int /*phiEB_comp*/,
   Vector<MultiFab const*> const& a_rhs,
   int rhs_comp,
   Vector<Array<MultiFab*, AMREX_SPACEDIM>> const& a_flux,
@@ -312,7 +312,7 @@ DiffusionOp::diffuse_scalar(
   Vector<MultiFab const*> const& a_bcoeff,
   int bcoeff_comp,
   Vector<MultiFab const*> const& a_bcoeffEB,
-  int bcoeffEB_comp,
+  int /*bcoeffEB_comp*/,
   Vector<BCRec> a_bcrec,
   int ncomp,
   int isPoissonSolve,
