+++ Exec/RegTests/HITDecay/pelelmex_prob.H
@@ -171,7 +171,7 @@ void
 bcnormal(
   const amrex::Real* /*x[AMREX_SPACEDIM]*/,
   const int /*m_nAux*/,
-  amrex::Real[NVAR] /*s_ext[NVAR]*/,
+  amrex::Real* /*s_ext[NVAR]*/,
   const int /*idir*/,
   const int /*sgn*/,
   const amrex::Real /*time*/,

[…]
