+++ Exec/RegTests/HotBubble/pelelmex_prob.H
@@ -17,7 +17,7 @@ pelelmex_initdata(
   int i,
   int j,
   int k,
-  int is_incompressible,
+  int /*is_incompressible*/,
   amrex::Array4<amrex::Real> const& state,
   amrex::Array4<amrex::Real> const& /*aux*/,
   amrex::GeometryData const& geomdata,
@@ -33,10 +33,8 @@ pelelmex_initdata(
                , const amrex::Real z = prob_lo[2] + (k + 0.5) * dx[2];);
 
   AMREX_D_TERM(const amrex::Real Lx = prob_hi[0] - prob_lo[0];
-               , const amrex::Real Ly = prob_hi[1] - prob_lo[1];
+               , /*const amrex::Real Ly = prob_hi[1] - prob_lo[1];*/
                , const amrex::Real Lz = prob_hi[2] - prob_lo[2]);
-
-  constexpr amrex::Real Pi = 3.14159265358979323846264338327950288;
 
   auto eos = pele::physics::PhysicsType::eos();
   amrex::Real massfrac[NUM_SPECIES] = {0.0};
