+++ Exec/RegTests/HotBubble/pelelmex_prob.H
@@ -17,12 +17,12 @@ pelelmex_initdata(
   int i,
   int j,
   int k,
-  int is_incompressible,
+  int /*is_incompressible*/,
   amrex::Array4<amrex::Real> const& state,
-  amrex::Array4<amrex::Real> const& aux,
+  amrex::Array4<amrex::Real> const& /*aux*/,
   amrex::GeometryData const& geomdata,
   ProbParm const& prob_parm,
-  pele::physics::PMF::PmfData::DataContainer const* pmf_data)
+  pele::physics::PMF::PmfData::DataContainer const* /*pmf_data*/)
 {
   const amrex::Real* prob_lo = geomdata.ProbLo();
   const amrex::Real* prob_hi = geomdata.ProbHi();
@@ -33,10 +33,8 @@ pelelmex_initdata(
                , const amrex::Real z = prob_lo[2] + (k + 0.5) * dx[2];);
 
   AMREX_D_TERM(const amrex::Real Lx = prob_hi[0] - prob_lo[0];
-               , const amrex::Real Ly = prob_hi[1] - prob_lo[1];
+               , /*const amrex::Real Ly = prob_hi[1] - prob_lo[1];*/
                , const amrex::Real Lz = prob_hi[2] - prob_lo[2]);
-
-  constexpr amrex::Real Pi = 3.14159265358979323846264338327950288;
 
   auto eos = pele::physics::PhysicsType::eos();
   amrex::Real massfrac[NUM_SPECIES] = {0.0};
@@ -47,7 +45,7 @@ pelelmex_initdata(
   state(i, j, k, TEMP) = prob_parm.T_mean;
   amrex::Real x_c = prob_parm.is_sym != 0 ? prob_lo[0] : prob_lo[0] + 0.5 * Lx;
 #if AMREX_SPACEDIM == 3
-  amrex::Real z_c = prob_parm.is_sym ? prob_lo[2] : prob_lo[2] + 0.5 * Lz;
+  amrex::Real z_c = prob_parm.is_sym != 0 ? prob_lo[2] : prob_lo[2] + 0.5 * Lz;
 #endif
   amrex::Real rad = std::sqrt(AMREX_D_TERM(
     (y - prob_parm.bubble_y0) * (y - prob_parm.bubble_y0),

[…]
