+++ Source/PeleLMeX_Advection.cpp
@@ -481,7 +481,7 @@ PeleLM::computeScalarAdvTerms(std::unique_ptr<AdvanceAdvData>& advData)
                 }
               }
             });
-        } else // Regular boxes
+        } else { // Regular boxes
 #endif
           amrex::ParallelFor(
             ebx,
@@ -491,6 +491,7 @@ PeleLM::computeScalarAdvTerms(std::unique_ptr<AdvanceAdvData>& advData)
                 rho_ed(i, j, k) += rhoY_ed(i, j, k, n);
               }
             });
+        }
       }
     }
 
@@ -629,9 +630,9 @@ PeleLM::computeScalarAdvTerms(std::unique_ptr<AdvanceAdvData>& advData)
       GetArrOfConstPtrs(fluxes[lev]), GetArrOfPtrs(fluxes[lev - 1]),
       refRatio(lev - 1), geom[lev - 1]);
 #else
-    average_down_faces(
-      GetArrOfConstPtrs(fluxes[lev]), GetArrOfPtrs(fluxes[lev - 1]),
-      refRatio(lev - 1), geom[lev - 1]);
+  average_down_faces(
+    GetArrOfConstPtrs(fluxes[lev]), GetArrOfPtrs(fluxes[lev - 1]),
+    refRatio(lev - 1), geom[lev - 1]);
 #endif
   }
 
@@ -714,14 +715,14 @@ PeleLM::computeScalarAdvTerms(std::unique_ptr<AdvanceAdvData>& advData)
       RHOH, 1, bcRecRhoH_d.dataPtr(), geom[lev]);
     EB_set_covered(advData->AofS[lev], 0.0);
 #else
-    //----------------------------------------------------------------
-    // Otherwise go directly into AofS
-    advFluxDivergence(
-      lev, advData->AofS[lev], FIRSTSPEC, divu, GetArrOfConstPtrs(fluxes[lev]),
-      0, GetArrOfConstPtrs(fluxes[lev]),
-      0, // This will not be used since none of rhoY/rhoH in convective
-      NUM_SPECIES + 1, AdvTypeAll_d.dataPtr(), geom[lev], -1.0,
-      fluxes_are_area_weighted);
+  //----------------------------------------------------------------
+  // Otherwise go directly into AofS
+  advFluxDivergence(
+    lev, advData->AofS[lev], FIRSTSPEC, divu, GetArrOfConstPtrs(fluxes[lev]), 0,
+    GetArrOfConstPtrs(fluxes[lev]),
+    0, // This will not be used since none of rhoY/rhoH in convective
+    NUM_SPECIES + 1, AdvTypeAll_d.dataPtr(), geom[lev], -1.0,
+    fluxes_are_area_weighted);
 #endif
   }
 
