+++ Source/PeleLMeX_Advection.cpp
@@ -145,7 +145,7 @@ PeleLM::computeVelocityAdvTerm(std::unique_ptr<AdvanceAdvData>& advData)
 
     bool fluxes_are_area_weighted = false;
 #ifdef AMREX_USE_EB
-    auto ldata_p = getLevelDataPtr(lev, AmrOldTime);
+    auto* ldata_p = getLevelDataPtr(lev, AmrOldTime);
     //----------------------------------------------------------------
     // Use a temporary MF to hold divergence before redistribution
     int nGrow_divT = 3;
@@ -483,6 +483,7 @@ PeleLM::computeScalarAdvTerms(std::unique_ptr<AdvanceAdvData>& advData)
             });
         } else // Regular boxes
 #endif
+        {
           amrex::ParallelFor(
             ebx,
             [rho_ed, rhoY_ed] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
@@ -491,6 +492,7 @@ PeleLM::computeScalarAdvTerms(std::unique_ptr<AdvanceAdvData>& advData)
                 rho_ed(i, j, k) += rhoY_ed(i, j, k, n);
               }
             });
+        }
       }
     }
 
@@ -687,7 +689,7 @@ PeleLM::computeScalarAdvTerms(std::unique_ptr<AdvanceAdvData>& advData)
 
     bool fluxes_are_area_weighted = false;
 #ifdef AMREX_USE_EB
-    auto ldata_p = getLevelDataPtr(lev, AmrOldTime);
+    auto* ldata_p = getLevelDataPtr(lev, AmrOldTime);
     //----------------------------------------------------------------
     // Use a temporary MF to hold divergence before redistribution
     int nGrow_divTmp = 3;
@@ -889,7 +891,7 @@ PeleLM::computePassiveAdvTerms(
 
     bool fluxes_are_area_weighted = false;
 #ifdef AMREX_USE_EB
-    auto ldata_p = getLevelDataPtr(lev, AmrOldTime);
+    auto* ldata_p = getLevelDataPtr(lev, AmrOldTime);
     //----------------------------------------------------------------
     // Use a temporary MF to hold divergence before redistribution
     int nGrow_divTmp = 3;
