+++ Source/PeleLMeX_Utils.cpp
@@ -1928,8 +1928,9 @@ PeleLM::extendSignedDistance(MultiFab* a_signDist, Real a_extendFactor)
                       +((j - jj) * dx[1] * (j - jj) * dx[1]),
                       +((k - kk) * dx[2] * (k - kk) * dx[2])));
                     Real distToEB = distToCell + sd_cc(ii, jj, kk);
-                    if (distToEB < closestEBDist)
+                    if (distToEB < closestEBDist) {
                       closestEBDist = distToEB;
+                    }
                   }
                 }
               }

[…]
