+++ Source/PeleLMeX_FlowController.cpp
@@ -59,77 +59,79 @@ PeleLM::initActiveControl()
   FlowControllerData* fcdata_host{nullptr};
   fcdata_host = getFCDataPtr(*prob_parm, hasFlowControllerData<ProbParm>{});
 
-  // Resize vector for temporal average
-  m_ctrl_time_pts.resize(m_ctrl_NavgPts + 1, -1.0);
-  m_ctrl_velo_pts.resize(m_ctrl_NavgPts + 1, -1.0);
-  m_ctrl_cntl_pts.resize(m_ctrl_NavgPts + 1, -1.0);
+  if (fcdata_host) {
+    // Resize vector for temporal average
+    m_ctrl_time_pts.resize(m_ctrl_NavgPts + 1, -1.0);
+    m_ctrl_velo_pts.resize(m_ctrl_NavgPts + 1, -1.0);
+    m_ctrl_cntl_pts.resize(m_ctrl_NavgPts + 1, -1.0);
 
-  // Extract data from BC: assumes flow comes in from lo side of ctrl_flameDir
-  ProbParm const* lprobparm = prob_parm_d;
-  pele::physics::PMF::PmfData::DataContainer const* lpmfdata =
-    pmf_data.getDeviceData();
+    // Extract data from BC: assumes flow comes in from lo side of ctrl_flameDir
+    ProbParm const* lprobparm = prob_parm_d;
+    pele::physics::PMF::PmfData::DataContainer const* lpmfdata =
+      pmf_data.getDeviceData();
 
-  Gpu::DeviceVector<Real> s_ext_v(NVAR);
-  Real* s_ext_d = s_ext_v.data();
-  Real x[AMREX_SPACEDIM] = {
-    AMREX_D_DECL(Geom(0).ProbLo(0), Geom(0).ProbLo(1), Geom(0).ProbLo(2))};
-  x[m_ctrl_flameDir] -= 1.0;
-  const int ctrl_flameDir_l = m_ctrl_flameDir;
-  const amrex::Real time_l = -1.0;
-  const auto geomdata = Geom(0).data();
+    Gpu::DeviceVector<Real> s_ext_v(NVAR);
+    Real* s_ext_d = s_ext_v.data();
+    Real x[AMREX_SPACEDIM] = {
+      AMREX_D_DECL(Geom(0).ProbLo(0), Geom(0).ProbLo(1), Geom(0).ProbLo(2))};
+    x[m_ctrl_flameDir] -= 1.0;
+    const int ctrl_flameDir_l = m_ctrl_flameDir;
+    const amrex::Real time_l = -1.0;
+    const auto geomdata = Geom(0).data();
 
-  Box dumbx({AMREX_D_DECL(0, 0, 0)}, {AMREX_D_DECL(0, 0, 0)});
-  amrex::ParallelFor(
-    dumbx,
-    [x, nAux = m_nAux, s_ext_d, ctrl_flameDir_l, time_l, geomdata, lprobparm,
-     lpmfdata] AMREX_GPU_DEVICE(int /*i*/, int /*j*/, int /*k*/) noexcept {
-      bcnormal(
-        x, nAux, s_ext_d, ctrl_flameDir_l, 1, time_l, geomdata, *lprobparm,
-        lpmfdata);
-    });
-  Vector<Real> s_ext(NVAR);
-  Gpu::copy(Gpu::deviceToHost, s_ext_v.begin(), s_ext_v.end(), s_ext.begin());
+    Box dumbx({AMREX_D_DECL(0, 0, 0)}, {AMREX_D_DECL(0, 0, 0)});
+    amrex::ParallelFor(
+      dumbx,
[…]
