### 2023-09-28 ebf4ea4
Add efield paper to bib.
