+++ paper/paper.bib
@@ -173,3 +173,15 @@
   year={2022},
   publisher={Elsevier}
 }
+
+@article{Esclapez:2020,
+  title = {A spectral deferred correction strategy for low Mach number reacting flows subject to electric fields},
+  author = {Esclapez, Lucas and Ricchiuti, Valentina and Bell, John B. and Day, Marcus S.},
+  journal = {Combustion Theory and Modelling},
+  volume = {24},
+  number = {2},
+  pages = {194-220},
+  year  = {2020},
+  publisher = {Taylor \& Francis},
+  doi = {10.1080/13647830.2019.1668060}
+}
