+++ Source/Efield/LinOps/AMReX_MLABecCecLaplacian.H
@@ -121,7 +121,7 @@ public:
   virtual std::unique_ptr<MLLinOp>
   makeNLinOp(int /*grid_size*/) const final override
   {
-    amrex::Abort("MLABecLaplacian::makeNLinOp: Not implmented");
+    amrex::Abort("MLABecLaplacian::makeNLinOp: Not implemented");
     return std::unique_ptr<MLLinOp>{};
   }
 
