+++ Source/Efield/LinOps/AMReX_MLABecCecLap_2D_K.H
@@ -736,7 +736,7 @@ abeccec_getnetfcoeff(
 {
   auto lo = amrex::lbound(box);
   auto hi = amrex::ubound(box);
-  // Gettting face coefficient
+  // Getting face coefficient
   if (dir == 0) {
     for (int n = 0; n < nc; ++n) {
       for (int j = lo.y; j <= hi.y; ++j) {
