+++ manual/html/_sources/LMeXControls.rst.txt
@@ -99,12 +99,12 @@ IO parameters
 ::
 
     #--------------------------IO CONTROL--------------------------
-    amr.plot_int         = 20              # [OPT, DEF=-1] Frequency (as step #) for writting plot file
-    amr.plot_per         = 0.002           # [OPT, DEF=-1] Period (time in s) for writting plot file
+    amr.plot_int         = 20              # [OPT, DEF=-1] Frequency (as step #) for writing plot file
+    amr.plot_per         = 0.002           # [OPT, DEF=-1] Period (time in s) for writing plot file
     amr.plot_per_exact   = 1               # [OPT, DEF=0] Flag to enforce exactly plt_per by shortening dt
     amr.plot_file        = "plt_"          # [OPT, DEF="plt_"] Plot file prefix
-    amr.check_int        = 100             # [OPT, DEF=-1] Frequency (as step #) for writting checkpoint file
-    amr.check_per        = 0.05            # [OPT, DEF=-1] Period (time in s) for writting checkpoint file
+    amr.check_int        = 100             # [OPT, DEF=-1] Frequency (as step #) for writing checkpoint file
+    amr.check_per        = 0.05            # [OPT, DEF=-1] Period (time in s) for writing checkpoint file
     amr.check_file       = "chk"           # [OPT, DEF="chk"] Checkpoint file prefix
     amr.file_stepDigits  = 6               # [OPT, DEF=5] Number of digits when adding nsteps to plt and chk names
     amr.derive_plot_vars = avg_pressure ...# [OPT, DEF=""] List of derived variable included in the plot files
@@ -269,7 +269,7 @@ Chemistry integrator
     peleLM.use_typ_vals_chem = 1                # [OPT, DEF=1] Use Typical values to scale components in the reactors
     peleLM.typical_values_reset_int = 5         # [OPT, DEF=10] Frequency at which the typical values are updated
     ode.rtol = 1.0e-6                           # [OPT, DEF=1e-10] Relative tolerance of the chem. reactor
-    ode.atol = 1.0e-6                           # [OPT, DEF=1e-10] Aboslute tolerance of the chem. reactor, or pre-factor of the typical values when used
+    ode.atol = 1.0e-6                           # [OPT, DEF=1e-10] Absolute tolerance of the chem. reactor, or pre-factor of the typical values when used
     cvode.solve_type = denseAJ_direct           # [OPT, DEF=GMRES] Linear solver employed for CVODE Newton direction
     cvode.max_order  = 4                        # [OPT, DEF=2] Maximum order of the BDF method in CVODE
 
@@ -278,7 +278,7 @@ Note that the last four parameters belong to the Reactor class of PelePhysics bu
 Embedded Geometry
 -----------------
 
-`PeleLMeX` geomtry relies on AMReX implementation of the EB method. Simple geometrical objects
+`PeleLMeX` geometry relies on AMReX implementation of the EB method. Simple geometrical objects
 can thus be constructed using `AMReX internal parser <https://amrex-codes.github.io/amrex/docs_html/EB.html>`_.
 For instance, setting up a sphere of radius 5 mm can be achieved:
 
@@ -397,7 +397,7 @@ Run-time diagnostics
 
 `PeleLMeX` provides a few diagnostics to check you simulations while it is running as well as adding basic analysis ingredients.
 
-It is often usefull to have an estimate of integrated quantities (kinetic energy, heat release rate, ,..), state extremas
+It is often useful to have an estimate of integrated quantities (kinetic energy, heat release rate, ,..), state extremas
 or other overall balance information to get a sense of the status and sanity of the simulation. To this end, it is possible
 to activate `temporal` diagnostics performing these reductions at given intervals:
 
@@ -412,7 +412,7 @@ to activate `temporal` diagnostics performing these reductions at given interval
 
[…]
