+++ manual/html/_sources/Tutorials_HotBubble.rst.txt
@@ -187,7 +187,7 @@ Building the executable
 Now that we have reviewed the basic ingredients required to setup the case, it is time to build the `PeleLMeX` executable.
 Although both GNUmake and CMake are available, it is advised to use GNUmake. The ``GNUmakefile`` file provides some compile-time options
 regarding the simulation we want to perform.
-The first few lines specify the paths towards the source codes of `PeleLMeX`, `AMReX`, `AMReX-Hydro` and `PelePhysics`, overwritting
+The first few lines specify the paths towards the source codes of `PeleLMeX`, `AMReX`, `AMReX-Hydro` and `PelePhysics`, overwriting
 any environment variable if necessary, and might have been already updated in :ref:`sec:TUTO_HB::PrepStep` earlier.
 
 The next few lines specify AMReX compilation options and compiler selection: ::
@@ -228,7 +228,7 @@ list of available mechanisms and more information regarding the EOS, chemistry a
     Transport_Model := Constant
 
 Finally, `PeleLMeX` utilizes the chemical kinetic ODE integrator `CVODE <https://computing.llnl.gov/projects/sundials/cvode>`_. This
-Third Party Librabry (TPL) is shipped as a submodule of the `PeleLMeX` distribution and can be readily installed through the makefile system
+Third Party Library (TPL) is shipped as a submodule of the `PeleLMeX` distribution and can be readily installed through the makefile system
 of `PeleLMeX`. To do so, type in the following command: ::
 
     make -j4 TPL
@@ -308,7 +308,7 @@ A typical `PeleLMeX` stdout for a time step now looks like: ::
     >> PeleLMeX::Advance() --> Time: 0.2141339779
 
 clearly showing the use of 1 SDC iteration. The first line at each step provides
-the time step contraint from the CFL
+the time step constraint from the CFL
 condition (``Conv:``) and from the density change condition (``divu:``).
 In the absence of reaction and diffusion, the ``divu:`` constraint is irrelevant and set to a
 large value.

[…]
