### 2023-10-03 #5450 @esclapez
Hi @kyleniemeyer, @mameehan5, @xzz105, @muellerm-pu,

Thank you for taking the time to handle and review this submission and sorry again for the delay in my replies.

Following the reviewers suggestions, we have implemented the following changes to the README:

- a more complete overview/statement of need has been added , to clarify what PeleLMeX is intended for, providing a couple of pointers to recent papers employing PeleLMeX.
- requirements for installing and running PeleLMeX (on linux/macOS) were added.
- multiple typographical errors were corrected.
- clarifications were brought regarding the process for reporting an issue/contributing/requiring support.
- the filiation with PeleLM is still mentioned, but no references are made to PeleLM documentation/sources.

, additions to the example/documentation:

- instead of a single README for all the examples, individual READMEs in each `Exec` subfolders were added, pointing towards the appropriate more complete piece of documentation when available.
- a brief tutorial on the `HotBubble` case was added, making it easier to move forward into using PeleLMeX from the main README.

 and changes to the sources:

-  following an unfortunate initial decision `PeleLM` was used instead of `PeleLMeX` in many places throughout the source code. This has been addressed.
- style and linting has been enforced on the code base.

I think these changes address all the reviewer comments appropriately. We thank the reviewers for their comments, as these improvements make it easier for new users to start with PeleLMeX.
