+++ paper/paper.bib
@@ -185,3 +185,51 @@
   publisher = {Taylor \& Francis},
   doi = {10.1080/13647830.2019.1668060}
 }
+
+@misc{NGA2,
+  author = {{NGA2 Team}},
+  title = {NGA2: Object-oriented multi-mesh version of the classic reacting turbulent multiphase flow solver},
+  year = {2023},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/desjardi/NGA2}
+}
+
+@article{Hassanaly:2018,
+  title={A minimally-dissipative low-Mach number solver for complex reacting flows in {OpenFOAM}},
+  author={Hassanaly, Malik and Koo, Heeseok and Lietz, Christopher F and Chong, Shao Teng and Raman, Venkat},
+  journal={Computers \& Fluids},
+  volume={162},
+  pages={11--25},
+  year={2018},
+  publisher={Elsevier},
+  doi = {10.1016/j.compfluid.2017.11.020}
+}
+
+@inproceedings{Domino:2003,
+  title={{SIERRA/Fuego}: A multi-mechanics fire environment simulation tool},
+  author={Domino, Stefan and Moen, Chris and Burns, Shawn and Evans, Greg},
+  booktitle={41st Aerospace Sciences Meeting and Exhibit},
+  pages={149},
+  year={2003},
+  doi = {10.2514/6.2003-149}
+}
+
+@article{Howarth:2023,
+  title={Thermodiffusively-unstable lean premixed hydrogen flames: Phenomenology, empirical modelling, and thermal leading points},
+  author={Howarth, Thomas and Hunt, EF and Aspden, Andy},
+  journal={Combustion and Flame},
+  volume={253},
+  pages={112811},
+  year={2023},
+  publisher={Elsevier},
+  doi = {10.1016/j.combustflame.2023.112811}
+}
+
+@inproceedings{Appukuttan:2023,
+  title={Simulations of fuel-air mixing in a 7 element Lean Direct Injection ({LDI}) aviation combustor},
+  author={Appukuttan, Sreejith Nadakkal and Perry, Bruce and Yellapantula, Shashank and Esclapez, Lucas and Sitaraman, Hariswaran and Day, Marc},
+  booktitle={Proceedings of the 13th U.S. National Combustion Meeting},
+  organization={Combustion Institute},
+  year={2023}
+}
