### 2023-10-03 0c8ff0c
paper acknowledgements and typos and fix some refs
