+++ paper/paper.bib
@@ -93,6 +93,15 @@
   url = {https://github.com/AMReX-Codes/AMReX-Hydro}
 }
 
+@misc{PeleLM,
+  author = {{PeleLM Team}},
+  title = {PeleLM},
+  year = {2022},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/AMReX-Combustion/PeleLM}
+}
+
 @inproceedings{Hypre2002,
   title={hypre: A library of high performance preconditioners},
   author={Falgout, Robert D and Yang, Ulrike Meier},
@@ -162,7 +171,7 @@
   year={2022},
   publisher={Elsevier}
 }
- 
+
 @article{Perry:2022,
   title={Co-optimized machine-learned manifold models for large eddy simulation of turbulent combustion},
   author={Perry, Bruce A and de Frahan, Marc T Henry and Yellapantula, Shashank},
@@ -206,7 +215,7 @@
   doi = {10.1016/j.compfluid.2017.11.020}
 }
 
-@inproceedings{Domino:2003,
+@inproceedings{Domingo:2003,
   title={{SIERRA/Fuego}: A multi-mechanics fire environment simulation tool},
   author={Domino, Stefan and Moen, Chris and Burns, Shawn and Evans, Greg},
   booktitle={41st Aerospace Sciences Meeting and Exhibit},
@@ -229,7 +238,7 @@
 @inproceedings{Appukuttan:2023,
   title={Simulations of fuel-air mixing in a 7 element Lean Direct Injection ({LDI}) aviation combustor},
   author={Appukuttan, Sreejith Nadakkal and Perry, Bruce and Yellapantula, Shashank and Esclapez, Lucas and Sitaraman, Hariswaran and Day, Marc},
-  booktitle={Proceedings of the 13th U.S. National Combustion Meeting},
+  booktitle={Proceedings of the 13th {U}.{S}. National Combustion Meeting},
   organization={Combustion Institute},
   year={2023}
 }
