### 2023-10-03 b6f7105
paper acknowledgements and typos and fix some refs (#278)
