### 2023-10-03 #5450 @esclapez
Hi @kyleniemeyer, @mameehan5, @xzz105, @muellerm-pu,

Thank you for your reviews of the software paper. We have addressed your suggestions and implemented the following improvements:

- clarify the filiation and differences between PeleLM and PeleLMeX
- provide a few pointers towards other low Mach combustion solvers available in the literature and highlight PeleLMeX distinctive features
- refactored the last paragraph of the `Statement of need` section to mention the targeted audience of PeleLMeX and where its use is particularly suited.

Note that we currently advice all users of PeleLM to transition towards PeleLMeX since the former has not been the target of developments over the past couple of years.

Thank you again!
