### 2023-10-05 a355dc3
Fix typos in joss paper (#280)
Co-authored-by: Bruce Perry <bruce.perry@nrel.gov>
