### 2023-10-05 11f1760
Add ccache to clang-tidy.
