+++ .github/workflows/ci.yml
@@ -508,41 +508,41 @@ jobs:
         run: |
           ccache -s
           du -hs ${HOME}/.cache/ccache
-  Lint-cppcheck:
-    needs: Formatting
-    runs-on: macos-12
-    steps:
-      - name: Clone
-        uses: actions/checkout@v3
-        with:
-          submodules: true
-      - name: Dependencies
-        run: |
-          export HOMEBREW_NO_INSTALLED_DEPENDENTS_CHECK=TRUE
-          wget https://raw.githubusercontent.com/Homebrew/homebrew-core/9b0e3d32f590458ae9e5dbf01943b39845b5315e/Formula/cppcheck.rb
-          brew install cppcheck.rb
-      - name: Configure
-        run: |
-          cmake -B${{runner.workspace}}/build-cppcheck \
-          -DPELELMEX_ENABLE_MPI:BOOL=OFF \
-          -DPELELMEX_ENABLE_FCOMPARE_FOR_TESTS:BOOL=OFF \
-          -DPELELMEX_ENABLE_MASA:BOOL=OFF \
-          -DPELELMEX_ENABLE_CPPCHECK:BOOL=ON \
-          ${{github.workspace}}
-      - name: Check
-        working-directory: ${{runner.workspace}}/build-cppcheck
-        run: make cppcheck
-      - name: Full report
-        working-directory: ${{runner.workspace}}/build-cppcheck/cppcheck
-        run: cat cppcheck-full-report.txt
-      - name: Short report
-        working-directory: ${{runner.workspace}}/build-cppcheck/cppcheck
-        run: |
-          echo "::add-matcher::.github/problem-matchers/cppcheck.json"
-          awk 'BEGIN{i=0}{print $0}{i++}END{print "Warnings: "i}' < cppcheck-report.txt > cppcheck-warnings.txt
-          cat cppcheck-warnings.txt
-          export return=$(tail -n 1 cppcheck-warnings.txt | awk '{print $2}')
-          exit ${return}
+#  Lint-cppcheck:
+#    needs: Formatting
+#    runs-on: macos-12
+#    steps:
+#      - name: Clone
+#        uses: actions/checkout@v3
+#        with:
+#          submodules: true
+#      - name: Dependencies
+#        run: |
+#          export HOMEBREW_NO_INSTALLED_DEPENDENTS_CHECK=TRUE
+#          wget https://raw.githubusercontent.com/Homebrew/homebrew-core/9b0e3d32f590458ae9e5dbf01943b39845b5315e/Formula/cppcheck.rb
+#          brew install cppcheck.rb
+#      - name: Configure
+#        run: |
+#          cmake -B${{runner.workspace}}/build-cppcheck \
+#          -DPELELMEX_ENABLE_MPI:BOOL=OFF \
+#          -DPELELMEX_ENABLE_FCOMPARE_FOR_TESTS:BOOL=OFF \
+#          -DPELELMEX_ENABLE_MASA:BOOL=OFF \
+#          -DPELELMEX_ENABLE_CPPCHECK:BOOL=ON \
+#          ${{github.workspace}}
+#      - name: Check
+#        working-directory: ${{runner.workspace}}/build-cppcheck
+#        run: make cppcheck
+#      - name: Full report
+#        working-directory: ${{runner.workspace}}/build-cppcheck/cppcheck
+#        run: cat cppcheck-full-report.txt
+#      - name: Short report
+#        working-directory: ${{runner.workspace}}/build-cppcheck/cppcheck
+#        run: |
+#          echo "::add-matcher::.github/problem-matchers/cppcheck.json"
[…]
