### 2023-10-05 1c7e4b1
Increase ccache compress level.
