### 2023-10-06 d4469e0
Fix codeql warnings (#279)
