+++ Exec/RegTests/EnclosedInjection/pelelmex_prob.H
@@ -24,20 +24,6 @@ pelelmex_initdata(
   ProbParm const& prob_parm,
   pele::physics::PMF::PmfData::DataContainer const* /*pmf_data*/)
 {
-  // const amrex::Real* prob_lo = geomdata.ProbLo();
-  // const amrex::Real* prob_hi = geomdata.ProbHi();
-  // const amrex::Real* dx = geomdata.CellSize();
-  /*
-    AMREX_D_TERM(const amrex::Real x = prob_lo[0] + (i + 0.5) * dx[0];
-                 , const amrex::Real y = prob_lo[1] + (j + 0.5) * dx[1];
-                 , const amrex::Real z = prob_lo[2] + (k + 0.5) * dx[2];);
-
-    AMREX_D_TERM(const amrex::Real Lx = prob_hi[0] - prob_lo[0];
-                 , const amrex::Real Ly = prob_hi[1] - prob_lo[1];
-                 , const amrex::Real Lz = prob_hi[2] - prob_lo[2]);
-
-    constexpr amrex::Real Pi = 3.14159265358979323846264338327950288;
-  */
   auto eos = pele::physics::PhysicsType::eos();
   amrex::Real massfrac[NUM_SPECIES] = {0.0};
 
@@ -88,10 +74,6 @@ bcnormal(
   amrex::Real X_jet[NUM_SPECIES] = {0.0};
   amrex::Real Y_jet[NUM_SPECIES] = {0.0};
   amrex::Real eta = 0.0;
-
-  // const amrex::Real* prob_lo = geomdata.ProbLo();
-  // const amrex::Real* prob_hi = geomdata.ProbHi();
-  // const amrex::Real* dx = geomdata.CellSize();
 
   // Simple, air into air injection
   X_air[O2_ID] = 0.21;
@@ -160,9 +142,6 @@ zero_visc(
   // velocity, ...)
   if (beta_comp == 0 && nComp == NUM_SPECIES) {
     if (dir == 2 && k <= domainBox.smallEnd(dir)) {
-      // const amrex::Real* prob_lo = geomdata.ProbLo();
-      // const amrex::Real* prob_hi = geomdata.ProbHi();
-      // const amrex::Real* dx = geomdata.CellSize();
 
       // Just over whole inlet
       for (int n = 0; n < nComp; n++) {
