### 2023-10-10 edb5c51
Reduce ccache cache level.
