### 2023-10-13 a85f73f
fixes for Efield after renaming LM to LMeX
