### 2023-10-16 a411767
fix compiling with Efield+EB
