### 2023-10-16 0210f10
fixes for Efield after renaming LM to LMeX (#283)
* fixes for Efield after renaming LM to LMeX

* fix compiling with Efield+EB

* don't allow Efield + EB compiling because it is not yet supported
