### 2023-10-30 #5450 @kyleniemeyer
@esclapez that isn't necessary, the archive just needs to include the accepted version of the software (since we archive the paper itself) - you can leave the paper in a separate branch if you'd like. We also may need to make some final minor edits to the paper after you archive the software.
