### 2023-10-31 #5450 @kyleniemeyer
@esclapez could you merge this PR for the paper? It has some minor fixes to the references: https://github.com/AMReX-Combustion/PeleLMeX/pull/298
